Comparing Two Estimated slca Models

Provides relative model fit test for two fitted SLCM models with deviance statistic.

Usage

compare(
  model1,
  model2,
  test = c("none", "chisq", "boot"),
  nboot = 100,
  method = c("hybrid", "em", "nlm"),
  maxiter = 1000,
  tol = 1e-08,
  verbose = FALSE
)

Arguments

model1

an object of class slca and estimated.

model2

another object of class slca and estimated.

test

a character string specifying the type of test to be conducted. If "chisq", a chi-squared test is conducted. If "boot", a bootstrap test is conducted.

nboot

an integer specifying the number of bootstrap rounds to be performed.

method

estimation method for bootstrapping.

maxiter

an integer specifying maximum number of iterations allowed for the estimation process of each bootstrapping round.

tol

a numeric value setting tolerance for the convergence of each bootstrapping round.

verbose

a logical value indicating whether to print progress updates on the number of bootstrapping rounds completed.

Value

A data.frame containing the number of parameters (Df), loglikelihood, AIC, BIC, G-squared statistics, and the residual degree of freedom for each object. Depending on the test argument, the p-value for the corresponding statistical test may also be included.

See also

Examples

library(magrittr)
data <- gss7677[gss7677$COHORT == "YOUNG", ]
stat2 <- slca(status(2) ~ PAPRES + PADEG + MADEG) %>%
   estimate(data = data)
#> EM iteration begin.
#> 1 iterations  logLik: -1846.42  diff: inf      2 iterations  logLik: -1298.14  diff: 548.27375      3 iterations  logLik: -1272.70  diff: 25.44642      4 iterations  logLik: -1270.40  diff: 2.29407      5 iterations  logLik: -1268.45  diff: 1.95201      6 iterations  logLik: -1263.98  diff: 4.47190      7 iterations  logLik: -1254.03  diff: 9.94912      8 iterations  logLik: -1236.22  diff: 17.80410      9 iterations  logLik: -1214.63  diff: 21.59465      10 iterations  logLik: -1197.76  diff: 16.87469      11 iterations  logLik: -1188.34  diff: 9.41972      12 iterations  logLik: -1183.92  diff: 4.41446      13 iterations  logLik: -1181.89  diff: 2.03582      14 iterations  logLik: -1180.85  diff: 1.03069      15 iterations  logLik: -1180.27  diff: 0.58792      16 iterations  logLik: -1179.90  diff: 0.36633      17 iterations  logLik: -1179.66  diff: 0.23936      18 iterations  logLik: -1179.50  diff: 0.15946      19 iterations  logLik: -1179.39  diff: 0.10676      20 iterations  logLik: -1179.32  diff: 0.07140      21 iterations  logLik: -1179.28  diff: 0.04760      22 iterations  logLik: -1179.24  diff: 0.03161      23 iterations  logLik: -1179.22  diff: 0.02093      24 iterations  logLik: -1179.21  diff: 0.01383      25 iterations  logLik: -1179.20  diff: 0.00913      26 iterations  logLik: -1179.19  diff: 0.00604      27 iterations  logLik: -1179.19  diff: 0.00401      28 iterations  logLik: -1179.19  diff: 0.00268      29 iterations  logLik: -1179.19  diff: 0.00180      30 iterations  logLik: -1179.18  diff: 0.00122      31 iterations  logLik: -1179.18  diff: 0.00084      32 iterations  logLik: -1179.18  diff: 0.00058      33 iterations  logLik: -1179.18  diff: 0.00041      34 iterations  logLik: -1179.18  diff: 0.00029      35 iterations  logLik: -1179.18  diff: 0.00021      36 iterations  logLik: -1179.18  diff: 0.00015      37 iterations  logLik: -1179.18  diff: 0.00011      38 iterations  logLik: -1179.18  diff: 0.00009      39 iterations  logLik: -1179.18  diff: 0.00006      40 iterations  logLik: -1179.18  diff: 0.00005      41 iterations  logLik: -1179.18  diff: 0.00004      42 iterations  logLik: -1179.18  diff: 0.00003      43 iterations  logLik: -1179.18  diff: 0.00002      44 iterations  logLik: -1179.18  diff: 0.00002      45 iterations  logLik: -1179.18  diff: 0.00001      46 iterations  logLik: -1179.18  diff: 0.00001      47 iterations  logLik: -1179.18  diff: 8.0e-06      48 iterations  logLik: -1179.18  diff: 6.3e-06      49 iterations  logLik: -1179.18  diff: 4.9e-06      50 iterations  logLik: -1179.18  diff: 3.9e-06      51 iterations  logLik: -1179.18  diff: 3.0e-06      52 iterations  logLik: -1179.18  diff: 2.4e-06      53 iterations  logLik: -1179.18  diff: 1.9e-06      54 iterations  logLik: -1179.18  diff: 1.5e-06      55 iterations  logLik: -1179.18  diff: 1.2e-06      56 iterations  logLik: -1179.18  diff: 9.3e-07      
#> .. done.
stat3 <- slca(status(3) ~ PAPRES + PADEG + MADEG) %>%
   estimate(data = data)
#> EM iteration begin.
#> 1 iterations  logLik: -1852.68  diff: inf      2 iterations  logLik: -1298.97  diff: 553.71376      3 iterations  logLik: -1273.28  diff: 25.68437      4 iterations  logLik: -1271.52  diff: 1.76571      5 iterations  logLik: -1270.89  diff: 0.62316      6 iterations  logLik: -1269.40  diff: 1.49072      7 iterations  logLik: -1265.40  diff: 3.99834      8 iterations  logLik: -1255.91  diff: 9.49289      9 iterations  logLik: -1238.45  diff: 17.46594      10 iterations  logLik: -1215.64  diff: 22.80804      11 iterations  logLik: -1194.25  diff: 21.38743      12 iterations  logLik: -1178.59  diff: 15.66249      13 iterations  logLik: -1168.35  diff: 10.23963      14 iterations  logLik: -1161.55  diff: 6.80062      15 iterations  logLik: -1156.79  diff: 4.75221      16 iterations  logLik: -1153.41  diff: 3.38024      17 iterations  logLik: -1151.03  diff: 2.38762      18 iterations  logLik: -1149.35  diff: 1.67878      19 iterations  logLik: -1148.15  diff: 1.19707      20 iterations  logLik: -1147.27  diff: 0.87955      21 iterations  logLik: -1146.61  diff: 0.66240      22 iterations  logLik: -1146.11  diff: 0.50012      23 iterations  logLik: -1145.74  diff: 0.37345      24 iterations  logLik: -1145.46  diff: 0.27953      25 iterations  logLik: -1145.24  diff: 0.21780      26 iterations  logLik: -1145.05  diff: 0.18372      27 iterations  logLik: -1144.88  diff: 0.16991      28 iterations  logLik: -1144.72  diff: 0.16921      29 iterations  logLik: -1144.54  diff: 0.17619      30 iterations  logLik: -1144.35  diff: 0.18728      31 iterations  logLik: -1144.15  diff: 0.20021      32 iterations  logLik: -1143.94  diff: 0.21349      33 iterations  logLik: -1143.71  diff: 0.22609      34 iterations  logLik: -1143.47  diff: 0.23726      35 iterations  logLik: -1143.23  diff: 0.24648      36 iterations  logLik: -1142.97  diff: 0.25344      37 iterations  logLik: -1142.72  diff: 0.25798      38 iterations  logLik: -1142.46  diff: 0.26007      39 iterations  logLik: -1142.20  diff: 0.25969      40 iterations  logLik: -1141.94  diff: 0.25682      41 iterations  logLik: -1141.69  diff: 0.25137      42 iterations  logLik: -1141.45  diff: 0.24331      43 iterations  logLik: -1141.21  diff: 0.23265      44 iterations  logLik: -1140.99  diff: 0.21963      45 iterations  logLik: -1140.79  diff: 0.20467      46 iterations  logLik: -1140.60  diff: 0.18834      47 iterations  logLik: -1140.43  diff: 0.17129      48 iterations  logLik: -1140.28  diff: 0.15411      49 iterations  logLik: -1140.14  diff: 0.13732      50 iterations  logLik: -1140.02  diff: 0.12131      51 iterations  logLik: -1139.91  diff: 0.10634      52 iterations  logLik: -1139.82  diff: 0.09259      53 iterations  logLik: -1139.74  diff: 0.08013      54 iterations  logLik: -1139.67  diff: 0.06897      55 iterations  logLik: -1139.61  diff: 0.05909      56 iterations  logLik: -1139.56  diff: 0.05040      57 iterations  logLik: -1139.52  diff: 0.04283      58 iterations  logLik: -1139.48  diff: 0.03627      59 iterations  logLik: -1139.45  diff: 0.03061      60 iterations  logLik: -1139.42  diff: 0.02577      61 iterations  logLik: -1139.40  diff: 0.02163      62 iterations  logLik: -1139.38  diff: 0.01811      63 iterations  logLik: -1139.37  diff: 0.01512      64 iterations  logLik: -1139.36  diff: 0.01260      65 iterations  logLik: -1139.35  diff: 0.01048      66 iterations  logLik: -1139.34  diff: 0.00869      67 iterations  logLik: -1139.33  diff: 0.00720      68 iterations  logLik: -1139.32  diff: 0.00596      69 iterations  logLik: -1139.32  diff: 0.00492      70 iterations  logLik: -1139.31  diff: 0.00406      71 iterations  logLik: -1139.31  diff: 0.00335      72 iterations  logLik: -1139.31  diff: 0.00277      73 iterations  logLik: -1139.31  diff: 0.00228      74 iterations  logLik: -1139.30  diff: 0.00189      75 iterations  logLik: -1139.30  diff: 0.00156      76 iterations  logLik: -1139.30  diff: 0.00130      77 iterations  logLik: -1139.30  diff: 0.00108      78 iterations  logLik: -1139.30  diff: 0.00090      79 iterations  logLik: -1139.30  diff: 0.00075      80 iterations  logLik: -1139.30  diff: 0.00063      81 iterations  logLik: -1139.30  diff: 0.00053      82 iterations  logLik: -1139.30  diff: 0.00045      83 iterations  logLik: -1139.30  diff: 0.00039      84 iterations  logLik: -1139.30  diff: 0.00033      85 iterations  logLik: -1139.30  diff: 0.00028      86 iterations  logLik: -1139.30  diff: 0.00024      87 iterations  logLik: -1139.30  diff: 0.00021      88 iterations  logLik: -1139.30  diff: 0.00019      89 iterations  logLik: -1139.30  diff: 0.00016      90 iterations  logLik: -1139.30  diff: 0.00014      91 iterations  logLik: -1139.30  diff: 0.00013      92 iterations  logLik: -1139.29  diff: 0.00011      93 iterations  logLik: -1139.29  diff: 0.00010      94 iterations  logLik: -1139.29  diff: 0.00009      95 iterations  logLik: -1139.29  diff: 0.00008      96 iterations  logLik: -1139.29  diff: 0.00007      97 iterations  logLik: -1139.29  diff: 0.00007      98 iterations  logLik: -1139.29  diff: 0.00006      99 iterations  logLik: -1139.29  diff: 0.00006      100 iterations  logLik: -1139.29  diff: 0.00005      
#> 101 iterations  logLik: -1139.29  diff: 0.00005      102 iterations  logLik: -1139.29  diff: 0.00004      103 iterations  logLik: -1139.29  diff: 0.00004      104 iterations  logLik: -1139.29  diff: 0.00004      105 iterations  logLik: -1139.29  diff: 0.00003      106 iterations  logLik: -1139.29  diff: 0.00003      107 iterations  logLik: -1139.29  diff: 0.00003      108 iterations  logLik: -1139.29  diff: 0.00003      109 iterations  logLik: -1139.29  diff: 0.00003      110 iterations  logLik: -1139.29  diff: 0.00002      111 iterations  logLik: -1139.29  diff: 0.00002      112 iterations  logLik: -1139.29  diff: 0.00002      113 iterations  logLik: -1139.29  diff: 0.00002      114 iterations  logLik: -1139.29  diff: 0.00002      115 iterations  logLik: -1139.29  diff: 0.00002      116 iterations  logLik: -1139.29  diff: 0.00002      117 iterations  logLik: -1139.29  diff: 0.00002      118 iterations  logLik: -1139.29  diff: 0.00001      119 iterations  logLik: -1139.29  diff: 0.00001      120 iterations  logLik: -1139.29  diff: 0.00001      121 iterations  logLik: -1139.29  diff: 0.00001      122 iterations  logLik: -1139.29  diff: 0.00001      123 iterations  logLik: -1139.29  diff: 0.00001      124 iterations  logLik: -1139.29  diff: 9.9e-06      125 iterations  logLik: -1139.29  diff: 9.4e-06      126 iterations  logLik: -1139.29  diff: 8.9e-06      127 iterations  logLik: -1139.29  diff: 8.4e-06      128 iterations  logLik: -1139.29  diff: 7.9e-06      129 iterations  logLik: -1139.29  diff: 7.5e-06      130 iterations  logLik: -1139.29  diff: 7.1e-06      131 iterations  logLik: -1139.29  diff: 6.7e-06      132 iterations  logLik: -1139.29  diff: 6.4e-06      133 iterations  logLik: -1139.29  diff: 6.1e-06      134 iterations  logLik: -1139.29  diff: 5.8e-06      135 iterations  logLik: -1139.29  diff: 5.5e-06      136 iterations  logLik: -1139.29  diff: 5.2e-06      137 iterations  logLik: -1139.29  diff: 4.9e-06      138 iterations  logLik: -1139.29  diff: 4.7e-06      139 iterations  logLik: -1139.29  diff: 4.4e-06      140 iterations  logLik: -1139.29  diff: 4.2e-06      141 iterations  logLik: -1139.29  diff: 4.0e-06      142 iterations  logLik: -1139.29  diff: 3.8e-06      143 iterations  logLik: -1139.29  diff: 3.6e-06      144 iterations  logLik: -1139.29  diff: 3.5e-06      145 iterations  logLik: -1139.29  diff: 3.3e-06      146 iterations  logLik: -1139.29  diff: 3.1e-06      147 iterations  logLik: -1139.29  diff: 3.0e-06      148 iterations  logLik: -1139.29  diff: 2.8e-06      149 iterations  logLik: -1139.29  diff: 2.7e-06      150 iterations  logLik: -1139.29  diff: 2.6e-06      151 iterations  logLik: -1139.29  diff: 2.5e-06      152 iterations  logLik: -1139.29  diff: 2.4e-06      153 iterations  logLik: -1139.29  diff: 2.2e-06      154 iterations  logLik: -1139.29  diff: 2.1e-06      155 iterations  logLik: -1139.29  diff: 2.0e-06      156 iterations  logLik: -1139.29  diff: 2.0e-06      157 iterations  logLik: -1139.29  diff: 1.9e-06      158 iterations  logLik: -1139.29  diff: 1.8e-06      159 iterations  logLik: -1139.29  diff: 1.7e-06      160 iterations  logLik: -1139.29  diff: 1.6e-06      161 iterations  logLik: -1139.29  diff: 1.5e-06      162 iterations  logLik: -1139.29  diff: 1.5e-06      163 iterations  logLik: -1139.29  diff: 1.4e-06      164 iterations  logLik: -1139.29  diff: 1.4e-06      165 iterations  logLik: -1139.29  diff: 1.3e-06      166 iterations  logLik: -1139.29  diff: 1.2e-06      167 iterations  logLik: -1139.29  diff: 1.2e-06      168 iterations  logLik: -1139.29  diff: 1.1e-06      169 iterations  logLik: -1139.29  diff: 1.1e-06      170 iterations  logLik: -1139.29  diff: 1.0e-06      171 iterations  logLik: -1139.29  diff: 9.9e-07      
#> .. done.
stat4 <- slca(status(4) ~ PAPRES + PADEG + MADEG) %>%
   estimate(data = data)
#> EM iteration begin.
#> 1 iterations  logLik: -1856.45  diff: inf      2 iterations  logLik: -1298.78  diff: 557.67206      3 iterations  logLik: -1273.32  diff: 25.45884      4 iterations  logLik: -1271.69  diff: 1.63464      5 iterations  logLik: -1271.37  diff: 0.32125      6 iterations  logLik: -1270.73  diff: 0.63041      7 iterations  logLik: -1269.10  diff: 1.63360      8 iterations  logLik: -1265.14  diff: 3.96191      9 iterations  logLik: -1256.76  diff: 8.37753      10 iterations  logLik: -1242.17  diff: 14.59379      11 iterations  logLik: -1222.03  diff: 20.13792      12 iterations  logLik: -1200.77  diff: 21.26152      13 iterations  logLik: -1183.26  diff: 17.50882      14 iterations  logLik: -1170.57  diff: 12.69052      15 iterations  logLik: -1161.45  diff: 9.12130      16 iterations  logLik: -1154.93  diff: 6.51745      17 iterations  logLik: -1150.50  diff: 4.43163      18 iterations  logLik: -1147.65  diff: 2.84969      19 iterations  logLik: -1145.86  diff: 1.78426      20 iterations  logLik: -1144.73  diff: 1.13415      21 iterations  logLik: -1143.97  diff: 0.76370      22 iterations  logLik: -1143.40  diff: 0.56381      23 iterations  logLik: -1142.94  diff: 0.46161      24 iterations  logLik: -1142.53  diff: 0.41331      25 iterations  logLik: -1142.13  diff: 0.39358      26 iterations  logLik: -1141.75  diff: 0.38741      27 iterations  logLik: -1141.36  diff: 0.38578      28 iterations  logLik: -1140.98  diff: 0.38341      29 iterations  logLik: -1140.60  diff: 0.37724      30 iterations  logLik: -1140.24  diff: 0.36553      31 iterations  logLik: -1139.89  diff: 0.34757      32 iterations  logLik: -1139.56  diff: 0.32357      33 iterations  logLik: -1139.27  diff: 0.29463      34 iterations  logLik: -1139.01  diff: 0.26250      35 iterations  logLik: -1138.78  diff: 0.22918      36 iterations  logLik: -1138.58  diff: 0.19658      37 iterations  logLik: -1138.42  diff: 0.16620      38 iterations  logLik: -1138.28  diff: 0.13904      39 iterations  logLik: -1138.16  diff: 0.11560      40 iterations  logLik: -1138.06  diff: 0.09595      41 iterations  logLik: -1137.98  diff: 0.07991      42 iterations  logLik: -1137.92  diff: 0.06711      43 iterations  logLik: -1137.86  diff: 0.05713      44 iterations  logLik: -1137.81  diff: 0.04950      45 iterations  logLik: -1137.77  diff: 0.04382      46 iterations  logLik: -1137.73  diff: 0.03969      47 iterations  logLik: -1137.69  diff: 0.03676      48 iterations  logLik: -1137.66  diff: 0.03473      49 iterations  logLik: -1137.62  diff: 0.03334      50 iterations  logLik: -1137.59  diff: 0.03236      51 iterations  logLik: -1137.56  diff: 0.03161      52 iterations  logLik: -1137.53  diff: 0.03097      53 iterations  logLik: -1137.50  diff: 0.03034      54 iterations  logLik: -1137.47  diff: 0.02967      55 iterations  logLik: -1137.44  diff: 0.02895      56 iterations  logLik: -1137.41  diff: 0.02818      57 iterations  logLik: -1137.38  diff: 0.02739      58 iterations  logLik: -1137.36  diff: 0.02659      59 iterations  logLik: -1137.33  diff: 0.02581      60 iterations  logLik: -1137.31  diff: 0.02508      61 iterations  logLik: -1137.28  diff: 0.02441      62 iterations  logLik: -1137.26  diff: 0.02381      63 iterations  logLik: -1137.23  diff: 0.02330      64 iterations  logLik: -1137.21  diff: 0.02288      65 iterations  logLik: -1137.19  diff: 0.02255      66 iterations  logLik: -1137.17  diff: 0.02231      67 iterations  logLik: -1137.14  diff: 0.02215      68 iterations  logLik: -1137.12  diff: 0.02209      69 iterations  logLik: -1137.10  diff: 0.02211      70 iterations  logLik: -1137.08  diff: 0.02220      71 iterations  logLik: -1137.06  diff: 0.02236      72 iterations  logLik: -1137.03  diff: 0.02258      73 iterations  logLik: -1137.01  diff: 0.02285      74 iterations  logLik: -1136.99  diff: 0.02315      75 iterations  logLik: -1136.96  diff: 0.02347      76 iterations  logLik: -1136.94  diff: 0.02380      77 iterations  logLik: -1136.92  diff: 0.02411      78 iterations  logLik: -1136.89  diff: 0.02441      79 iterations  logLik: -1136.87  diff: 0.02466      80 iterations  logLik: -1136.84  diff: 0.02486      81 iterations  logLik: -1136.82  diff: 0.02499      82 iterations  logLik: -1136.79  diff: 0.02503      83 iterations  logLik: -1136.77  diff: 0.02497      84 iterations  logLik: -1136.74  diff: 0.02481      85 iterations  logLik: -1136.72  diff: 0.02455      86 iterations  logLik: -1136.69  diff: 0.02418      87 iterations  logLik: -1136.67  diff: 0.02370      88 iterations  logLik: -1136.65  diff: 0.02314      89 iterations  logLik: -1136.62  diff: 0.02250      90 iterations  logLik: -1136.60  diff: 0.02180      91 iterations  logLik: -1136.58  diff: 0.02107      92 iterations  logLik: -1136.56  diff: 0.02032      93 iterations  logLik: -1136.54  diff: 0.01958      94 iterations  logLik: -1136.52  diff: 0.01885      95 iterations  logLik: -1136.50  diff: 0.01816      96 iterations  logLik: -1136.49  diff: 0.01751      97 iterations  logLik: -1136.47  diff: 0.01690      98 iterations  logLik: -1136.45  diff: 0.01635      99 iterations  logLik: -1136.44  diff: 0.01584      100 iterations  logLik: -1136.42  diff: 0.01538      
#> 101 iterations  logLik: -1136.41  diff: 0.01497      102 iterations  logLik: -1136.39  diff: 0.01459      103 iterations  logLik: -1136.38  diff: 0.01425      104 iterations  logLik: -1136.36  diff: 0.01393      105 iterations  logLik: -1136.35  diff: 0.01364      106 iterations  logLik: -1136.34  diff: 0.01338      107 iterations  logLik: -1136.32  diff: 0.01313      108 iterations  logLik: -1136.31  diff: 0.01290      109 iterations  logLik: -1136.30  diff: 0.01268      110 iterations  logLik: -1136.29  diff: 0.01248      111 iterations  logLik: -1136.27  diff: 0.01228      112 iterations  logLik: -1136.26  diff: 0.01209      113 iterations  logLik: -1136.25  diff: 0.01191      114 iterations  logLik: -1136.24  diff: 0.01174      115 iterations  logLik: -1136.23  diff: 0.01157      116 iterations  logLik: -1136.21  diff: 0.01141      117 iterations  logLik: -1136.20  diff: 0.01125      118 iterations  logLik: -1136.19  diff: 0.01109      119 iterations  logLik: -1136.18  diff: 0.01094      120 iterations  logLik: -1136.17  diff: 0.01079      121 iterations  logLik: -1136.16  diff: 0.01064      122 iterations  logLik: -1136.15  diff: 0.01050      123 iterations  logLik: -1136.14  diff: 0.01036      124 iterations  logLik: -1136.13  diff: 0.01022      125 iterations  logLik: -1136.12  diff: 0.01009      126 iterations  logLik: -1136.11  diff: 0.00995      127 iterations  logLik: -1136.10  diff: 0.00982      128 iterations  logLik: -1136.09  diff: 0.00969      129 iterations  logLik: -1136.08  diff: 0.00956      130 iterations  logLik: -1136.07  diff: 0.00944      131 iterations  logLik: -1136.06  diff: 0.00931      132 iterations  logLik: -1136.05  diff: 0.00919      133 iterations  logLik: -1136.04  diff: 0.00907      134 iterations  logLik: -1136.03  diff: 0.00895      135 iterations  logLik: -1136.03  diff: 0.00883      136 iterations  logLik: -1136.02  diff: 0.00871      137 iterations  logLik: -1136.01  diff: 0.00860      138 iterations  logLik: -1136.00  diff: 0.00848      139 iterations  logLik: -1135.99  diff: 0.00837      140 iterations  logLik: -1135.98  diff: 0.00826      141 iterations  logLik: -1135.97  diff: 0.00815      142 iterations  logLik: -1135.97  diff: 0.00804      143 iterations  logLik: -1135.96  diff: 0.00793      144 iterations  logLik: -1135.95  diff: 0.00782      145 iterations  logLik: -1135.94  diff: 0.00771      146 iterations  logLik: -1135.94  diff: 0.00761      147 iterations  logLik: -1135.93  diff: 0.00751      148 iterations  logLik: -1135.92  diff: 0.00740      149 iterations  logLik: -1135.91  diff: 0.00730      150 iterations  logLik: -1135.91  diff: 0.00720      151 iterations  logLik: -1135.90  diff: 0.00710      152 iterations  logLik: -1135.89  diff: 0.00700      153 iterations  logLik: -1135.89  diff: 0.00690      154 iterations  logLik: -1135.88  diff: 0.00681      155 iterations  logLik: -1135.87  diff: 0.00671      156 iterations  logLik: -1135.87  diff: 0.00662      157 iterations  logLik: -1135.86  diff: 0.00652      158 iterations  logLik: -1135.85  diff: 0.00643      159 iterations  logLik: -1135.85  diff: 0.00634      160 iterations  logLik: -1135.84  diff: 0.00625      161 iterations  logLik: -1135.83  diff: 0.00616      162 iterations  logLik: -1135.83  diff: 0.00607      163 iterations  logLik: -1135.82  diff: 0.00599      164 iterations  logLik: -1135.82  diff: 0.00590      165 iterations  logLik: -1135.81  diff: 0.00581      166 iterations  logLik: -1135.80  diff: 0.00573      167 iterations  logLik: -1135.80  diff: 0.00565      168 iterations  logLik: -1135.79  diff: 0.00556      169 iterations  logLik: -1135.79  diff: 0.00548      170 iterations  logLik: -1135.78  diff: 0.00540      171 iterations  logLik: -1135.78  diff: 0.00532      172 iterations  logLik: -1135.77  diff: 0.00524      173 iterations  logLik: -1135.77  diff: 0.00516      174 iterations  logLik: -1135.76  diff: 0.00509      175 iterations  logLik: -1135.76  diff: 0.00501      176 iterations  logLik: -1135.75  diff: 0.00493      177 iterations  logLik: -1135.75  diff: 0.00486      178 iterations  logLik: -1135.74  diff: 0.00478      179 iterations  logLik: -1135.74  diff: 0.00471      180 iterations  logLik: -1135.73  diff: 0.00464      181 iterations  logLik: -1135.73  diff: 0.00457      182 iterations  logLik: -1135.72  diff: 0.00450      183 iterations  logLik: -1135.72  diff: 0.00443      184 iterations  logLik: -1135.71  diff: 0.00436      185 iterations  logLik: -1135.71  diff: 0.00429      186 iterations  logLik: -1135.71  diff: 0.00422      187 iterations  logLik: -1135.70  diff: 0.00415      188 iterations  logLik: -1135.70  diff: 0.00408      189 iterations  logLik: -1135.69  diff: 0.00402      190 iterations  logLik: -1135.69  diff: 0.00395      191 iterations  logLik: -1135.69  diff: 0.00389      192 iterations  logLik: -1135.68  diff: 0.00382      193 iterations  logLik: -1135.68  diff: 0.00376      194 iterations  logLik: -1135.67  diff: 0.00369      195 iterations  logLik: -1135.67  diff: 0.00363      196 iterations  logLik: -1135.67  diff: 0.00357      197 iterations  logLik: -1135.66  diff: 0.00351      198 iterations  logLik: -1135.66  diff: 0.00345      199 iterations  logLik: -1135.66  diff: 0.00339      200 iterations  logLik: -1135.65  diff: 0.00333      
#> 201 iterations  logLik: -1135.65  diff: 0.00327      202 iterations  logLik: -1135.65  diff: 0.00321      203 iterations  logLik: -1135.64  diff: 0.00315      204 iterations  logLik: -1135.64  diff: 0.00309      205 iterations  logLik: -1135.64  diff: 0.00304      206 iterations  logLik: -1135.63  diff: 0.00298      207 iterations  logLik: -1135.63  diff: 0.00293      208 iterations  logLik: -1135.63  diff: 0.00287      209 iterations  logLik: -1135.63  diff: 0.00282      210 iterations  logLik: -1135.62  diff: 0.00276      211 iterations  logLik: -1135.62  diff: 0.00271      212 iterations  logLik: -1135.62  diff: 0.00266      213 iterations  logLik: -1135.62  diff: 0.00261      214 iterations  logLik: -1135.61  diff: 0.00255      215 iterations  logLik: -1135.61  diff: 0.00250      216 iterations  logLik: -1135.61  diff: 0.00245      217 iterations  logLik: -1135.61  diff: 0.00240      218 iterations  logLik: -1135.60  diff: 0.00236      219 iterations  logLik: -1135.60  diff: 0.00231      220 iterations  logLik: -1135.60  diff: 0.00226      221 iterations  logLik: -1135.60  diff: 0.00221      222 iterations  logLik: -1135.59  diff: 0.00217      223 iterations  logLik: -1135.59  diff: 0.00212      224 iterations  logLik: -1135.59  diff: 0.00208      225 iterations  logLik: -1135.59  diff: 0.00203      226 iterations  logLik: -1135.59  diff: 0.00199      227 iterations  logLik: -1135.58  diff: 0.00194      228 iterations  logLik: -1135.58  diff: 0.00190      229 iterations  logLik: -1135.58  diff: 0.00186      230 iterations  logLik: -1135.58  diff: 0.00182      231 iterations  logLik: -1135.58  diff: 0.00178      232 iterations  logLik: -1135.57  diff: 0.00174      233 iterations  logLik: -1135.57  diff: 0.00170      234 iterations  logLik: -1135.57  diff: 0.00166      235 iterations  logLik: -1135.57  diff: 0.00162      236 iterations  logLik: -1135.57  diff: 0.00159      237 iterations  logLik: -1135.57  diff: 0.00155      238 iterations  logLik: -1135.57  diff: 0.00151      239 iterations  logLik: -1135.56  diff: 0.00148      240 iterations  logLik: -1135.56  diff: 0.00144      241 iterations  logLik: -1135.56  diff: 0.00141      242 iterations  logLik: -1135.56  diff: 0.00137      243 iterations  logLik: -1135.56  diff: 0.00134      244 iterations  logLik: -1135.56  diff: 0.00131      245 iterations  logLik: -1135.56  diff: 0.00128      246 iterations  logLik: -1135.55  diff: 0.00125      247 iterations  logLik: -1135.55  diff: 0.00122      248 iterations  logLik: -1135.55  diff: 0.00119      249 iterations  logLik: -1135.55  diff: 0.00116      250 iterations  logLik: -1135.55  diff: 0.00113      251 iterations  logLik: -1135.55  diff: 0.00110      252 iterations  logLik: -1135.55  diff: 0.00107      253 iterations  logLik: -1135.55  diff: 0.00105      254 iterations  logLik: -1135.55  diff: 0.00102      255 iterations  logLik: -1135.54  diff: 0.00099      256 iterations  logLik: -1135.54  diff: 0.00097      257 iterations  logLik: -1135.54  diff: 0.00094      258 iterations  logLik: -1135.54  diff: 0.00092      259 iterations  logLik: -1135.54  diff: 0.00090      260 iterations  logLik: -1135.54  diff: 0.00087      261 iterations  logLik: -1135.54  diff: 0.00085      262 iterations  logLik: -1135.54  diff: 0.00083      263 iterations  logLik: -1135.54  diff: 0.00081      264 iterations  logLik: -1135.54  diff: 0.00079      265 iterations  logLik: -1135.54  diff: 0.00076      266 iterations  logLik: -1135.54  diff: 0.00074      267 iterations  logLik: -1135.53  diff: 0.00073      268 iterations  logLik: -1135.53  diff: 0.00071      269 iterations  logLik: -1135.53  diff: 0.00069      270 iterations  logLik: -1135.53  diff: 0.00067      271 iterations  logLik: -1135.53  diff: 0.00065      272 iterations  logLik: -1135.53  diff: 0.00063      273 iterations  logLik: -1135.53  diff: 0.00062      274 iterations  logLik: -1135.53  diff: 0.00060      275 iterations  logLik: -1135.53  diff: 0.00058      276 iterations  logLik: -1135.53  diff: 0.00057      277 iterations  logLik: -1135.53  diff: 0.00055      278 iterations  logLik: -1135.53  diff: 0.00054      279 iterations  logLik: -1135.53  diff: 0.00052      280 iterations  logLik: -1135.53  diff: 0.00051      281 iterations  logLik: -1135.53  diff: 0.00049      282 iterations  logLik: -1135.53  diff: 0.00048      283 iterations  logLik: -1135.53  diff: 0.00047      284 iterations  logLik: -1135.52  diff: 0.00046      285 iterations  logLik: -1135.52  diff: 0.00044      286 iterations  logLik: -1135.52  diff: 0.00043      287 iterations  logLik: -1135.52  diff: 0.00042      288 iterations  logLik: -1135.52  diff: 0.00041      289 iterations  logLik: -1135.52  diff: 0.00040      290 iterations  logLik: -1135.52  diff: 0.00038      291 iterations  logLik: -1135.52  diff: 0.00037      292 iterations  logLik: -1135.52  diff: 0.00036      293 iterations  logLik: -1135.52  diff: 0.00035      294 iterations  logLik: -1135.52  diff: 0.00034      295 iterations  logLik: -1135.52  diff: 0.00033      296 iterations  logLik: -1135.52  diff: 0.00032      297 iterations  logLik: -1135.52  diff: 0.00032      298 iterations  logLik: -1135.52  diff: 0.00031      299 iterations  logLik: -1135.52  diff: 0.00030      300 iterations  logLik: -1135.52  diff: 0.00029      
#> 301 iterations  logLik: -1135.52  diff: 0.00028      302 iterations  logLik: -1135.52  diff: 0.00027      303 iterations  logLik: -1135.52  diff: 0.00027      304 iterations  logLik: -1135.52  diff: 0.00026      305 iterations  logLik: -1135.52  diff: 0.00025      306 iterations  logLik: -1135.52  diff: 0.00024      307 iterations  logLik: -1135.52  diff: 0.00024      308 iterations  logLik: -1135.52  diff: 0.00023      309 iterations  logLik: -1135.52  diff: 0.00022      310 iterations  logLik: -1135.52  diff: 0.00022      311 iterations  logLik: -1135.52  diff: 0.00021      312 iterations  logLik: -1135.52  diff: 0.00021      313 iterations  logLik: -1135.52  diff: 0.00020      314 iterations  logLik: -1135.52  diff: 0.00019      315 iterations  logLik: -1135.52  diff: 0.00019      316 iterations  logLik: -1135.52  diff: 0.00018      317 iterations  logLik: -1135.51  diff: 0.00018      318 iterations  logLik: -1135.51  diff: 0.00017      319 iterations  logLik: -1135.51  diff: 0.00017      320 iterations  logLik: -1135.51  diff: 0.00016      321 iterations  logLik: -1135.51  diff: 0.00016      322 iterations  logLik: -1135.51  diff: 0.00015      323 iterations  logLik: -1135.51  diff: 0.00015      324 iterations  logLik: -1135.51  diff: 0.00015      325 iterations  logLik: -1135.51  diff: 0.00014      326 iterations  logLik: -1135.51  diff: 0.00014      327 iterations  logLik: -1135.51  diff: 0.00013      328 iterations  logLik: -1135.51  diff: 0.00013      329 iterations  logLik: -1135.51  diff: 0.00013      330 iterations  logLik: -1135.51  diff: 0.00012      331 iterations  logLik: -1135.51  diff: 0.00012      332 iterations  logLik: -1135.51  diff: 0.00012      333 iterations  logLik: -1135.51  diff: 0.00011      334 iterations  logLik: -1135.51  diff: 0.00011      335 iterations  logLik: -1135.51  diff: 0.00011      336 iterations  logLik: -1135.51  diff: 0.00010      337 iterations  logLik: -1135.51  diff: 0.00010      338 iterations  logLik: -1135.51  diff: 0.00010      339 iterations  logLik: -1135.51  diff: 0.00009      340 iterations  logLik: -1135.51  diff: 0.00009      341 iterations  logLik: -1135.51  diff: 0.00009      342 iterations  logLik: -1135.51  diff: 0.00009      343 iterations  logLik: -1135.51  diff: 0.00008      344 iterations  logLik: -1135.51  diff: 0.00008      345 iterations  logLik: -1135.51  diff: 0.00008      346 iterations  logLik: -1135.51  diff: 0.00008      347 iterations  logLik: -1135.51  diff: 0.00008      348 iterations  logLik: -1135.51  diff: 0.00007      349 iterations  logLik: -1135.51  diff: 0.00007      350 iterations  logLik: -1135.51  diff: 0.00007      351 iterations  logLik: -1135.51  diff: 0.00007      352 iterations  logLik: -1135.51  diff: 0.00007      353 iterations  logLik: -1135.51  diff: 0.00006      354 iterations  logLik: -1135.51  diff: 0.00006      355 iterations  logLik: -1135.51  diff: 0.00006      356 iterations  logLik: -1135.51  diff: 0.00006      357 iterations  logLik: -1135.51  diff: 0.00006      358 iterations  logLik: -1135.51  diff: 0.00005      359 iterations  logLik: -1135.51  diff: 0.00005      360 iterations  logLik: -1135.51  diff: 0.00005      361 iterations  logLik: -1135.51  diff: 0.00005      362 iterations  logLik: -1135.51  diff: 0.00005      363 iterations  logLik: -1135.51  diff: 0.00005      364 iterations  logLik: -1135.51  diff: 0.00005      365 iterations  logLik: -1135.51  diff: 0.00004      366 iterations  logLik: -1135.51  diff: 0.00004      367 iterations  logLik: -1135.51  diff: 0.00004      368 iterations  logLik: -1135.51  diff: 0.00004      369 iterations  logLik: -1135.51  diff: 0.00004      370 iterations  logLik: -1135.51  diff: 0.00004      371 iterations  logLik: -1135.51  diff: 0.00004      372 iterations  logLik: -1135.51  diff: 0.00004      373 iterations  logLik: -1135.51  diff: 0.00004      374 iterations  logLik: -1135.51  diff: 0.00003      375 iterations  logLik: -1135.51  diff: 0.00003      376 iterations  logLik: -1135.51  diff: 0.00003      377 iterations  logLik: -1135.51  diff: 0.00003      378 iterations  logLik: -1135.51  diff: 0.00003      379 iterations  logLik: -1135.51  diff: 0.00003      380 iterations  logLik: -1135.51  diff: 0.00003      381 iterations  logLik: -1135.51  diff: 0.00003      382 iterations  logLik: -1135.51  diff: 0.00003      383 iterations  logLik: -1135.51  diff: 0.00003      384 iterations  logLik: -1135.51  diff: 0.00003      385 iterations  logLik: -1135.51  diff: 0.00003      386 iterations  logLik: -1135.51  diff: 0.00002      387 iterations  logLik: -1135.51  diff: 0.00002      388 iterations  logLik: -1135.51  diff: 0.00002      389 iterations  logLik: -1135.51  diff: 0.00002      390 iterations  logLik: -1135.51  diff: 0.00002      391 iterations  logLik: -1135.51  diff: 0.00002      392 iterations  logLik: -1135.51  diff: 0.00002      393 iterations  logLik: -1135.51  diff: 0.00002      394 iterations  logLik: -1135.51  diff: 0.00002      395 iterations  logLik: -1135.51  diff: 0.00002      396 iterations  logLik: -1135.51  diff: 0.00002      397 iterations  logLik: -1135.51  diff: 0.00002      398 iterations  logLik: -1135.51  diff: 0.00002      399 iterations  logLik: -1135.51  diff: 0.00002      400 iterations  logLik: -1135.51  diff: 0.00002      
#> 401 iterations  logLik: -1135.51  diff: 0.00002      402 iterations  logLik: -1135.51  diff: 0.00002      403 iterations  logLik: -1135.51  diff: 0.00002      404 iterations  logLik: -1135.51  diff: 0.00002      405 iterations  logLik: -1135.51  diff: 0.00001      406 iterations  logLik: -1135.51  diff: 0.00001      407 iterations  logLik: -1135.51  diff: 0.00001      408 iterations  logLik: -1135.51  diff: 0.00001      409 iterations  logLik: -1135.51  diff: 0.00001      410 iterations  logLik: -1135.51  diff: 0.00001      411 iterations  logLik: -1135.51  diff: 0.00001      412 iterations  logLik: -1135.51  diff: 0.00001      413 iterations  logLik: -1135.51  diff: 0.00001      414 iterations  logLik: -1135.51  diff: 0.00001      415 iterations  logLik: -1135.51  diff: 0.00001      416 iterations  logLik: -1135.51  diff: 0.00001      417 iterations  logLik: -1135.51  diff: 0.00001      418 iterations  logLik: -1135.51  diff: 0.00001      419 iterations  logLik: -1135.51  diff: 0.00001      420 iterations  logLik: -1135.51  diff: 9.8e-06      421 iterations  logLik: -1135.51  diff: 9.6e-06      422 iterations  logLik: -1135.51  diff: 9.3e-06      423 iterations  logLik: -1135.51  diff: 9.1e-06      424 iterations  logLik: -1135.51  diff: 8.9e-06      425 iterations  logLik: -1135.51  diff: 8.6e-06      426 iterations  logLik: -1135.51  diff: 8.4e-06      427 iterations  logLik: -1135.51  diff: 8.2e-06      428 iterations  logLik: -1135.51  diff: 8.0e-06      429 iterations  logLik: -1135.51  diff: 7.8e-06      430 iterations  logLik: -1135.51  diff: 7.6e-06      431 iterations  logLik: -1135.51  diff: 7.4e-06      432 iterations  logLik: -1135.51  diff: 7.2e-06      433 iterations  logLik: -1135.51  diff: 7.1e-06      434 iterations  logLik: -1135.51  diff: 6.9e-06      435 iterations  logLik: -1135.51  diff: 6.7e-06      436 iterations  logLik: -1135.51  diff: 6.6e-06      437 iterations  logLik: -1135.51  diff: 6.4e-06      438 iterations  logLik: -1135.51  diff: 6.2e-06      439 iterations  logLik: -1135.51  diff: 6.1e-06      440 iterations  logLik: -1135.51  diff: 5.9e-06      441 iterations  logLik: -1135.51  diff: 5.8e-06      442 iterations  logLik: -1135.51  diff: 5.7e-06      443 iterations  logLik: -1135.51  diff: 5.5e-06      444 iterations  logLik: -1135.51  diff: 5.4e-06      445 iterations  logLik: -1135.51  diff: 5.3e-06      446 iterations  logLik: -1135.51  diff: 5.2e-06      447 iterations  logLik: -1135.51  diff: 5.0e-06      448 iterations  logLik: -1135.51  diff: 4.9e-06      449 iterations  logLik: -1135.51  diff: 4.8e-06      450 iterations  logLik: -1135.51  diff: 4.7e-06      451 iterations  logLik: -1135.51  diff: 4.6e-06      452 iterations  logLik: -1135.51  diff: 4.5e-06      453 iterations  logLik: -1135.51  diff: 4.4e-06      454 iterations  logLik: -1135.51  diff: 4.3e-06      455 iterations  logLik: -1135.51  diff: 4.2e-06      456 iterations  logLik: -1135.51  diff: 4.1e-06      457 iterations  logLik: -1135.51  diff: 4.0e-06      458 iterations  logLik: -1135.51  diff: 4.0e-06      459 iterations  logLik: -1135.51  diff: 3.9e-06      460 iterations  logLik: -1135.51  diff: 3.8e-06      461 iterations  logLik: -1135.51  diff: 3.7e-06      462 iterations  logLik: -1135.51  diff: 3.6e-06      463 iterations  logLik: -1135.51  diff: 3.6e-06      464 iterations  logLik: -1135.51  diff: 3.5e-06      465 iterations  logLik: -1135.51  diff: 3.4e-06      466 iterations  logLik: -1135.51  diff: 3.4e-06      467 iterations  logLik: -1135.51  diff: 3.3e-06      468 iterations  logLik: -1135.51  diff: 3.2e-06      469 iterations  logLik: -1135.51  diff: 3.2e-06      470 iterations  logLik: -1135.51  diff: 3.1e-06      471 iterations  logLik: -1135.51  diff: 3.0e-06      472 iterations  logLik: -1135.51  diff: 3.0e-06      473 iterations  logLik: -1135.51  diff: 2.9e-06      474 iterations  logLik: -1135.51  diff: 2.9e-06      475 iterations  logLik: -1135.51  diff: 2.8e-06      476 iterations  logLik: -1135.51  diff: 2.8e-06      477 iterations  logLik: -1135.51  diff: 2.7e-06      478 iterations  logLik: -1135.51  diff: 2.7e-06      479 iterations  logLik: -1135.51  diff: 2.6e-06      480 iterations  logLik: -1135.51  diff: 2.6e-06      481 iterations  logLik: -1135.51  diff: 2.6e-06      482 iterations  logLik: -1135.51  diff: 2.5e-06      483 iterations  logLik: -1135.51  diff: 2.5e-06      484 iterations  logLik: -1135.51  diff: 2.4e-06      485 iterations  logLik: -1135.51  diff: 2.4e-06      486 iterations  logLik: -1135.51  diff: 2.4e-06      487 iterations  logLik: -1135.51  diff: 2.3e-06      488 iterations  logLik: -1135.51  diff: 2.3e-06      489 iterations  logLik: -1135.51  diff: 2.2e-06      490 iterations  logLik: -1135.51  diff: 2.2e-06      491 iterations  logLik: -1135.51  diff: 2.2e-06      492 iterations  logLik: -1135.51  diff: 2.2e-06      493 iterations  logLik: -1135.51  diff: 2.1e-06      494 iterations  logLik: -1135.51  diff: 2.1e-06      495 iterations  logLik: -1135.51  diff: 2.1e-06      496 iterations  logLik: -1135.51  diff: 2.0e-06      497 iterations  logLik: -1135.51  diff: 2.0e-06      498 iterations  logLik: -1135.51  diff: 2.0e-06      499 iterations  logLik: -1135.51  diff: 2.0e-06      500 iterations  logLik: -1135.51  diff: 1.9e-06      
#> 501 iterations  logLik: -1135.51  diff: 1.9e-06      502 iterations  logLik: -1135.51  diff: 1.9e-06      503 iterations  logLik: -1135.51  diff: 1.9e-06      504 iterations  logLik: -1135.51  diff: 1.8e-06      505 iterations  logLik: -1135.51  diff: 1.8e-06      506 iterations  logLik: -1135.51  diff: 1.8e-06      507 iterations  logLik: -1135.51  diff: 1.8e-06      508 iterations  logLik: -1135.51  diff: 1.8e-06      509 iterations  logLik: -1135.51  diff: 1.7e-06      510 iterations  logLik: -1135.51  diff: 1.7e-06      511 iterations  logLik: -1135.51  diff: 1.7e-06      512 iterations  logLik: -1135.51  diff: 1.7e-06      513 iterations  logLik: -1135.51  diff: 1.7e-06      514 iterations  logLik: -1135.51  diff: 1.7e-06      515 iterations  logLik: -1135.51  diff: 1.6e-06      516 iterations  logLik: -1135.51  diff: 1.6e-06      517 iterations  logLik: -1135.51  diff: 1.6e-06      518 iterations  logLik: -1135.51  diff: 1.6e-06      519 iterations  logLik: -1135.51  diff: 1.6e-06      520 iterations  logLik: -1135.51  diff: 1.6e-06      521 iterations  logLik: -1135.51  diff: 1.6e-06      522 iterations  logLik: -1135.51  diff: 1.5e-06      523 iterations  logLik: -1135.51  diff: 1.5e-06      524 iterations  logLik: -1135.51  diff: 1.5e-06      525 iterations  logLik: -1135.51  diff: 1.5e-06      526 iterations  logLik: -1135.51  diff: 1.5e-06      527 iterations  logLik: -1135.51  diff: 1.5e-06      528 iterations  logLik: -1135.51  diff: 1.5e-06      529 iterations  logLik: -1135.51  diff: 1.5e-06      530 iterations  logLik: -1135.51  diff: 1.5e-06      531 iterations  logLik: -1135.51  diff: 1.4e-06      532 iterations  logLik: -1135.51  diff: 1.4e-06      533 iterations  logLik: -1135.51  diff: 1.4e-06      534 iterations  logLik: -1135.51  diff: 1.4e-06      535 iterations  logLik: -1135.51  diff: 1.4e-06      536 iterations  logLik: -1135.51  diff: 1.4e-06      537 iterations  logLik: -1135.51  diff: 1.4e-06      538 iterations  logLik: -1135.51  diff: 1.4e-06      539 iterations  logLik: -1135.51  diff: 1.4e-06      540 iterations  logLik: -1135.51  diff: 1.4e-06      541 iterations  logLik: -1135.51  diff: 1.4e-06      542 iterations  logLik: -1135.51  diff: 1.4e-06      543 iterations  logLik: -1135.51  diff: 1.3e-06      544 iterations  logLik: -1135.51  diff: 1.3e-06      545 iterations  logLik: -1135.51  diff: 1.3e-06      546 iterations  logLik: -1135.51  diff: 1.3e-06      547 iterations  logLik: -1135.51  diff: 1.3e-06      548 iterations  logLik: -1135.51  diff: 1.3e-06      549 iterations  logLik: -1135.51  diff: 1.3e-06      550 iterations  logLik: -1135.51  diff: 1.3e-06      551 iterations  logLik: -1135.51  diff: 1.3e-06      552 iterations  logLik: -1135.51  diff: 1.3e-06      553 iterations  logLik: -1135.51  diff: 1.3e-06      554 iterations  logLik: -1135.51  diff: 1.3e-06      555 iterations  logLik: -1135.51  diff: 1.3e-06      556 iterations  logLik: -1135.51  diff: 1.3e-06      557 iterations  logLik: -1135.51  diff: 1.3e-06      558 iterations  logLik: -1135.51  diff: 1.3e-06      559 iterations  logLik: -1135.51  diff: 1.3e-06      560 iterations  logLik: -1135.51  diff: 1.3e-06      561 iterations  logLik: -1135.51  diff: 1.3e-06      562 iterations  logLik: -1135.51  diff: 1.2e-06      563 iterations  logLik: -1135.51  diff: 1.2e-06      564 iterations  logLik: -1135.51  diff: 1.2e-06      565 iterations  logLik: -1135.51  diff: 1.2e-06      566 iterations  logLik: -1135.51  diff: 1.2e-06      567 iterations  logLik: -1135.51  diff: 1.2e-06      568 iterations  logLik: -1135.51  diff: 1.2e-06      569 iterations  logLik: -1135.51  diff: 1.2e-06      570 iterations  logLik: -1135.51  diff: 1.2e-06      571 iterations  logLik: -1135.51  diff: 1.2e-06      572 iterations  logLik: -1135.51  diff: 1.2e-06      573 iterations  logLik: -1135.51  diff: 1.2e-06      574 iterations  logLik: -1135.51  diff: 1.2e-06      575 iterations  logLik: -1135.51  diff: 1.2e-06      576 iterations  logLik: -1135.51  diff: 1.2e-06      577 iterations  logLik: -1135.51  diff: 1.2e-06      578 iterations  logLik: -1135.51  diff: 1.2e-06      579 iterations  logLik: -1135.51  diff: 1.2e-06      580 iterations  logLik: -1135.51  diff: 1.2e-06      581 iterations  logLik: -1135.51  diff: 1.2e-06      582 iterations  logLik: -1135.51  diff: 1.2e-06      583 iterations  logLik: -1135.51  diff: 1.2e-06      584 iterations  logLik: -1135.51  diff: 1.2e-06      585 iterations  logLik: -1135.51  diff: 1.2e-06      586 iterations  logLik: -1135.51  diff: 1.2e-06      587 iterations  logLik: -1135.51  diff: 1.2e-06      588 iterations  logLik: -1135.51  diff: 1.2e-06      589 iterations  logLik: -1135.51  diff: 1.2e-06      590 iterations  logLik: -1135.51  diff: 1.2e-06      591 iterations  logLik: -1135.51  diff: 1.2e-06      592 iterations  logLik: -1135.51  diff: 1.2e-06      593 iterations  logLik: -1135.51  diff: 1.2e-06      594 iterations  logLik: -1135.51  diff: 1.2e-06      595 iterations  logLik: -1135.51  diff: 1.2e-06      596 iterations  logLik: -1135.51  diff: 1.2e-06      597 iterations  logLik: -1135.51  diff: 1.2e-06      598 iterations  logLik: -1135.51  diff: 1.2e-06      599 iterations  logLik: -1135.51  diff: 1.2e-06      600 iterations  logLik: -1135.51  diff: 1.2e-06      
#> 601 iterations  logLik: -1135.51  diff: 1.2e-06      602 iterations  logLik: -1135.51  diff: 1.2e-06      603 iterations  logLik: -1135.51  diff: 1.2e-06      604 iterations  logLik: -1135.51  diff: 1.2e-06      605 iterations  logLik: -1135.51  diff: 1.1e-06      606 iterations  logLik: -1135.51  diff: 1.1e-06      607 iterations  logLik: -1135.51  diff: 1.1e-06      608 iterations  logLik: -1135.51  diff: 1.1e-06      609 iterations  logLik: -1135.51  diff: 1.1e-06      610 iterations  logLik: -1135.51  diff: 1.1e-06      611 iterations  logLik: -1135.51  diff: 1.1e-06      612 iterations  logLik: -1135.51  diff: 1.1e-06      613 iterations  logLik: -1135.51  diff: 1.1e-06      614 iterations  logLik: -1135.51  diff: 1.1e-06      615 iterations  logLik: -1135.51  diff: 1.1e-06      616 iterations  logLik: -1135.51  diff: 1.1e-06      617 iterations  logLik: -1135.51  diff: 1.1e-06      618 iterations  logLik: -1135.51  diff: 1.1e-06      619 iterations  logLik: -1135.51  diff: 1.1e-06      620 iterations  logLik: -1135.51  diff: 1.1e-06      621 iterations  logLik: -1135.51  diff: 1.1e-06      622 iterations  logLik: -1135.51  diff: 1.1e-06      623 iterations  logLik: -1135.51  diff: 1.1e-06      624 iterations  logLik: -1135.51  diff: 1.1e-06      625 iterations  logLik: -1135.51  diff: 1.1e-06      626 iterations  logLik: -1135.51  diff: 1.1e-06      627 iterations  logLik: -1135.51  diff: 1.1e-06      628 iterations  logLik: -1135.51  diff: 1.1e-06      629 iterations  logLik: -1135.51  diff: 1.1e-06      630 iterations  logLik: -1135.51  diff: 1.1e-06      631 iterations  logLik: -1135.51  diff: 1.1e-06      632 iterations  logLik: -1135.51  diff: 1.1e-06      633 iterations  logLik: -1135.51  diff: 1.1e-06      634 iterations  logLik: -1135.51  diff: 1.1e-06      635 iterations  logLik: -1135.51  diff: 1.1e-06      636 iterations  logLik: -1135.51  diff: 1.1e-06      637 iterations  logLik: -1135.51  diff: 1.1e-06      638 iterations  logLik: -1135.51  diff: 1.1e-06      639 iterations  logLik: -1135.51  diff: 1.1e-06      640 iterations  logLik: -1135.51  diff: 1.1e-06      641 iterations  logLik: -1135.51  diff: 1.1e-06      642 iterations  logLik: -1135.51  diff: 1.1e-06      643 iterations  logLik: -1135.51  diff: 1.1e-06      644 iterations  logLik: -1135.51  diff: 1.1e-06      645 iterations  logLik: -1135.51  diff: 1.1e-06      646 iterations  logLik: -1135.51  diff: 1.1e-06      647 iterations  logLik: -1135.51  diff: 1.1e-06      648 iterations  logLik: -1135.51  diff: 1.1e-06      649 iterations  logLik: -1135.51  diff: 1.1e-06      650 iterations  logLik: -1135.51  diff: 1.1e-06      651 iterations  logLik: -1135.51  diff: 1.1e-06      652 iterations  logLik: -1135.51  diff: 1.1e-06      653 iterations  logLik: -1135.51  diff: 1.1e-06      654 iterations  logLik: -1135.51  diff: 1.1e-06      655 iterations  logLik: -1135.51  diff: 1.1e-06      656 iterations  logLik: -1135.51  diff: 1.1e-06      657 iterations  logLik: -1135.51  diff: 1.1e-06      658 iterations  logLik: -1135.51  diff: 1.1e-06      659 iterations  logLik: -1135.51  diff: 1.1e-06      660 iterations  logLik: -1135.51  diff: 1.1e-06      661 iterations  logLik: -1135.51  diff: 1.1e-06      662 iterations  logLik: -1135.51  diff: 1.1e-06      663 iterations  logLik: -1135.51  diff: 1.1e-06      664 iterations  logLik: -1135.51  diff: 1.1e-06      665 iterations  logLik: -1135.51  diff: 1.1e-06      666 iterations  logLik: -1135.51  diff: 1.1e-06      667 iterations  logLik: -1135.51  diff: 1.1e-06      668 iterations  logLik: -1135.51  diff: 1.1e-06      669 iterations  logLik: -1135.51  diff: 1.1e-06      670 iterations  logLik: -1135.51  diff: 1.1e-06      671 iterations  logLik: -1135.51  diff: 1.1e-06      672 iterations  logLik: -1135.51  diff: 1.1e-06      673 iterations  logLik: -1135.51  diff: 1.1e-06      674 iterations  logLik: -1135.51  diff: 1.1e-06      675 iterations  logLik: -1135.51  diff: 1.1e-06      676 iterations  logLik: -1135.51  diff: 1.1e-06      677 iterations  logLik: -1135.51  diff: 1.1e-06      678 iterations  logLik: -1135.51  diff: 1.1e-06      679 iterations  logLik: -1135.51  diff: 1.0e-06      680 iterations  logLik: -1135.51  diff: 1.0e-06      681 iterations  logLik: -1135.51  diff: 1.0e-06      682 iterations  logLik: -1135.51  diff: 1.0e-06      683 iterations  logLik: -1135.51  diff: 1.0e-06      684 iterations  logLik: -1135.51  diff: 1.0e-06      685 iterations  logLik: -1135.51  diff: 1.0e-06      686 iterations  logLik: -1135.51  diff: 1.0e-06      687 iterations  logLik: -1135.51  diff: 1.0e-06      688 iterations  logLik: -1135.51  diff: 1.0e-06      689 iterations  logLik: -1135.51  diff: 1.0e-06      690 iterations  logLik: -1135.51  diff: 1.0e-06      691 iterations  logLik: -1135.51  diff: 1.0e-06      692 iterations  logLik: -1135.51  diff: 1.0e-06      693 iterations  logLik: -1135.51  diff: 1.0e-06      694 iterations  logLik: -1135.51  diff: 1.0e-06      695 iterations  logLik: -1135.51  diff: 1.0e-06      696 iterations  logLik: -1135.51  diff: 1.0e-06      697 iterations  logLik: -1135.51  diff: 1.0e-06      698 iterations  logLik: -1135.51  diff: 1.0e-06      699 iterations  logLik: -1135.51  diff: 1.0e-06      700 iterations  logLik: -1135.51  diff: 1.0e-06      
#> 701 iterations  logLik: -1135.51  diff: 1.0e-06      702 iterations  logLik: -1135.51  diff: 1.0e-06      703 iterations  logLik: -1135.51  diff: 1.0e-06      704 iterations  logLik: -1135.51  diff: 1.0e-06      705 iterations  logLik: -1135.51  diff: 1.0e-06      706 iterations  logLik: -1135.51  diff: 1.0e-06      707 iterations  logLik: -1135.51  diff: 1.0e-06      708 iterations  logLik: -1135.51  diff: 1.0e-06      
#> .. done.

gof(stat2, stat3, stat4)
#> Analysis of Goodness of Fit Table
#> 
#>       Df  logLik    AIC    BIC     Gsq Res. Df
#> stat2 21 -1179.2 2400.4 2489.5 115.979      53
#> stat3 32 -1139.3 2342.6 2478.5  36.204      42
#> stat4 43 -1135.5 2357.0 2539.6  28.633      31
gof(stat2, stat3, stat4, test = "chisq")
#> Analysis of Goodness of Fit Table
#> 
#>       Df  logLik    AIC    BIC     Gsq Res. Df Pr(>Chi)    
#> stat2 21 -1179.2 2400.4 2489.5 115.979      53 1.34e-06 ***
#> stat3 32 -1139.3 2342.6 2478.5  36.204      42   0.7225    
#> stat4 43 -1135.5 2357.0 2539.6  28.633      31   0.5883    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# \donttest{
gof(stat2, stat3, stat4, test = "boot")
#> Analysis of Goodness of Fit Table
#> 
#>       Df  logLik    AIC    BIC     Gsq Res. Df Pr(Boot)    
#> stat2 21 -1179.2 2400.4 2489.5 115.979      53   <2e-16 ***
#> stat3 32 -1139.3 2342.6 2478.5  36.204      42     0.19    
#> stat4 43 -1135.5 2357.0 2539.6  28.633      31     0.30    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# }

compare(stat3, stat4)
#> Model H0: stat3 
#> Model H1: stat4 
#> 
#> Analysis of Relative Model Fit
#> 
#>       Df  logLik    AIC    BIC    Gsq Res. Df
#> stat3 32 -1139.3 2342.6 2478.5               
#> stat4 43 -1135.5 2357.0 2539.6 7.5703      11
compare(stat3, stat4, test = "chisq")
#> Model H0: stat3 
#> Model H1: stat4 
#> 
#> Analysis of Relative Model Fit
#> 
#>       Df  logLik    AIC    BIC    Gsq Res. Df Pr(>Chi)
#> stat3 32 -1139.3 2342.6 2478.5                        
#> stat4 43 -1135.5 2357.0 2539.6 7.5703      11   0.7512
# \donttest{
compare(stat3, stat4, test = "boot")
#> Model H0: stat3 
#> Model H1: stat4 
#> 
#> Analysis of Relative Model Fit
#> 
#>       Df  logLik    AIC    BIC    Gsq Res. Df Pr(Boot)
#> stat3 32 -1139.3 2342.6 2478.5                        
#> stat4 43 -1135.5 2357.0 2539.6 7.5703      11     0.38
# }