Confidence Intervals for Model Parameters

Computes confidence intervals for one or more parameters of fitted model. Package slca adds methods for slca fits.

Usage

# S3 method for slca
confint(object, parm, level = 0.95, type = c("param", "logit"), ...)

Arguments

object

an object of class slca and estimated.

parm

an integer string specifying parameters to be given confidence intervals.

level

numeric value representing the desired confidence level for the intervals, with a default of 0.95.

type

a character string specifying the format in which the results should be returned. Options include "probs" for probability format and "logit" for log-odds (logit) format, with the default being "probs".

...

additional arguments.

Value

A matrix with two columns representing the confidence intervals for the selected parameters. The columns are named based on the specified confidence level (level):

100 * (level / 2) %: This column shows the lower bound of the confidence interval.

100 * (1 - level / 2) %: This column shows the upper bound of the confidence interval.

The level parameter specifies the confidence level, with common values being 0.05 for a 95% confidence interval and 0.01 for a 99% confidence interval.

Examples

param(nlsy_jlcpa, index = TRUE)
#> PI :
#> (prof)
#>   class
#>              1           2           3           4
#>     0.2729 (1)  0.1762 (2)  0.3187 (3)  0.2322 (4)
#> 
#> TAU :
#> (A)
#>      parent
#> child           1            2            3            4
#>     1  0.2753 (5)  0.0000 (10)  0.0000 (15)  0.0770 (20)
#>     2  0.0000 (6)  0.0264 (11)  0.1894 (16)  0.0000 (21)
#>     3  0.0591 (7)  0.4030 (12)  0.1762 (17)  0.0602 (22)
#>     4  0.3077 (8)  0.3655 (13)  0.4496 (18)  0.1440 (23)
#>     5  0.3579 (9)  0.2050 (14)  0.1848 (19)  0.7188 (24)
#>             
#> parent prof 
#> child  use98
#> (B)
#>      parent
#> child            1            2            3            4
#>     1  0.9875 (25)  0.0000 (30)  0.0000 (35)  0.1110 (40)
#>     2  0.0013 (26)  0.0000 (31)  0.8876 (36)  0.0000 (41)
#>     3  0.0000 (27)  0.9749 (32)  0.0000 (37)  0.0069 (42)
#>     4  0.0112 (28)  0.0251 (33)  0.0312 (38)  0.0000 (43)
#>     5  0.0000 (29)  0.0000 (34)  0.0812 (39)  0.8821 (44)
#>             
#> parent prof 
#> child  use03
#> (C)
#>      parent
#> child            1            2            3            4
#>     1  0.9414 (45)  0.0361 (50)  0.0167 (55)  0.2579 (60)
#>     2  0.0109 (46)  0.1570 (51)  0.8165 (56)  0.0343 (61)
#>     3  0.0002 (47)  0.7507 (52)  0.0000 (57)  0.0000 (62)
#>     4  0.0107 (48)  0.0060 (53)  0.0057 (58)  0.0000 (63)
#>     5  0.0368 (49)  0.0502 (54)  0.1611 (59)  0.7077 (64)
#>             
#> parent prof 
#> child  use08
#> (D)
#>      parent
#> child            1            2            3            4            5
#>     1  0.0000 (65)  0.7675 (68)  0.6486 (71)  0.1701 (74)  0.0101 (77)
#>     2  0.7970 (66)  0.0137 (69)  0.0782 (72)  0.1233 (75)  0.9407 (78)
#>     3  0.2030 (67)  0.2188 (70)  0.2732 (73)  0.7066 (76)  0.0491 (79)
#>                         
#> parent use98 use03 use08
#> child  smk98 smk03 smk08
#> (E)
#>      parent
#> child            1            2            3            4            5
#>     1  0.2658 (80)  0.2176 (83)  0.1174 (86)  0.5843 (89)  0.2308 (92)
#>     2  0.0000 (81)  0.1910 (84)  0.0205 (87)  0.0820 (90)  0.7341 (93)
#>     3  0.7342 (82)  0.5913 (85)  0.8622 (88)  0.3337 (91)  0.0351 (94)
#>                         
#> parent use98 use03 use08
#> child  drk98 drk03 drk08
#> (F)
#>      parent
#> child            1             2             3             4             5
#>     1  0.0928 (95)   0.0520 (98)  0.7870 (101)  0.0013 (104)  0.0109 (107)
#>     2  0.1898 (96)   0.1097 (99)  0.1930 (102)  0.5129 (105)  0.0000 (108)
#>     3  0.7174 (97)  0.8383 (100)  0.0200 (103)  0.4858 (106)  0.9891 (109)
#>                         
#> parent use98 use03 use08
#> child  mrj98 mrj03 mrj08
#> 
#> RHO :
#> (a)
#>         class
#> response             1             2             3
#>    1(V1)  1.0000 (110)  0.0152 (118)  1.0000 (126)
#>    2      0.0000 (111)  0.9848 (119)  0.0000 (127)
#>    1(V2)  1.0000 (112)  0.0000 (120)  0.6728 (128)
#>    2      0.0000 (113)  1.0000 (121)  0.3272 (129)
#>    1(V3)  0.8836 (114)  0.0000 (122)  0.0000 (130)
#>    2      0.1164 (115)  1.0000 (123)  1.0000 (131)
#>    1(V4)  0.6463 (116)  0.0000 (124)  0.0000 (132)
#>    2      0.3537 (117)  1.0000 (125)  1.0000 (133)
#> 
#>       V1      V2      V3      V4     
#> smk98 ESMK_98 FSMK_98 DSMK_98 HSMK_98
#> smk03 ESMK_03 FSMK_03 DSMK_03 HSMK_03
#> smk08 ESMK_08 FSMK_08 DSMK_08 HSMK_08
#> (b)
#>         class
#> response             1             2             3
#>    1(V1)  1.0000 (134)  0.0990 (142)  1.0000 (150)
#>    2      0.0000 (135)  0.9010 (143)  0.0000 (151)
#>    1(V2)  0.5928 (136)  0.0000 (144)  1.0000 (152)
#>    2      0.4072 (137)  1.0000 (145)  0.0000 (153)
#>    1(V3)  0.0000 (138)  0.0000 (146)  0.6957 (154)
#>    2      1.0000 (139)  1.0000 (147)  0.3043 (155)
#>    1(V4)  0.0000 (140)  0.0000 (148)  0.8088 (156)
#>    2      1.0000 (141)  1.0000 (149)  0.1912 (157)
#> 
#>       V1      V2      V3      V4     
#> drk98 EDRK_98 CDRK_98 WDRK_98 BDRK_98
#> drk03 EDRK_03 CDRK_03 WDRK_03 BDRK_03
#> drk08 EDRK_08 CDRK_08 WDRK_08 BDRK_08
#> (c)
#>         class
#> response             1             2             3
#>    1(V1)  1.0000 (158)  1.0000 (166)  0.0157 (174)
#>    2      0.0000 (159)  0.0000 (167)  0.9843 (175)
#>    1(V2)  1.0000 (160)  0.5016 (168)  0.0000 (176)
#>    2      0.0000 (161)  0.4984 (169)  1.0000 (177)
#>    1(V3)  0.7903 (162)  0.0000 (170)  0.0000 (178)
#>    2      0.2097 (163)  1.0000 (171)  1.0000 (179)
#>    1(V4)  0.5591 (164)  0.0000 (172)  0.0000 (180)
#>    2      0.4409 (165)  1.0000 (173)  1.0000 (181)
#> 
#>       V1      V2      V3      V4     
#> mrj98 EMRJ_98 CMRJ_98 OMRJ_98 SMRJ_98
#> mrj03 EMRJ_03 CMRJ_03 OMRJ_03 SMRJ_03
#> mrj08 EMRJ_08 CMRJ_08 OMRJ_08 SMRJ_08
confint(nlsy_jlcpa)
#>               2.5 %        97.5 %
#> (1)    2.333152e-01  3.113322e-01
#> (2)    1.627691e-01  1.859232e-01
#> (3)    3.009433e-01  3.292266e-01
#> (4)    3.029724e-01  1.735179e-01
#> (5)    2.125055e-01  3.291963e-01
#> (6)    0.000000e+00  0.000000e+00
#> (7)    3.416965e-02  9.429820e-02
#> (8)    2.588218e-01  3.374739e-01
#> (9)    4.945030e-01  2.390317e-01
#> (10)   8.927238e-91  2.497194e-91
#> (11)   1.656739e-03  2.951713e-01
#> (12)   3.669302e-01  3.097948e-01
#> (13)   3.071979e-01  3.043421e-01
#> (14)   3.242151e-01  9.069191e-02
#> (15)   1.886593e-94  8.014075e-95
#> (16)   1.554286e-01  2.209086e-01
#> (17)   1.517025e-01  1.958372e-01
#> (18)   4.155141e-01  4.654365e-01
#> (19)   2.773548e-01  1.178178e-01
#> (20)   2.568500e-02  1.946630e-01
#> (21)  6.103542e-262 3.643339e-262
#> (22)   3.347585e-02  9.125628e-02
#> (23)   8.595630e-02  2.037822e-01
#> (24)   8.548828e-01  5.102984e-01
#> (25)   9.972229e-01  8.467190e-04
#> (26)   1.382622e-09  9.991124e-01
#> (27)  3.705492e-322  0.000000e+00
#> (28)   2.658109e-03  4.089855e-05
#> (29)   1.189698e-04  2.245702e-19
#> (30)  8.230203e-181 2.166451e-182
#> (31)   1.586760e-15  4.176856e-17
#> (32)   9.748686e-01  9.748686e-01
#> (33)   2.513139e-02  2.513139e-02
#> (34)   0.000000e+00  0.000000e+00
#> (35)  2.870893e-100 5.896147e-101
#> (36)   8.292102e-01  8.101420e-01
#> (37)   4.217897e-36  8.662581e-37
#> (38)   5.327816e-03  1.558760e-01
#> (39)   1.654619e-01  3.398204e-02
#> (40)   2.153626e-02  1.675537e-01
#> (41)   1.567634e-68  3.731209e-69
#> (42)   2.333416e-05  5.995625e-01
#> (43)  6.488630e-142 1.544393e-142
#> (44)   9.784404e-01  2.328838e-01
#> (45)   5.247851e-01  2.203611e-50
#> (46)   2.436720e-03  6.307881e-52
#> (47)   5.217179e-58  1.000000e+00
#> (48)   2.238540e-03  6.690588e-52
#> (49)   4.705397e-01  3.757429e-53
#> (50)   1.534097e-02  1.438472e-03
#> (51)   1.158490e-01  3.605438e-03
#> (52)   7.006088e-01  1.362658e-02
#> (53)   6.269688e-07  9.810755e-01
#> (54)   1.682006e-01  2.539993e-04
#> (55)   6.039055e-04  1.353538e-02
#> (56)   7.650449e-01  2.567945e-02
#> (57)  8.808576e-158 1.226697e-159
#> (58)   1.009542e-06  9.575216e-01
#> (59)   2.343502e-01  3.263601e-03
#> (60)   1.745184e-01  3.233359e-01
#> (61)   6.327725e-03  1.580121e-01
#> (62)   0.000000e+00  0.000000e+00
#> (63)  3.866199e-203 2.447906e-203
#> (64)   8.191539e-01  5.186520e-01
#> (65)   1.549833e-10  9.621466e-11
#> (66)   7.441673e-01  8.411877e-01
#> (67)   2.558327e-01  1.588123e-01
#> (68)   7.144512e-01  7.251610e-01
#> (69)   1.294679e-03  1.266935e-01
#> (70)   2.842541e-01  1.481455e-01
#> (71)   6.089869e-01  6.772912e-01
#> (72)   5.829514e-02  1.028050e-01
#> (73)   3.327180e-01  2.199038e-01
#> (74)   1.175155e-01  2.309643e-01
#> (75)   7.557364e-02  1.886478e-01
#> (76)   8.069108e-01  5.803880e-01
#> (77)   3.763638e-03  2.642568e-02
#> (78)   9.069473e-01  9.473111e-01
#> (79)   8.928906e-02  2.626323e-02
#> (80)   2.182819e-01  3.195050e-01
#> (81)   2.415154e-27  2.102420e-27
#> (82)   7.817181e-01  6.804950e-01
#> (83)   1.829593e-01  2.546591e-01
#> (84)   1.647110e-01  2.179856e-01
#> (85)   6.523298e-01  5.273553e-01
#> (86)   8.093687e-02  1.634612e-01
#> (87)   7.553106e-03  5.329044e-02
#> (88)   9.115100e-01  7.832484e-01
#> (89)   5.232065e-01  6.252035e-01
#> (90)   5.203063e-02  1.236569e-01
#> (91)   4.247629e-01  2.511395e-01
#> (92)   2.058107e-01  2.335276e-01
#> (93)   6.407064e-01  7.592284e-01
#> (94)   1.534829e-01  7.243993e-03
#> (95)   7.212655e-02  1.175760e-01
#> (96)   1.594105e-01  2.226436e-01
#> (97)   7.684630e-01  6.597804e-01
#> (98)   2.784916e-02  9.318867e-02
#> (99)   7.453603e-02  1.549981e-01
#> (100)  8.976148e-01  7.518133e-01
#> (101)  4.211653e-01  8.082271e-01
#> (102)  1.070549e-01  1.913064e-01
#> (103)  4.717798e-01  4.665556e-04
#> (104)  2.455966e-18  1.000000e+00
#> (105)  4.320614e-01  8.649072e-13
#> (106)  5.679386e-01  5.903454e-13
#> (107)  4.409475e-03  2.676174e-02
#> (108) 9.832297e-147 9.611549e-147
#> (109)  9.955905e-01  9.732383e-01
#> (110)  1.000000e+00  1.000000e+00
#> (111)  0.000000e+00  0.000000e+00
#> (112)  1.000000e+00  1.000000e+00
#> (113)  0.000000e+00  0.000000e+00
#> (114)  8.514482e-01  9.095975e-01
#> (115)  1.485518e-01  9.040250e-02
#> (116)  6.145082e-01  6.768744e-01
#> (117)  3.854918e-01  3.231256e-01
#> (118)  3.776843e-03  5.923633e-02
#> (119)  9.962232e-01  9.407637e-01
#> (120)  0.000000e+00  0.000000e+00
#> (121)  1.000000e+00  1.000000e+00
#> (122)  0.000000e+00  0.000000e+00
#> (123)  1.000000e+00  1.000000e+00
#> (124)  0.000000e+00  0.000000e+00
#> (125)  1.000000e+00  1.000000e+00
#> (126)  1.000000e+00  1.000000e+00
#> (127)  0.000000e+00  0.000000e+00
#> (128)  6.195857e-01  7.218352e-01
#> (129)  3.804143e-01  2.781648e-01
#> (130) 1.095174e-274 1.095174e-274
#> (131)  1.000000e+00  1.000000e+00
#> (132) 1.629762e-297 1.629762e-297
#> (133)  1.000000e+00  1.000000e+00
#> (134)  1.000000e+00  1.000000e+00
#> (135)  0.000000e+00  0.000000e+00
#> (136)  5.339593e-01  6.490855e-01
#> (137)  4.660407e-01  3.509145e-01
#> (138)  0.000000e+00  0.000000e+00
#> (139)  1.000000e+00  1.000000e+00
#> (140) 1.526863e-261 1.526863e-261
#> (141)  1.000000e+00  1.000000e+00
#> (142)  4.769964e-02  1.942945e-01
#> (143)  9.523004e-01  8.057055e-01
#> (144)  0.000000e+00  0.000000e+00
#> (145)  1.000000e+00  1.000000e+00
#> (146)  0.000000e+00  0.000000e+00
#> (147)  1.000000e+00  1.000000e+00
#> (148)  0.000000e+00  0.000000e+00
#> (149)  1.000000e+00  1.000000e+00
#> (150)  1.000000e+00  1.000000e+00
#> (151)  0.000000e+00  0.000000e+00
#> (152)  1.000000e+00  1.000000e+00
#> (153)  0.000000e+00  0.000000e+00
#> (154)  6.681142e-01  7.219062e-01
#> (155)  3.318858e-01  2.780938e-01
#> (156)  7.845176e-01  8.308794e-01
#> (157)  2.154824e-01  1.691206e-01
#> (158)  1.000000e+00  1.000000e+00
#> (159)  0.000000e+00  0.000000e+00
#> (160)  1.000000e+00  1.000000e+00
#> (161)  0.000000e+00  0.000000e+00
#> (162)  7.296830e-01  8.402829e-01
#> (163)  2.703170e-01  1.597171e-01
#> (164)  5.082244e-01  6.087412e-01
#> (165)  4.917756e-01  3.912588e-01
#> (166)  1.000000e+00  1.000000e+00
#> (167)  0.000000e+00  0.000000e+00
#> (168)  4.434991e-01  5.596139e-01
#> (169)  5.565009e-01  4.403861e-01
#> (170)  0.000000e+00  0.000000e+00
#> (171)  1.000000e+00  1.000000e+00
#> (172)  6.526489e-66  6.526489e-66
#> (173)  1.000000e+00  1.000000e+00
#> (174)  7.457063e-03  3.265249e-02
#> (175)  9.925429e-01  9.673475e-01
#> (176)  0.000000e+00  0.000000e+00
#> (177)  1.000000e+00  1.000000e+00
#> (178)  0.000000e+00  0.000000e+00
#> (179)  1.000000e+00  1.000000e+00
#> (180)  0.000000e+00  0.000000e+00
#> (181)  1.000000e+00  1.000000e+00
confint(nlsy_jlcpa, 1:4)
#>         2.5 %    97.5 %
#> (1) 0.2333152 0.3113322
#> (2) 0.1627691 0.1859232
#> (3) 0.3009433 0.3292266
#> (4) 0.3029724 0.1735179