Confidence Intervals for Model Parameters

Computes confidence intervals for one or more parameters of a fitted model.

Usage

# S3 method for class 'slcafit'
confint(object, parm, level = 0.95, type = c("probs", "logit"), ...)

Arguments

object

an object of class slcafit.

parm

an integer or string specifying the parameters for which confidence intervals are to be computed.

level

a numeric value representing the confidence level for the intervals. The default is 0.95 (95% confidence level).

type

a character string specifying the format in which the results should be returned. Options include "probs" for probability format and "logit" for log-odds (logit) format, with the default being "probs".

...

additional arguments.

Value

A matrix with two columns representing the confidence intervals for the selected parameters. The column names correspond to the specified confidence level:

100 * (level / 2)%: The lower bound of the confidence interval.

100 * (1 - level / 2)%: The upper bound of the confidence interval.

The level argument determines the confidence level, with common values being 0.95 for a 95% confidence interval and 0.99 for a 99% confidence interval.

Examples

param(nlsy_jlcpa, index = TRUE)
#> PI :
#> (PROF)
#>   class
#>              1           2           3           4
#>     0.2790 (1)  0.2146 (2)  0.2869 (3)  0.2195 (4)
#> 
#> TAU :
#> (A)
#>      parent
#> child           1            2            3            4
#>     1  0.0000 (5)  0.0000 (10)  0.1745 (15)  0.0000 (20)
#>     2  0.0901 (6)  0.4456 (11)  0.1939 (16)  0.0403 (21)
#>     3  0.3771 (7)  0.1675 (12)  0.1825 (17)  0.7018 (22)
#>     4  0.2620 (8)  0.0000 (13)  0.0000 (18)  0.0813 (23)
#>     5  0.2709 (9)  0.3869 (14)  0.4491 (19)  0.1766 (24)
#>              
#> parent PROF  
#> child  SUB_98
#> (B)
#>      parent
#> child            1            2            3            4
#>     1  0.0000 (25)  0.0114 (30)  0.8719 (35)  0.0000 (40)
#>     2  0.0000 (26)  0.9617 (31)  0.0000 (36)  0.0079 (41)
#>     3  0.0000 (27)  0.0000 (32)  0.0965 (37)  0.9210 (42)
#>     4  0.9941 (28)  0.0000 (33)  0.0000 (38)  0.0711 (43)
#>     5  0.0059 (29)  0.0269 (34)  0.0316 (39)  0.0000 (44)
#>              
#> parent PROF  
#> child  SUB_03
#> (C)
#>      parent
#> child            1            2            3            4
#>     1  0.0019 (45)  0.1821 (50)  0.8024 (55)  0.0319 (60)
#>     2  0.0097 (46)  0.7202 (51)  0.0000 (56)  0.0000 (61)
#>     3  0.0573 (47)  0.0442 (52)  0.1844 (57)  0.7168 (62)
#>     4  0.9311 (48)  0.0356 (53)  0.0084 (58)  0.2473 (63)
#>     5  0.0000 (49)  0.0179 (54)  0.0048 (59)  0.0040 (64)
#>              
#> parent PROF  
#> child  SUB_08
#> (D)
#>      parent
#> child            1            2            3            4            5
#>     1  0.0254 (65)  0.0457 (68)  0.9505 (71)  0.7958 (74)  0.1613 (77)
#>     2  0.7516 (66)  0.6690 (69)  0.0051 (72)  0.0000 (75)  0.1631 (78)
#>     3  0.2230 (67)  0.2853 (70)  0.0444 (73)  0.2042 (76)  0.6755 (79)
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  SMK_98 SMK_03 SMK_08
#> (E)
#>      parent
#> child            1            2            3            4            5
#>     1  0.5664 (80)  0.8533 (83)  0.0335 (86)  0.7521 (89)  0.2872 (92)
#>     2  0.2229 (81)  0.1264 (84)  0.2589 (87)  0.2414 (90)  0.6127 (93)
#>     3  0.2107 (82)  0.0203 (85)  0.7075 (88)  0.0065 (91)  0.1001 (94)
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  DRK_98 DRK_03 DRK_08
#> (F)
#>      parent
#> child            1             2             3             4             5
#>     1  0.0871 (95)   0.1442 (98)  0.0317 (101)  0.1525 (104)  0.7278 (107)
#>     2  0.0617 (96)   0.8387 (99)  0.0095 (102)  0.2302 (105)  0.1258 (108)
#>     3  0.8512 (97)  0.0170 (100)  0.9588 (103)  0.6173 (106)  0.1463 (109)
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  MRJ_98 MRJ_03 MRJ_08
#> 
#> RHO :
#> (a)
#>         class
#> response             1             2             3
#>    1(V1)  0.0134 (110)  1.0000 (118)  1.0000 (126)
#>    2      0.9866 (111)  0.0000 (119)  0.0000 (127)
#>    1(V2)  0.0000 (112)  1.0000 (120)  0.6701 (128)
#>    2      1.0000 (113)  0.0000 (121)  0.3299 (129)
#>    1(V3)  0.0000 (114)  0.8833 (122)  0.0000 (130)
#>    2      1.0000 (115)  0.1167 (123)  1.0000 (131)
#>    1(V4)  0.0000 (116)  0.6460 (124)  0.0000 (132)
#>    2      1.0000 (117)  0.3540 (125)  1.0000 (133)
#> 
#>        V1      V2      V3      V4     
#> SMK_98 ESMK_98 FSMK_98 DSMK_98 HSMK_98
#> SMK_03 ESMK_03 FSMK_03 DSMK_03 HSMK_03
#> SMK_08 ESMK_08 FSMK_08 DSMK_08 HSMK_08
#> (b)
#>         class
#> response             1             2             3
#>    1(V1)  1.0000 (134)  1.0000 (142)  0.0759 (150)
#>    2      0.0000 (135)  0.0000 (143)  0.9241 (151)
#>    1(V2)  1.0000 (136)  0.5760 (144)  0.0000 (152)
#>    2      0.0000 (137)  0.4240 (145)  1.0000 (153)
#>    1(V3)  0.6934 (138)  0.0000 (146)  0.0000 (154)
#>    2      0.3066 (139)  1.0000 (147)  1.0000 (155)
#>    1(V4)  0.8061 (140)  0.0000 (148)  0.0000 (156)
#>    2      0.1939 (141)  1.0000 (149)  1.0000 (157)
#> 
#>        V1      V2      V3      V4     
#> DRK_98 EDRK_98 CDRK_98 WDRK_98 BDRK_98
#> DRK_03 EDRK_03 CDRK_03 WDRK_03 BDRK_03
#> DRK_08 EDRK_08 CDRK_08 WDRK_08 BDRK_08
#> (c)
#>         class
#> response             1             2             3
#>    1(V1)  0.5189 (158)  1.0000 (166)  0.0002 (174)
#>    2      0.4811 (159)  0.0000 (167)  0.9998 (175)
#>    1(V2)  0.0000 (160)  1.0000 (168)  0.0000 (176)
#>    2      1.0000 (161)  0.0000 (169)  1.0000 (177)
#>    1(V3)  0.0000 (162)  0.5259 (170)  0.0000 (178)
#>    2      1.0000 (163)  0.4741 (171)  1.0000 (179)
#>    1(V4)  0.0000 (164)  0.3720 (172)  0.0000 (180)
#>    2      1.0000 (165)  0.6280 (173)  1.0000 (181)
#> 
#>        V1      V2      V3      V4     
#> MRJ_98 EMRJ_98 CMRJ_98 OMRJ_98 SMRJ_98
#> MRJ_03 EMRJ_03 CMRJ_03 OMRJ_03 SMRJ_03
#> MRJ_08 EMRJ_08 CMRJ_08 OMRJ_08 SMRJ_08
confint(nlsy_jlcpa)
#>                2.5 %        97.5 %
#> (1)     2.196632e-01  3.384007e-01
#> (2)     1.708777e-01  2.582819e-01
#> (3)     2.377812e-01  3.359991e-01
#> (4)     1.607946e-01  2.782015e-01
#> (5)    1.316219e-147 2.962555e-147
#> (6)     2.865600e-02  1.514643e-01
#> (7)     2.817499e-01  4.724122e-01
#> (8)     1.577645e-01  3.661524e-01
#> (9)     1.666665e-01  3.751342e-01
#> (10)    7.265367e-18  1.356439e-17
#> (11)    3.351151e-01  5.560190e-01
#> (12)    9.349363e-02  2.415669e-01
#> (13)    3.052471e-09  5.698944e-09
#> (14)    2.699013e-01  5.039040e-01
#> (15)    8.636855e-02  2.625399e-01
#> (16)    1.136580e-01  2.741749e-01
#> (17)    1.163294e-01  2.487285e-01
#> (18)    2.414352e-25  4.143108e-25
#> (19)    3.307031e-01  5.674976e-01
#> (20)    7.034261e-81  2.663338e-80
#> (21)   -8.554393e-03  8.913669e-02
#> (22)    6.014447e-01  8.021353e-01
#> (23)   -5.208181e-03  1.679023e-01
#> (24)    7.378316e-02  2.793605e-01
#> (25)   -3.365407e-26  4.112143e-26
#> (26)   -2.996881e-75  3.661846e-75
#> (27)   -9.303904e-07  1.136831e-06
#> (28)    9.351496e-01  1.053074e+00
#> (29)   -5.307307e-02  6.484923e-02
#> (30)   -9.982077e-02  1.226243e-01
#> (31)    8.482744e-01  1.075050e+00
#> (32)   1.523339e-199 1.523339e-199
#> (33)  -8.289540e-100  1.757511e-99
#> (34)   -4.809295e-02  1.019645e-01
#> (35)    7.829478e-01  9.608499e-01
#> (36)   -1.195434e-11  3.513308e-11
#> (37)    2.886429e-02  1.641170e-01
#> (38)   -4.341679e-82  1.275993e-81
#> (39)   -3.260598e-02  9.582698e-02
#> (40)   -4.963485e-26  1.642774e-25
#> (41)   -3.376576e-02  4.962831e-02
#> (42)    7.269635e-01  1.114959e+00
#> (43)   -1.217313e-01  2.639468e-01
#> (44)   -8.170873e-49  2.704329e-48
#> (45)   -1.934174e-02  2.313701e-02
#> (46)   -1.574984e-02  3.515286e-02
#> (47)   -6.158788e-02  1.760997e-01
#> (48)    8.071224e-01  1.055167e+00
#> (49)  -4.583996e-129 9.547265e-129
#> (50)    6.279219e-02  3.014505e-01
#> (51)    5.887958e-01  8.516152e-01
#> (52)   -8.285468e-03  9.661303e-02
#> (53)   -1.985109e-02  9.110167e-02
#> (54)   -3.202804e-02  6.779624e-02
#> (55)    7.123386e-01  8.925108e-01
#> (56)   -6.418826e-39  7.687447e-39
#> (57)    1.170281e-01  2.517426e-01
#> (58)   -4.374604e-02  6.054986e-02
#> (59)   -4.845173e-02  5.802777e-02
#> (60)   -2.924168e-02  9.305138e-02
#> (61)   7.216797e-232 7.216797e-232
#> (62)    6.056283e-01  8.279084e-01
#> (63)    1.390506e-01  3.556173e-01
#> (64)   -3.099270e-02  3.897843e-02
#> (65)   -1.939799e-02  7.022763e-02
#> (66)    6.818585e-01  8.214010e-01
#> (67)    1.666418e-01  2.792690e-01
#> (68)    1.832904e-02  7.314001e-02
#> (69)    6.162033e-01  7.218193e-01
#> (70)    2.371028e-01  3.334056e-01
#> (71)    9.163161e-01  9.846204e-01
#> (72)   -8.141767e-03  1.830527e-02
#> (73)    1.412244e-02  7.477748e-02
#> (74)    7.491208e-01  8.425406e-01
#> (75)    1.196215e-11  1.905923e-11
#> (76)    1.574594e-01  2.508792e-01
#> (77)    8.931824e-02  2.333613e-01
#> (78)    9.639739e-02  2.298591e-01
#> (79)    5.870730e-01  7.639909e-01
#> (80)    5.133754e-01  6.193661e-01
#> (81)    1.727128e-01  2.730651e-01
#> (82)    1.682509e-01  2.532297e-01
#> (83)    8.074661e-01  8.991574e-01
#> (84)    8.326637e-02  1.695867e-01
#> (85)    2.061184e-03  3.846230e-02
#> (86)   -1.991548e-02  8.701094e-02
#> (87)    1.836337e-01  3.342563e-01
#> (88)    6.194496e-01  7.955649e-01
#> (89)    6.997628e-01  8.044282e-01
#> (90)    1.919496e-01  2.909129e-01
#> (91)   -1.021852e-02  2.316502e-02
#> (92)    2.049584e-01  3.694083e-01
#> (93)    5.265049e-01  6.988820e-01
#> (94)    4.371735e-02  1.565291e-01
#> (95)    1.996969e-02  1.541847e-01
#> (96)    1.501435e-02  1.084684e-01
#> (97)    7.695210e-01  9.328419e-01
#> (98)    7.254894e-02  2.159447e-01
#> (99)    7.496561e-01  9.278350e-01
#> (100)  -1.313749e-02  4.715276e-02
#> (101)   2.213799e-03  6.109436e-02
#> (102)  -6.743734e-04  1.971026e-02
#> (103)   9.271295e-01  9.905265e-01
#> (104)   7.665047e-02  2.283733e-01
#> (105)   1.953820e-01  2.649226e-01
#> (106)   5.368540e-01  6.978177e-01
#> (107)   3.927324e-01  1.062935e+00
#> (108)   5.567487e-02  1.959617e-01
#> (109)  -1.837854e-01  4.764817e-01
#> (110)  -7.588212e-03  3.436836e-02
#> (111)   9.656316e-01  1.007588e+00
#> (112)   0.000000e+00  0.000000e+00
#> (113)   1.000000e+00  1.000000e+00
#> (114)   0.000000e+00  0.000000e+00
#> (115)   1.000000e+00  1.000000e+00
#> (116)   0.000000e+00  0.000000e+00
#> (117)   1.000000e+00  1.000000e+00
#> (118)   1.000000e+00  1.000000e+00
#> (119)   0.000000e+00  0.000000e+00
#> (120)   1.000000e+00  1.000000e+00
#> (121)   0.000000e+00  0.000000e+00
#> (122)   8.542618e-01  9.122978e-01
#> (123)   8.770220e-02  1.457382e-01
#> (124)   6.151722e-01  6.769268e-01
#> (125)   3.230732e-01  3.848278e-01
#> (126)   1.000000e+00  1.000000e+00
#> (127)   0.000000e+00  0.000000e+00
#> (128)   6.184276e-01  7.216748e-01
#> (129)   2.783252e-01  3.815724e-01
#> (130)   0.000000e+00  0.000000e+00
#> (131)   1.000000e+00  1.000000e+00
#> (132)   0.000000e+00  0.000000e+00
#> (133)   1.000000e+00  1.000000e+00
#> (134)   1.000000e+00  1.000000e+00
#> (135)   0.000000e+00  0.000000e+00
#> (136)   1.000000e+00  1.000000e+00
#> (137)   0.000000e+00  0.000000e+00
#> (138)   6.665516e-01  7.202240e-01
#> (139)   2.797760e-01  3.334484e-01
#> (140)   7.827662e-01  8.294459e-01
#> (141)   1.705541e-01  2.172338e-01
#> (142)   1.000000e+00  1.000000e+00
#> (143)   0.000000e+00  0.000000e+00
#> (144)   5.150970e-01  6.369518e-01
#> (145)   3.630482e-01  4.849030e-01
#> (146)  2.266886e-173 2.266886e-173
#> (147)   1.000000e+00  1.000000e+00
#> (148)  3.659486e-127 3.659486e-127
#> (149)   1.000000e+00  1.000000e+00
#> (150)  -4.844300e-03  1.567197e-01
#> (151)   8.432803e-01  1.004844e+00
#> (152)   0.000000e+00  0.000000e+00
#> (153)   1.000000e+00  1.000000e+00
#> (154)   0.000000e+00  0.000000e+00
#> (155)   1.000000e+00  1.000000e+00
#> (156)   0.000000e+00  0.000000e+00
#> (157)   1.000000e+00  1.000000e+00
#> (158)   3.129676e-01  7.248330e-01
#> (159)   2.751670e-01  6.870324e-01
#> (160)  9.689486e-106 9.689486e-106
#> (161)   1.000000e+00  1.000000e+00
#> (162)   0.000000e+00  0.000000e+00
#> (163)   1.000000e+00  1.000000e+00
#> (164)   0.000000e+00  0.000000e+00
#> (165)   1.000000e+00  1.000000e+00
#> (166)   1.000000e+00  1.000000e+00
#> (167)   0.000000e+00  0.000000e+00
#> (168)   1.000000e+00  1.000000e+00
#> (169)   0.000000e+00  0.000000e+00
#> (170)   4.825718e-01  5.691765e-01
#> (171)   4.308235e-01  5.174282e-01
#> (172)   3.318994e-01  4.121566e-01
#> (173)   5.878434e-01  6.681006e-01
#> (174)   2.279783e-04  2.354243e-04
#> (175)   9.997646e-01  9.997720e-01
#> (176)   0.000000e+00  0.000000e+00
#> (177)   1.000000e+00  1.000000e+00
#> (178)   0.000000e+00  0.000000e+00
#> (179)   1.000000e+00  1.000000e+00
#> (180)   0.000000e+00  0.000000e+00
#> (181)   1.000000e+00  1.000000e+00
confint(nlsy_jlcpa, 1:4)
#>         2.5 %    97.5 %
#> (1) 0.2196632 0.3384007
#> (2) 0.1708777 0.2582819
#> (3) 0.2377812 0.3359991
#> (4) 0.1607946 0.2782015