Reorder Latent Class Membership of Class Variables

This function reorders the latent class membership for specified latent class variables.

Usage

# S3 method for class 'slcafit'
reorder(x, ...)

Arguments

x

an object of class slcafit.

...

additional arguments specifying the new order for the latent class variables.

Value

Returns the modified slcafit object with the reordered latent classes.

Examples

library(magrittr)
nlsy_jlcpa %>% param
#> PI :
#> (PROF)
#>   class
#>          1       2       3       4
#>     0.2790  0.2146  0.2869  0.2195
#> 
#> TAU :
#> (A)
#>      parent
#> child       1       2       3       4
#>     1  0.0000  0.0000  0.1745  0.0000
#>     2  0.0901  0.4456  0.1939  0.0403
#>     3  0.3771  0.1675  0.1825  0.7018
#>     4  0.2620  0.0000  0.0000  0.0813
#>     5  0.2709  0.3869  0.4491  0.1766
#>              
#> parent PROF  
#> child  SUB_98
#> (B)
#>      parent
#> child       1       2       3       4
#>     1  0.0000  0.0114  0.8719  0.0000
#>     2  0.0000  0.9617  0.0000  0.0079
#>     3  0.0000  0.0000  0.0965  0.9210
#>     4  0.9941  0.0000  0.0000  0.0711
#>     5  0.0059  0.0269  0.0316  0.0000
#>              
#> parent PROF  
#> child  SUB_03
#> (C)
#>      parent
#> child       1       2       3       4
#>     1  0.0019  0.1821  0.8024  0.0319
#>     2  0.0097  0.7202  0.0000  0.0000
#>     3  0.0573  0.0442  0.1844  0.7168
#>     4  0.9311  0.0356  0.0084  0.2473
#>     5  0.0000  0.0179  0.0048  0.0040
#>              
#> parent PROF  
#> child  SUB_08
#> (D)
#>      parent
#> child       1       2       3       4       5
#>     1  0.0254  0.0457  0.9505  0.7958  0.1613
#>     2  0.7516  0.6690  0.0051  0.0000  0.1631
#>     3  0.2230  0.2853  0.0444  0.2042  0.6755
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  SMK_98 SMK_03 SMK_08
#> (E)
#>      parent
#> child       1       2       3       4       5
#>     1  0.5664  0.8533  0.0335  0.7521  0.2872
#>     2  0.2229  0.1264  0.2589  0.2414  0.6127
#>     3  0.2107  0.0203  0.7075  0.0065  0.1001
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  DRK_98 DRK_03 DRK_08
#> (F)
#>      parent
#> child       1       2       3       4       5
#>     1  0.0871  0.1442  0.0317  0.1525  0.7278
#>     2  0.0617  0.8387  0.0095  0.2302  0.1258
#>     3  0.8512  0.0170  0.9588  0.6173  0.1463
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  MRJ_98 MRJ_03 MRJ_08
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  0.0134  1.0000  1.0000
#>    2      0.9866  0.0000  0.0000
#>    1(V2)  0.0000  1.0000  0.6701
#>    2      1.0000  0.0000  0.3299
#>    1(V3)  0.0000  0.8833  0.0000
#>    2      1.0000  0.1167  1.0000
#>    1(V4)  0.0000  0.6460  0.0000
#>    2      1.0000  0.3540  1.0000
#> 
#>        V1      V2      V3      V4     
#> SMK_98 ESMK_98 FSMK_98 DSMK_98 HSMK_98
#> SMK_03 ESMK_03 FSMK_03 DSMK_03 HSMK_03
#> SMK_08 ESMK_08 FSMK_08 DSMK_08 HSMK_08
#> (b)
#>         class
#> response       1       2       3
#>    1(V1)  1.0000  1.0000  0.0759
#>    2      0.0000  0.0000  0.9241
#>    1(V2)  1.0000  0.5760  0.0000
#>    2      0.0000  0.4240  1.0000
#>    1(V3)  0.6934  0.0000  0.0000
#>    2      0.3066  1.0000  1.0000
#>    1(V4)  0.8061  0.0000  0.0000
#>    2      0.1939  1.0000  1.0000
#> 
#>        V1      V2      V3      V4     
#> DRK_98 EDRK_98 CDRK_98 WDRK_98 BDRK_98
#> DRK_03 EDRK_03 CDRK_03 WDRK_03 BDRK_03
#> DRK_08 EDRK_08 CDRK_08 WDRK_08 BDRK_08
#> (c)
#>         class
#> response       1       2       3
#>    1(V1)  0.5189  1.0000  0.0002
#>    2      0.4811  0.0000  0.9998
#>    1(V2)  0.0000  1.0000  0.0000
#>    2      1.0000  0.0000  1.0000
#>    1(V3)  0.0000  0.5259  0.0000
#>    2      1.0000  0.4741  1.0000
#>    1(V4)  0.0000  0.3720  0.0000
#>    2      1.0000  0.6280  1.0000
#> 
#>        V1      V2      V3      V4     
#> MRJ_98 EMRJ_98 CMRJ_98 OMRJ_98 SMRJ_98
#> MRJ_03 EMRJ_03 CMRJ_03 OMRJ_03 SMRJ_03
#> MRJ_08 EMRJ_08 CMRJ_08 OMRJ_08 SMRJ_08

# Reorder the RHO parameters as ascending order
reordered1 <- nlsy_jlcpa %>%
   reorder(SMK_98 = c(1, 3, 2),
           DRK_98 = c(3, 2, 1),
           MRJ_98 = c(3, 1, 2))
reordered1 %>% param
#> PI :
#> (PROF)
#>   class
#>          1       2       3       4
#>     0.2790  0.2146  0.2869  0.2195
#> 
#> TAU :
#> (A)
#>      parent
#> child       1       2       3       4
#>     1  0.0000  0.0000  0.1745  0.0000
#>     2  0.0901  0.4456  0.1939  0.0403
#>     3  0.3771  0.1675  0.1825  0.7018
#>     4  0.2620  0.0000  0.0000  0.0813
#>     5  0.2709  0.3869  0.4491  0.1766
#>              
#> parent PROF  
#> child  SUB_98
#> (B)
#>      parent
#> child       1       2       3       4
#>     1  0.0000  0.0114  0.8719  0.0000
#>     2  0.0000  0.9617  0.0000  0.0079
#>     3  0.0000  0.0000  0.0965  0.9210
#>     4  0.9941  0.0000  0.0000  0.0711
#>     5  0.0059  0.0269  0.0316  0.0000
#>              
#> parent PROF  
#> child  SUB_03
#> (C)
#>      parent
#> child       1       2       3       4
#>     1  0.0019  0.1821  0.8024  0.0319
#>     2  0.0097  0.7202  0.0000  0.0000
#>     3  0.0573  0.0442  0.1844  0.7168
#>     4  0.9311  0.0356  0.0084  0.2473
#>     5  0.0000  0.0179  0.0048  0.0040
#>              
#> parent PROF  
#> child  SUB_08
#> (D)
#>      parent
#> child       1       2       3       4       5
#>     1  0.0254  0.0457  0.9505  0.7958  0.1613
#>     2  0.2230  0.2853  0.0444  0.2042  0.6755
#>     3  0.7516  0.6690  0.0051  0.0000  0.1631
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  SMK_98 SMK_03 SMK_08
#> (E)
#>      parent
#> child       1       2       3       4       5
#>     1  0.2107  0.0203  0.7075  0.0065  0.1001
#>     2  0.2229  0.1264  0.2589  0.2414  0.6127
#>     3  0.5664  0.8533  0.0335  0.7521  0.2872
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  DRK_98 DRK_03 DRK_08
#> (F)
#>      parent
#> child       1       2       3       4       5
#>     1  0.8512  0.0170  0.9588  0.6173  0.1463
#>     2  0.0871  0.1442  0.0317  0.1525  0.7278
#>     3  0.0617  0.8387  0.0095  0.2302  0.1258
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  MRJ_98 MRJ_03 MRJ_08
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  0.0134  1.0000  1.0000
#>    2      0.9866  0.0000  0.0000
#>    1(V2)  0.0000  0.6701  1.0000
#>    2      1.0000  0.3299  0.0000
#>    1(V3)  0.0000  0.0000  0.8833
#>    2      1.0000  1.0000  0.1167
#>    1(V4)  0.0000  0.0000  0.6460
#>    2      1.0000  1.0000  0.3540
#> 
#>        V1      V2      V3      V4     
#> SMK_98 ESMK_98 FSMK_98 DSMK_98 HSMK_98
#> SMK_03 ESMK_03 FSMK_03 DSMK_03 HSMK_03
#> SMK_08 ESMK_08 FSMK_08 DSMK_08 HSMK_08
#> (b)
#>         class
#> response       1       2       3
#>    1(V1)  0.0759  1.0000  1.0000
#>    2      0.9241  0.0000  0.0000
#>    1(V2)  0.0000  0.5760  1.0000
#>    2      1.0000  0.4240  0.0000
#>    1(V3)  0.0000  0.0000  0.6934
#>    2      1.0000  1.0000  0.3066
#>    1(V4)  0.0000  0.0000  0.8061
#>    2      1.0000  1.0000  0.1939
#> 
#>        V1      V2      V3      V4     
#> DRK_98 EDRK_98 CDRK_98 WDRK_98 BDRK_98
#> DRK_03 EDRK_03 CDRK_03 WDRK_03 BDRK_03
#> DRK_08 EDRK_08 CDRK_08 WDRK_08 BDRK_08
#> (c)
#>         class
#> response       1       2       3
#>    1(V1)  0.0002  0.5189  1.0000
#>    2      0.9998  0.4811  0.0000
#>    1(V2)  0.0000  0.0000  1.0000
#>    2      1.0000  1.0000  0.0000
#>    1(V3)  0.0000  0.0000  0.5259
#>    2      1.0000  1.0000  0.4741
#>    1(V4)  0.0000  0.0000  0.3720
#>    2      1.0000  1.0000  0.6280
#> 
#>        V1      V2      V3      V4     
#> MRJ_98 EMRJ_98 CMRJ_98 OMRJ_98 SMRJ_98
#> MRJ_03 EMRJ_03 CMRJ_03 OMRJ_03 SMRJ_03
#> MRJ_08 EMRJ_08 CMRJ_08 OMRJ_08 SMRJ_08
# Label class1: nonuse
#       class2: lifetime use
#       class3: current use

# Reorder the TAU parameters for joint classes as ascending order
reordered2 <- reordered1 %>%
   reorder(SUB_98 = c(3, 4, 5, 1, 2))
reordered2 %>% param
#> PI :
#> (PROF)
#>   class
#>          1       2       3       4
#>     0.2790  0.2146  0.2869  0.2195
#> 
#> TAU :
#> (A)
#>      parent
#> child       1       2       3       4
#>     1  0.3771  0.1675  0.1825  0.7018
#>     2  0.2620  0.0000  0.0000  0.0813
#>     3  0.2709  0.3869  0.4491  0.1766
#>     4  0.0000  0.0000  0.1745  0.0000
#>     5  0.0901  0.4456  0.1939  0.0403
#>              
#> parent PROF  
#> child  SUB_98
#> (B)
#>      parent
#> child       1       2       3       4
#>     1  0.0000  0.0000  0.0965  0.9210
#>     2  0.9941  0.0000  0.0000  0.0711
#>     3  0.0059  0.0269  0.0316  0.0000
#>     4  0.0000  0.0114  0.8719  0.0000
#>     5  0.0000  0.9617  0.0000  0.0079
#>              
#> parent PROF  
#> child  SUB_03
#> (C)
#>      parent
#> child       1       2       3       4
#>     1  0.0573  0.0442  0.1844  0.7168
#>     2  0.9311  0.0356  0.0084  0.2473
#>     3  0.0000  0.0179  0.0048  0.0040
#>     4  0.0019  0.1821  0.8024  0.0319
#>     5  0.0097  0.7202  0.0000  0.0000
#>              
#> parent PROF  
#> child  SUB_08
#> (D)
#>      parent
#> child       1       2       3       4       5
#>     1  0.9505  0.7958  0.1613  0.0254  0.0457
#>     2  0.0444  0.2042  0.6755  0.2230  0.2853
#>     3  0.0051  0.0000  0.1631  0.7516  0.6690
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  SMK_98 SMK_03 SMK_08
#> (E)
#>      parent
#> child       1       2       3       4       5
#>     1  0.7075  0.0065  0.1001  0.2107  0.0203
#>     2  0.2589  0.2414  0.6127  0.2229  0.1264
#>     3  0.0335  0.7521  0.2872  0.5664  0.8533
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  DRK_98 DRK_03 DRK_08
#> (F)
#>      parent
#> child       1       2       3       4       5
#>     1  0.9588  0.6173  0.1463  0.8512  0.0170
#>     2  0.0317  0.1525  0.7278  0.0871  0.1442
#>     3  0.0095  0.2302  0.1258  0.0617  0.8387
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  MRJ_98 MRJ_03 MRJ_08
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  0.0134  1.0000  1.0000
#>    2      0.9866  0.0000  0.0000
#>    1(V2)  0.0000  0.6701  1.0000
#>    2      1.0000  0.3299  0.0000
#>    1(V3)  0.0000  0.0000  0.8833
#>    2      1.0000  1.0000  0.1167
#>    1(V4)  0.0000  0.0000  0.6460
#>    2      1.0000  1.0000  0.3540
#> 
#>        V1      V2      V3      V4     
#> SMK_98 ESMK_98 FSMK_98 DSMK_98 HSMK_98
#> SMK_03 ESMK_03 FSMK_03 DSMK_03 HSMK_03
#> SMK_08 ESMK_08 FSMK_08 DSMK_08 HSMK_08
#> (b)
#>         class
#> response       1       2       3
#>    1(V1)  0.0759  1.0000  1.0000
#>    2      0.9241  0.0000  0.0000
#>    1(V2)  0.0000  0.5760  1.0000
#>    2      1.0000  0.4240  0.0000
#>    1(V3)  0.0000  0.0000  0.6934
#>    2      1.0000  1.0000  0.3066
#>    1(V4)  0.0000  0.0000  0.8061
#>    2      1.0000  1.0000  0.1939
#> 
#>        V1      V2      V3      V4     
#> DRK_98 EDRK_98 CDRK_98 WDRK_98 BDRK_98
#> DRK_03 EDRK_03 CDRK_03 WDRK_03 BDRK_03
#> DRK_08 EDRK_08 CDRK_08 WDRK_08 BDRK_08
#> (c)
#>         class
#> response       1       2       3
#>    1(V1)  0.0002  0.5189  1.0000
#>    2      0.9998  0.4811  0.0000
#>    1(V2)  0.0000  0.0000  1.0000
#>    2      1.0000  1.0000  0.0000
#>    1(V3)  0.0000  0.0000  0.5259
#>    2      1.0000  1.0000  0.4741
#>    1(V4)  0.0000  0.0000  0.3720
#>    2      1.0000  1.0000  0.6280
#> 
#>        V1      V2      V3      V4     
#> MRJ_98 EMRJ_98 CMRJ_98 OMRJ_98 SMRJ_98
#> MRJ_03 EMRJ_03 CMRJ_03 OMRJ_03 SMRJ_03
#> MRJ_08 EMRJ_08 CMRJ_08 OMRJ_08 SMRJ_08
# Label class1: nonuse
#       class2: heavy drinking only
#       class3: not heavy use
#       class4: heavy drinking & smoking
#       class5: heavy use

# Reorder the TAU paramters for profiles as ascending order
reordered3 <- reordered2 %>%
   reorder(PROF = c(4, 1, 3, 2))
reordered3 %>% param
#> PI :
#> (PROF)
#>   class
#>          1       2       3       4
#>     0.2195  0.2790  0.2869  0.2146
#> 
#> TAU :
#> (A)
#>      parent
#> child       1       2       3       4
#>     1  0.7018  0.3771  0.1825  0.1675
#>     2  0.0813  0.2620  0.0000  0.0000
#>     3  0.1766  0.2709  0.4491  0.3869
#>     4  0.0000  0.0000  0.1745  0.0000
#>     5  0.0403  0.0901  0.1939  0.4456
#>              
#> parent PROF  
#> child  SUB_98
#> (B)
#>      parent
#> child       1       2       3       4
#>     1  0.9210  0.0000  0.0965  0.0000
#>     2  0.0711  0.9941  0.0000  0.0000
#>     3  0.0000  0.0059  0.0316  0.0269
#>     4  0.0000  0.0000  0.8719  0.0114
#>     5  0.0079  0.0000  0.0000  0.9617
#>              
#> parent PROF  
#> child  SUB_03
#> (C)
#>      parent
#> child       1       2       3       4
#>     1  0.7168  0.0573  0.1844  0.0442
#>     2  0.2473  0.9311  0.0084  0.0356
#>     3  0.0040  0.0000  0.0048  0.0179
#>     4  0.0319  0.0019  0.8024  0.1821
#>     5  0.0000  0.0097  0.0000  0.7202
#>              
#> parent PROF  
#> child  SUB_08
#> (D)
#>      parent
#> child       1       2       3       4       5
#>     1  0.9505  0.7958  0.1613  0.0254  0.0457
#>     2  0.0444  0.2042  0.6755  0.2230  0.2853
#>     3  0.0051  0.0000  0.1631  0.7516  0.6690
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  SMK_98 SMK_03 SMK_08
#> (E)
#>      parent
#> child       1       2       3       4       5
#>     1  0.7075  0.0065  0.1001  0.2107  0.0203
#>     2  0.2589  0.2414  0.6127  0.2229  0.1264
#>     3  0.0335  0.7521  0.2872  0.5664  0.8533
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  DRK_98 DRK_03 DRK_08
#> (F)
#>      parent
#> child       1       2       3       4       5
#>     1  0.9588  0.6173  0.1463  0.8512  0.0170
#>     2  0.0317  0.1525  0.7278  0.0871  0.1442
#>     3  0.0095  0.2302  0.1258  0.0617  0.8387
#>                            
#> parent SUB_98 SUB_03 SUB_08
#> child  MRJ_98 MRJ_03 MRJ_08
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  0.0134  1.0000  1.0000
#>    2      0.9866  0.0000  0.0000
#>    1(V2)  0.0000  0.6701  1.0000
#>    2      1.0000  0.3299  0.0000
#>    1(V3)  0.0000  0.0000  0.8833
#>    2      1.0000  1.0000  0.1167
#>    1(V4)  0.0000  0.0000  0.6460
#>    2      1.0000  1.0000  0.3540
#> 
#>        V1      V2      V3      V4     
#> SMK_98 ESMK_98 FSMK_98 DSMK_98 HSMK_98
#> SMK_03 ESMK_03 FSMK_03 DSMK_03 HSMK_03
#> SMK_08 ESMK_08 FSMK_08 DSMK_08 HSMK_08
#> (b)
#>         class
#> response       1       2       3
#>    1(V1)  0.0759  1.0000  1.0000
#>    2      0.9241  0.0000  0.0000
#>    1(V2)  0.0000  0.5760  1.0000
#>    2      1.0000  0.4240  0.0000
#>    1(V3)  0.0000  0.0000  0.6934
#>    2      1.0000  1.0000  0.3066
#>    1(V4)  0.0000  0.0000  0.8061
#>    2      1.0000  1.0000  0.1939
#> 
#>        V1      V2      V3      V4     
#> DRK_98 EDRK_98 CDRK_98 WDRK_98 BDRK_98
#> DRK_03 EDRK_03 CDRK_03 WDRK_03 BDRK_03
#> DRK_08 EDRK_08 CDRK_08 WDRK_08 BDRK_08
#> (c)
#>         class
#> response       1       2       3
#>    1(V1)  0.0002  0.5189  1.0000
#>    2      0.9998  0.4811  0.0000
#>    1(V2)  0.0000  0.0000  1.0000
#>    2      1.0000  1.0000  0.0000
#>    1(V3)  0.0000  0.0000  0.5259
#>    2      1.0000  1.0000  0.4741
#>    1(V4)  0.0000  0.0000  0.3720
#>    2      1.0000  1.0000  0.6280
#> 
#>        V1      V2      V3      V4     
#> MRJ_98 EMRJ_98 CMRJ_98 OMRJ_98 SMRJ_98
#> MRJ_03 EMRJ_03 CMRJ_03 OMRJ_03 SMRJ_03
#> MRJ_08 EMRJ_08 CMRJ_08 OMRJ_08 SMRJ_08
# Label class1: nonuse stayer
#       class2: heavy drinking advancer
#       class3: heavy drk & smk advancer
#       class4: heavy use advancer