Add Health data about adolescents delinquency.

This dataset contains responses from the National Longitudinal Study of Adolescent Health (Add Health) study. The subjects were in Grades 10 and 11 during the 1994-1995 academic year and provided data on at least one variable measuring adolescent delinquency at Wave 1. The latent class models can be fitted using this dataset to replicate the analysis conducted by Collins and Lanza (2008).
The data includes two covariates: grade and sex of respondents. The variables measuring adolescent delinquency include: lying to parents about where or with whom they were, acting loud, rowdy, or unruly in public, damaging property, stealing something from a store, stealing something worth less than $50, and taking part in a group fight. These variables were originally coded as "Never", “1-2 times,” “3-4 times,” or “5 or more times.” In this dataset, those responses are recoded to "No" for "Never" and "Yes" for all other responses from both Waves I and II, providing a longitudinal pattern of adolescent delinquency. Variables ending in "w1" were obtained from Wave I, while variables ending in "w2" were obtained from Wave II.

Format

A data frame with 2087 rows and 14 variables:

GRADE

Respondent's grade level at Wave I

SEX

Respondent's sex
levels: (1)Male, (2)Female

LIEw1, LIEw2

Lying to parents about where or with whom they were

LOUDw1, LOUDw2

Acting loud, rowdy, or unruly in public

DAMAGEw1, DAMAGEw2

Damaging property

SHOPLIFTw1, SHOPLIFTw2

Stealing something from a store

STEALw1, STEALw2

Stealing something worth less than $50

FIGHTw1, FIGHTw2

Taking part in a group fight

Source

https://addhealth.cpc.unc.edu/data/#public-use

References

Collins, L.M., & Lanza, S.T. (2009). Latent Class and Latent Transition Analysis: With Applications in the Social, Behavioral, and Health Sciences.

J.R. Udry. The National Longitudinal Study of Adolescent Health (Add Health), Waves I & II, 1994-1996 machine-readable data file and documentation. Carolina Population Center, University of North Carolina at Chapel Hill, Chapel Hill, NC, 2003.

Examples

library(magrittr)
f5_var <- slca(
   DLQ1(5) ~ LIEw1 + LOUDw1 + DAMAGEw1 + SHOPLIFTw1 + STEALw1 + FIGHTw1,
   DLQ2(5) ~ LIEw2 + LOUDw2 + DAMAGEw2 + SHOPLIFTw2 + STEALw2 + FIGHTw2,
   DLQ1 ~ DLQ2
) %>% estimate(addhealth[1:300,])
#> Error in eval(predvars, data, env): object 'LIEw1' not found

f5_inv <- slca(
   DLQ1(5) ~ LIEw1 + LOUDw1 + DAMAGEw1 + SHOPLIFTw1 + STEALw1 + FIGHTw1,
   DLQ2(5) ~ LIEw2 + LOUDw2 + DAMAGEw2 + SHOPLIFTw2 + STEALw2 + FIGHTw2,
   DLQ1 ~ DLQ2,
   constraints = c("DLQ1", "DLQ2")
) %>% estimate(addhealth[1:300,])
#> Error in eval(predvars, data, env): object 'LIEw1' not found

compare(f5_var, f5_inv, test = "chisq")
#> Error: object 'f5_var' not found
param(f5_inv)
#> Error: object 'f5_inv' not found