Adolescent Depression Data from the Add Health Study

This dataset includes responses from the National Longitudinal Study of Adolescent Health (Add Health), focusing on adolescents' experiences with depression. The subjects, who were in Grades 10 and 11 during the 1994-1995 academic year, provided data on at least one measure of adolescent delinquency in Wave 1. These data can be used to replicate the latent class analysis conducted by Collins and Lanza (2008).
The dataset includes five covariates, notably grade level and sex of respondents, along with variables capturing depressive emotions: sadness (S1-S4), feeling disliked (D1-D2), and feelings of failure (F1-F2).

Format

A data frame with 2061 rows and 18 variables:

GRADE

Respondent's grade level at Wave I

SEX

Respondent's sex
levels: (1)Male, (2)Female

S1w1, S1w2

I felt that I could not shake off the blues even with help from my family and friends.

S2w1, S2w2

I felt depressed.

S3w1, S3w2

I felt lonely.

S4w1, S4w2

I felt sad.

D1w1, D1w2

People were unfriendly to me.

D2w1, D2w2

I felt that people disliked me

F1w1, F1w2

I thought my life had been a failure.

F2w1, F2w2

I felt life was not worth living

Source

https://addhealth.cpc.unc.edu/data/#public-use

Details

Responses for these variables were initially categorized as "Never," "Sometimes," "Often," or "Most or All of the Time." In this dataset, responses are recoded as "No" for "Never" and "Yes" for all other responses, providing a longitudinal perspective on adolescent depression from Waves I and II. Variables with suffix "w1" are from Wave I, and those with suffix "w2" are from Wave II.

References

Collins, L.M., & Lanza, S.T. (2009). Latent Class and Latent Transition Analysis: With Applications in the Social, Behavioral, and Health Sciences.

J.R. Udry. The National Longitudinal Study of Adolescent Health (Add Health), Waves I & II, 1994-1996. Carolina Population Center, University of North Carolina at Chapel Hill, Chapel Hill, NC, 2003.

Examples

library(magrittr)
lta5 <- slca(
   DEP1(5) ~ S1w1 + S2w1 + S3w1 + S4w1 + D1w1 + D2w1 + F1w1 + F2w1,
   DEP2(5) ~ S1w2 + S2w2 + S3w2 + S4w2 + D1w2 + D2w2 + F1w2 + F2w2,
   DEP1 ~ DEP2
) %>% estimate(addhealth)
#> EM iteration begin.
#> 1 iterations  logLik: -21440.38  diff: inf      2 iterations  logLik: -18186.33  diff: 3254.04887      3 iterations  logLik: -17945.44  diff: 240.89628      4 iterations  logLik: -16907.69  diff: 1037.74852      5 iterations  logLik: -15510.87  diff: 1396.81328      6 iterations  logLik: -15197.66  diff: 313.21581      7 iterations  logLik: -15124.73  diff: 72.92762      8 iterations  logLik: -15054.47  diff: 70.25983      9 iterations  logLik: -14980.31  diff: 74.16477      10 iterations  logLik: -14911.19  diff: 69.11724      11 iterations  logLik: -14852.73  diff: 58.45643      12 iterations  logLik: -14806.30  diff: 46.43064      13 iterations  logLik: -14771.11  diff: 35.18984      14 iterations  logLik: -14744.93  diff: 26.18353      15 iterations  logLik: -14725.08  diff: 19.84665      16 iterations  logLik: -14709.55  diff: 15.52937      17 iterations  logLik: -14697.21  diff: 12.34490      18 iterations  logLik: -14687.43  diff: 9.77394      19 iterations  logLik: -14679.76  diff: 7.67020      20 iterations  logLik: -14673.72  diff: 6.04292      21 iterations  logLik: -14668.83  diff: 4.89305      22 iterations  logLik: -14664.66  diff: 4.16415      23 iterations  logLik: -14660.90  diff: 3.76532      24 iterations  logLik: -14657.29  diff: 3.60605      25 iterations  logLik: -14653.68  diff: 3.61307      26 iterations  logLik: -14649.95  diff: 3.73056      27 iterations  logLik: -14646.03  diff: 3.91389      28 iterations  logLik: -14641.91  diff: 4.12423      29 iterations  logLik: -14637.58  diff: 4.32629      30 iterations  logLik: -14633.09  diff: 4.48843      31 iterations  logLik: -14628.51  diff: 4.58377      32 iterations  logLik: -14623.92  diff: 4.59225      33 iterations  logLik: -14619.41  diff: 4.50453      34 iterations  logLik: -14615.09  diff: 4.32599      35 iterations  logLik: -14611.01  diff: 4.07665      36 iterations  logLik: -14607.23  diff: 3.78504      37 iterations  logLik: -14603.75  diff: 3.47923      38 iterations  logLik: -14600.57  diff: 3.18020      39 iterations  logLik: -14597.67  diff: 2.90003      40 iterations  logLik: -14595.02  diff: 2.64373      41 iterations  logLik: -14592.61  diff: 2.41233      42 iterations  logLik: -14590.41  diff: 2.20510      43 iterations  logLik: -14588.39  diff: 2.02059      44 iterations  logLik: -14586.53  diff: 1.85681      45 iterations  logLik: -14584.82  diff: 1.71127      46 iterations  logLik: -14583.24  diff: 1.58106      47 iterations  logLik: -14581.77  diff: 1.46314      48 iterations  logLik: -14580.42  diff: 1.35466      49 iterations  logLik: -14579.17  diff: 1.25328      50 iterations  logLik: -14578.01  diff: 1.15730      51 iterations  logLik: -14576.94  diff: 1.06573      52 iterations  logLik: -14575.96  diff: 0.97819      53 iterations  logLik: -14575.07  diff: 0.89480      54 iterations  logLik: -14574.25  diff: 0.81591      55 iterations  logLik: -14573.51  diff: 0.74199      56 iterations  logLik: -14572.84  diff: 0.67343      57 iterations  logLik: -14572.23  diff: 0.61053      58 iterations  logLik: -14571.67  diff: 0.55336      59 iterations  logLik: -14571.17  diff: 0.50188      60 iterations  logLik: -14570.72  diff: 0.45584      61 iterations  logLik: -14570.30  diff: 0.41494      62 iterations  logLik: -14569.92  diff: 0.37876      63 iterations  logLik: -14569.58  diff: 0.34688      64 iterations  logLik: -14569.26  diff: 0.31883      65 iterations  logLik: -14568.96  diff: 0.29420      66 iterations  logLik: -14568.69  diff: 0.27256      67 iterations  logLik: -14568.44  diff: 0.25355      68 iterations  logLik: -14568.20  diff: 0.23683      69 iterations  logLik: -14567.98  diff: 0.22209      70 iterations  logLik: -14567.77  diff: 0.20910      71 iterations  logLik: -14567.57  diff: 0.19762      72 iterations  logLik: -14567.38  diff: 0.18748      73 iterations  logLik: -14567.20  diff: 0.17851      74 iterations  logLik: -14567.03  diff: 0.17058      75 iterations  logLik: -14566.87  diff: 0.16359      76 iterations  logLik: -14566.71  diff: 0.15745      77 iterations  logLik: -14566.56  diff: 0.15209      78 iterations  logLik: -14566.41  diff: 0.14745      79 iterations  logLik: -14566.27  diff: 0.14348      80 iterations  logLik: -14566.13  diff: 0.14016      81 iterations  logLik: -14565.99  diff: 0.13747      82 iterations  logLik: -14565.86  diff: 0.13538      83 iterations  logLik: -14565.72  diff: 0.13390      84 iterations  logLik: -14565.59  diff: 0.13303      85 iterations  logLik: -14565.46  diff: 0.13277      86 iterations  logLik: -14565.32  diff: 0.13316      87 iterations  logLik: -14565.19  diff: 0.13421      88 iterations  logLik: -14565.05  diff: 0.13598      89 iterations  logLik: -14564.92  diff: 0.13851      90 iterations  logLik: -14564.77  diff: 0.14187      91 iterations  logLik: -14564.63  diff: 0.14613      92 iterations  logLik: -14564.48  diff: 0.15141      93 iterations  logLik: -14564.32  diff: 0.15781      94 iterations  logLik: -14564.15  diff: 0.16547      95 iterations  logLik: -14563.98  diff: 0.17457      96 iterations  logLik: -14563.79  diff: 0.18530      97 iterations  logLik: -14563.60  diff: 0.19791      98 iterations  logLik: -14563.38  diff: 0.21266      99 iterations  logLik: -14563.15  diff: 0.22989      100 iterations  logLik: -14562.90  diff: 0.24995      101 iterations  logLik: -14562.63  diff: 0.27329      102 iterations  logLik: -14562.33  diff: 0.30037      103 iterations  logLik: -14562.00  diff: 0.33172      104 iterations  logLik: -14561.63  diff: 0.36789      105 iterations  logLik: -14561.22  diff: 0.40943      106 iterations  logLik: -14560.76  diff: 0.45686      107 iterations  logLik: -14560.25  diff: 0.51058      108 iterations  logLik: -14559.68  diff: 0.57082      109 iterations  logLik: -14559.04  diff: 0.63751      110 iterations  logLik: -14558.33  diff: 0.71016      111 iterations  logLik: -14557.55  diff: 0.78779      112 iterations  logLik: -14556.68  diff: 0.86883      113 iterations  logLik: -14555.73  diff: 0.95116      114 iterations  logLik: -14554.69  diff: 1.03228      115 iterations  logLik: -14553.58  diff: 1.10955      116 iterations  logLik: -14552.40  diff: 1.18057      117 iterations  logLik: -14551.16  diff: 1.24348      118 iterations  logLik: -14549.86  diff: 1.29714      119 iterations  logLik: -14548.52  diff: 1.34112      120 iterations  logLik: -14547.15  diff: 1.37541      121 iterations  logLik: -14545.75  diff: 1.40018      122 iterations  logLik: -14544.33  diff: 1.41553      123 iterations  logLik: -14542.91  diff: 1.42141      124 iterations  logLik: -14541.49  diff: 1.41776      125 iterations  logLik: -14540.09  diff: 1.40481      126 iterations  logLik: -14538.70  diff: 1.38333      127 iterations  logLik: -14537.35  diff: 1.35485      128 iterations  logLik: -14536.03  diff: 1.32143      129 iterations  logLik: -14534.74  diff: 1.28529      130 iterations  logLik: -14533.49  diff: 1.24826      131 iterations  logLik: -14532.28  diff: 1.21139      132 iterations  logLik: -14531.11  diff: 1.17491      133 iterations  logLik: -14529.97  diff: 1.13840      134 iterations  logLik: -14528.87  diff: 1.10109      135 iterations  logLik: -14527.81  diff: 1.06216      136 iterations  logLik: -14526.78  diff: 1.02097      137 iterations  logLik: -14525.81  diff: 0.97716      138 iterations  logLik: -14524.88  diff: 0.93070      139 iterations  logLik: -14524.00  diff: 0.88186      140 iterations  logLik: -14523.16  diff: 0.83117      141 iterations  logLik: -14522.38  diff: 0.77930      142 iterations  logLik: -14521.66  diff: 0.72702      143 iterations  logLik: -14520.98  diff: 0.67507      144 iterations  logLik: -14520.36  diff: 0.62419      145 iterations  logLik: -14519.78  diff: 0.57499      146 iterations  logLik: -14519.26  diff: 0.52799      147 iterations  logLik: -14518.77  diff: 0.48358      148 iterations  logLik: -14518.33  diff: 0.44201      149 iterations  logLik: -14517.93  diff: 0.40344      150 iterations  logLik: -14517.56  diff: 0.36789      151 iterations  logLik: -14517.22  diff: 0.33534      152 iterations  logLik: -14516.92  diff: 0.30565      153 iterations  logLik: -14516.64  diff: 0.27866      154 iterations  logLik: -14516.38  diff: 0.25419      155 iterations  logLik: -14516.15  diff: 0.23201      156 iterations  logLik: -14515.94  diff: 0.21194      157 iterations  logLik: -14515.75  diff: 0.19375      158 iterations  logLik: -14515.57  diff: 0.17727      159 iterations  logLik: -14515.41  diff: 0.16231      160 iterations  logLik: -14515.26  diff: 0.14873      161 iterations  logLik: -14515.12  diff: 0.13639      162 iterations  logLik: -14515.00  diff: 0.12516      163 iterations  logLik: -14514.88  diff: 0.11493      164 iterations  logLik: -14514.78  diff: 0.10561      165 iterations  logLik: -14514.68  diff: 0.09710      166 iterations  logLik: -14514.59  diff: 0.08933      167 iterations  logLik: -14514.51  diff: 0.08223      168 iterations  logLik: -14514.43  diff: 0.07574      169 iterations  logLik: -14514.36  diff: 0.06980      170 iterations  logLik: -14514.30  diff: 0.06436      171 iterations  logLik: -14514.24  diff: 0.05937      172 iterations  logLik: -14514.18  diff: 0.05480      173 iterations  logLik: -14514.13  diff: 0.05061      174 iterations  logLik: -14514.09  diff: 0.04677      175 iterations  logLik: -14514.04  diff: 0.04323      176 iterations  logLik: -14514.00  diff: 0.03998      177 iterations  logLik: -14513.97  diff: 0.03700      178 iterations  logLik: -14513.93  diff: 0.03425      179 iterations  logLik: -14513.90  diff: 0.03172      180 iterations  logLik: -14513.87  diff: 0.02938      181 iterations  logLik: -14513.84  diff: 0.02723      182 iterations  logLik: -14513.82  diff: 0.02525      183 iterations  logLik: -14513.80  diff: 0.02342      184 iterations  logLik: -14513.77  diff: 0.02173      185 iterations  logLik: -14513.75  diff: 0.02017      186 iterations  logLik: -14513.73  diff: 0.01872      187 iterations  logLik: -14513.72  diff: 0.01739      188 iterations  logLik: -14513.70  diff: 0.01615      189 iterations  logLik: -14513.69  diff: 0.01501      190 iterations  logLik: -14513.67  diff: 0.01395      191 iterations  logLik: -14513.66  diff: 0.01297      192 iterations  logLik: -14513.65  diff: 0.01206      193 iterations  logLik: -14513.64  diff: 0.01122      194 iterations  logLik: -14513.63  diff: 0.01043      195 iterations  logLik: -14513.62  diff: 0.00971      196 iterations  logLik: -14513.61  diff: 0.00904      197 iterations  logLik: -14513.60  diff: 0.00841      198 iterations  logLik: -14513.59  diff: 0.00783      199 iterations  logLik: -14513.58  diff: 0.00729      200 iterations  logLik: -14513.58  diff: 0.00679      201 iterations  logLik: -14513.57  diff: 0.00632      202 iterations  logLik: -14513.56  diff: 0.00589      203 iterations  logLik: -14513.56  diff: 0.00549      204 iterations  logLik: -14513.55  diff: 0.00511      205 iterations  logLik: -14513.55  diff: 0.00477      206 iterations  logLik: -14513.54  diff: 0.00444      207 iterations  logLik: -14513.54  diff: 0.00414      208 iterations  logLik: -14513.54  diff: 0.00386      209 iterations  logLik: -14513.53  diff: 0.00360      210 iterations  logLik: -14513.53  diff: 0.00336      211 iterations  logLik: -14513.53  diff: 0.00313      212 iterations  logLik: -14513.52  diff: 0.00292      213 iterations  logLik: -14513.52  diff: 0.00272      214 iterations  logLik: -14513.52  diff: 0.00254      215 iterations  logLik: -14513.52  diff: 0.00237      216 iterations  logLik: -14513.51  diff: 0.00221      217 iterations  logLik: -14513.51  diff: 0.00206      218 iterations  logLik: -14513.51  diff: 0.00192      219 iterations  logLik: -14513.51  diff: 0.00179      220 iterations  logLik: -14513.51  diff: 0.00167      221 iterations  logLik: -14513.50  diff: 0.00156      222 iterations  logLik: -14513.50  diff: 0.00146      223 iterations  logLik: -14513.50  diff: 0.00136      224 iterations  logLik: -14513.50  diff: 0.00127      225 iterations  logLik: -14513.50  diff: 0.00119      226 iterations  logLik: -14513.50  diff: 0.00111      227 iterations  logLik: -14513.50  diff: 0.00103      228 iterations  logLik: -14513.50  diff: 0.00096      229 iterations  logLik: -14513.49  diff: 0.00090      230 iterations  logLik: -14513.49  diff: 0.00084      231 iterations  logLik: -14513.49  diff: 0.00078      232 iterations  logLik: -14513.49  diff: 0.00073      233 iterations  logLik: -14513.49  diff: 0.00068      234 iterations  logLik: -14513.49  diff: 0.00064      235 iterations  logLik: -14513.49  diff: 0.00060      236 iterations  logLik: -14513.49  diff: 0.00056      237 iterations  logLik: -14513.49  diff: 0.00052      238 iterations  logLik: -14513.49  diff: 0.00049      239 iterations  logLik: -14513.49  diff: 0.00045      240 iterations  logLik: -14513.49  diff: 0.00042      241 iterations  logLik: -14513.49  diff: 0.00040      242 iterations  logLik: -14513.49  diff: 0.00037      243 iterations  logLik: -14513.49  diff: 0.00035      244 iterations  logLik: -14513.49  diff: 0.00032      245 iterations  logLik: -14513.49  diff: 0.00030      246 iterations  logLik: -14513.49  diff: 0.00028      247 iterations  logLik: -14513.49  diff: 0.00026      248 iterations  logLik: -14513.49  diff: 0.00025      249 iterations  logLik: -14513.49  diff: 0.00023      250 iterations  logLik: -14513.49  diff: 0.00021      251 iterations  logLik: -14513.49  diff: 0.00020      252 iterations  logLik: -14513.48  diff: 0.00019      253 iterations  logLik: -14513.48  diff: 0.00017      254 iterations  logLik: -14513.48  diff: 0.00016      255 iterations  logLik: -14513.48  diff: 0.00015      256 iterations  logLik: -14513.48  diff: 0.00014      257 iterations  logLik: -14513.48  diff: 0.00013      258 iterations  logLik: -14513.48  diff: 0.00012      259 iterations  logLik: -14513.48  diff: 0.00012      260 iterations  logLik: -14513.48  diff: 0.00011      261 iterations  logLik: -14513.48  diff: 0.00010      262 iterations  logLik: -14513.48  diff: 0.00009      263 iterations  logLik: -14513.48  diff: 0.00009      264 iterations  logLik: -14513.48  diff: 0.00008      265 iterations  logLik: -14513.48  diff: 0.00008      266 iterations  logLik: -14513.48  diff: 0.00007      267 iterations  logLik: -14513.48  diff: 0.00007      268 iterations  logLik: -14513.48  diff: 0.00006      269 iterations  logLik: -14513.48  diff: 0.00006      270 iterations  logLik: -14513.48  diff: 0.00006      271 iterations  logLik: -14513.48  diff: 0.00005      272 iterations  logLik: -14513.48  diff: 0.00005      273 iterations  logLik: -14513.48  diff: 0.00004      274 iterations  logLik: -14513.48  diff: 0.00004      275 iterations  logLik: -14513.48  diff: 0.00004      276 iterations  logLik: -14513.48  diff: 0.00004      277 iterations  logLik: -14513.48  diff: 0.00003      278 iterations  logLik: -14513.48  diff: 0.00003      279 iterations  logLik: -14513.48  diff: 0.00003      280 iterations  logLik: -14513.48  diff: 0.00003      281 iterations  logLik: -14513.48  diff: 0.00003      282 iterations  logLik: -14513.48  diff: 0.00002      283 iterations  logLik: -14513.48  diff: 0.00002      284 iterations  logLik: -14513.48  diff: 0.00002      285 iterations  logLik: -14513.48  diff: 0.00002      286 iterations  logLik: -14513.48  diff: 0.00002      287 iterations  logLik: -14513.48  diff: 0.00002      288 iterations  logLik: -14513.48  diff: 0.00002      289 iterations  logLik: -14513.48  diff: 0.00002      290 iterations  logLik: -14513.48  diff: 0.00001      291 iterations  logLik: -14513.48  diff: 0.00001      292 iterations  logLik: -14513.48  diff: 0.00001      293 iterations  logLik: -14513.48  diff: 0.00001      294 iterations  logLik: -14513.48  diff: 0.00001      295 iterations  logLik: -14513.48  diff: 0.00001      296 iterations  logLik: -14513.48  diff: 9.5e-06      297 iterations  logLik: -14513.48  diff: 8.8e-06      298 iterations  logLik: -14513.48  diff: 8.3e-06      299 iterations  logLik: -14513.48  diff: 7.7e-06      300 iterations  logLik: -14513.48  diff: 7.2e-06      301 iterations  logLik: -14513.48  diff: 6.7e-06      302 iterations  logLik: -14513.48  diff: 6.3e-06      303 iterations  logLik: -14513.48  diff: 5.9e-06      304 iterations  logLik: -14513.48  diff: 5.5e-06      305 iterations  logLik: -14513.48  diff: 5.1e-06      306 iterations  logLik: -14513.48  diff: 4.8e-06      307 iterations  logLik: -14513.48  diff: 4.5e-06      308 iterations  logLik: -14513.48  diff: 4.2e-06      309 iterations  logLik: -14513.48  diff: 3.9e-06      310 iterations  logLik: -14513.48  diff: 3.7e-06      311 iterations  logLik: -14513.48  diff: 3.4e-06      312 iterations  logLik: -14513.48  diff: 3.2e-06      313 iterations  logLik: -14513.48  diff: 3.0e-06      314 iterations  logLik: -14513.48  diff: 2.8e-06      315 iterations  logLik: -14513.48  diff: 2.6e-06      316 iterations  logLik: -14513.48  diff: 2.4e-06      317 iterations  logLik: -14513.48  diff: 2.3e-06      318 iterations  logLik: -14513.48  diff: 2.1e-06      319 iterations  logLik: -14513.48  diff: 2.0e-06      320 iterations  logLik: -14513.48  diff: 1.9e-06      321 iterations  logLik: -14513.48  diff: 1.7e-06      322 iterations  logLik: -14513.48  diff: 1.6e-06      323 iterations  logLik: -14513.48  diff: 1.5e-06      324 iterations  logLik: -14513.48  diff: 1.4e-06      325 iterations  logLik: -14513.48  diff: 1.3e-06      326 iterations  logLik: -14513.48  diff: 1.2e-06      327 iterations  logLik: -14513.48  diff: 1.2e-06      328 iterations  logLik: -14513.48  diff: 1.1e-06      329 iterations  logLik: -14513.48  diff: 1.0e-06      330 iterations  logLik: -14513.48  diff: 9.5e-07      
#> .. done.

lta5inv <- slca(
   DEP1(5) ~ S1w1 + S2w1 + S3w1 + S4w1 + D1w1 + D2w1 + F1w1 + F2w1,
   DEP2(5) ~ S1w2 + S2w2 + S3w2 + S4w2 + D1w2 + D2w2 + F1w2 + F2w2,
   DEP1 ~ DEP2,
   constraints = c("DEP1", "DEP2")
) %>% estimate(addhealth)
#> EM iteration begin.
#> 1 iterations  logLik: -21467.05  diff: inf      2 iterations  logLik: -18163.28  diff: 3303.76619      3 iterations  logLik: -17842.96  diff: 320.32477      4 iterations  logLik: -16519.91  diff: 1323.04826      5 iterations  logLik: -15399.66  diff: 1120.24584      6 iterations  logLik: -15244.04  diff: 155.62252      7 iterations  logLik: -15186.68  diff: 57.35903      8 iterations  logLik: -15107.58  diff: 79.10244      9 iterations  logLik: -15006.58  diff: 101.00272      10 iterations  logLik: -14907.30  diff: 99.27860      11 iterations  logLik: -14828.85  diff: 78.44784      12 iterations  logLik: -14772.08  diff: 56.77356      13 iterations  logLik: -14729.81  diff: 42.26479      14 iterations  logLik: -14696.98  diff: 32.83357      15 iterations  logLik: -14671.60  diff: 25.37834      16 iterations  logLik: -14652.51  diff: 19.09570      17 iterations  logLik: -14638.22  diff: 14.28400      18 iterations  logLik: -14627.24  diff: 10.98418      19 iterations  logLik: -14618.42  diff: 8.81629      20 iterations  logLik: -14611.09  diff: 7.33283      21 iterations  logLik: -14604.86  diff: 6.23082      22 iterations  logLik: -14599.50  diff: 5.35570      23 iterations  logLik: -14594.86  diff: 4.63959      24 iterations  logLik: -14590.81  diff: 4.05174      25 iterations  logLik: -14587.24  diff: 3.57272      26 iterations  logLik: -14584.05  diff: 3.18521      27 iterations  logLik: -14581.18  diff: 2.87280      28 iterations  logLik: -14578.56  diff: 2.62099      29 iterations  logLik: -14576.14  diff: 2.41786      30 iterations  logLik: -14573.89  diff: 2.25403      31 iterations  logLik: -14571.76  diff: 2.12216      32 iterations  logLik: -14569.75  diff: 2.01633      33 iterations  logLik: -14567.82  diff: 1.93156      34 iterations  logLik: -14565.95  diff: 1.86348      35 iterations  logLik: -14564.15  diff: 1.80813      36 iterations  logLik: -14562.38  diff: 1.76190      37 iterations  logLik: -14560.66  diff: 1.72140      38 iterations  logLik: -14558.98  diff: 1.68357      39 iterations  logLik: -14557.33  diff: 1.64560      40 iterations  logLik: -14555.73  diff: 1.60504      41 iterations  logLik: -14554.17  diff: 1.55989      42 iterations  logLik: -14552.66  diff: 1.50867      43 iterations  logLik: -14551.21  diff: 1.45051      44 iterations  logLik: -14549.82  diff: 1.38522      45 iterations  logLik: -14548.51  diff: 1.31328      46 iterations  logLik: -14547.27  diff: 1.23577      47 iterations  logLik: -14546.12  diff: 1.15419      48 iterations  logLik: -14545.05  diff: 1.07033      49 iterations  logLik: -14544.06  diff: 0.98603      50 iterations  logLik: -14543.16  diff: 0.90304      51 iterations  logLik: -14542.34  diff: 0.82285      52 iterations  logLik: -14541.59  diff: 0.74665      53 iterations  logLik: -14540.92  diff: 0.67527      54 iterations  logLik: -14540.31  diff: 0.60922      55 iterations  logLik: -14539.76  diff: 0.54870      56 iterations  logLik: -14539.26  diff: 0.49371      57 iterations  logLik: -14538.82  diff: 0.44404      58 iterations  logLik: -14538.42  diff: 0.39938      59 iterations  logLik: -14538.06  diff: 0.35936      60 iterations  logLik: -14537.74  diff: 0.32357      61 iterations  logLik: -14537.45  diff: 0.29158      62 iterations  logLik: -14537.18  diff: 0.26299      63 iterations  logLik: -14536.95  diff: 0.23744      64 iterations  logLik: -14536.73  diff: 0.21457      65 iterations  logLik: -14536.54  diff: 0.19408      66 iterations  logLik: -14536.36  diff: 0.17569      67 iterations  logLik: -14536.20  diff: 0.15916      68 iterations  logLik: -14536.06  diff: 0.14428      69 iterations  logLik: -14535.93  diff: 0.13086      70 iterations  logLik: -14535.81  diff: 0.11874      71 iterations  logLik: -14535.70  diff: 0.10779      72 iterations  logLik: -14535.60  diff: 0.09787      73 iterations  logLik: -14535.51  diff: 0.08888      74 iterations  logLik: -14535.43  diff: 0.08072      75 iterations  logLik: -14535.36  diff: 0.07332      76 iterations  logLik: -14535.29  diff: 0.06659      77 iterations  logLik: -14535.23  diff: 0.06048      78 iterations  logLik: -14535.18  diff: 0.05492      79 iterations  logLik: -14535.13  diff: 0.04986      80 iterations  logLik: -14535.08  diff: 0.04526      81 iterations  logLik: -14535.04  diff: 0.04108      82 iterations  logLik: -14535.00  diff: 0.03728      83 iterations  logLik: -14534.97  diff: 0.03381      84 iterations  logLik: -14534.94  diff: 0.03067      85 iterations  logLik: -14534.91  diff: 0.02781      86 iterations  logLik: -14534.89  diff: 0.02520      87 iterations  logLik: -14534.86  diff: 0.02284      88 iterations  logLik: -14534.84  diff: 0.02069      89 iterations  logLik: -14534.82  diff: 0.01874      90 iterations  logLik: -14534.81  diff: 0.01697      91 iterations  logLik: -14534.79  diff: 0.01537      92 iterations  logLik: -14534.78  diff: 0.01391      93 iterations  logLik: -14534.77  diff: 0.01259      94 iterations  logLik: -14534.75  diff: 0.01139      95 iterations  logLik: -14534.74  diff: 0.01030      96 iterations  logLik: -14534.74  diff: 0.00931      97 iterations  logLik: -14534.73  diff: 0.00842      98 iterations  logLik: -14534.72  diff: 0.00761      99 iterations  logLik: -14534.71  diff: 0.00688      100 iterations  logLik: -14534.71  diff: 0.00622      101 iterations  logLik: -14534.70  diff: 0.00562      102 iterations  logLik: -14534.70  diff: 0.00508      103 iterations  logLik: -14534.69  diff: 0.00459      104 iterations  logLik: -14534.69  diff: 0.00414      105 iterations  logLik: -14534.68  diff: 0.00374      106 iterations  logLik: -14534.68  diff: 0.00338      107 iterations  logLik: -14534.68  diff: 0.00305      108 iterations  logLik: -14534.67  diff: 0.00275      109 iterations  logLik: -14534.67  diff: 0.00249      110 iterations  logLik: -14534.67  diff: 0.00224      111 iterations  logLik: -14534.67  diff: 0.00202      112 iterations  logLik: -14534.67  diff: 0.00183      113 iterations  logLik: -14534.66  diff: 0.00165      114 iterations  logLik: -14534.66  diff: 0.00149      115 iterations  logLik: -14534.66  diff: 0.00134      116 iterations  logLik: -14534.66  diff: 0.00121      117 iterations  logLik: -14534.66  diff: 0.00109      118 iterations  logLik: -14534.66  diff: 0.00099      119 iterations  logLik: -14534.66  diff: 0.00089      120 iterations  logLik: -14534.66  diff: 0.00080      121 iterations  logLik: -14534.65  diff: 0.00072      122 iterations  logLik: -14534.65  diff: 0.00065      123 iterations  logLik: -14534.65  diff: 0.00059      124 iterations  logLik: -14534.65  diff: 0.00053      125 iterations  logLik: -14534.65  diff: 0.00048      126 iterations  logLik: -14534.65  diff: 0.00043      127 iterations  logLik: -14534.65  diff: 0.00039      128 iterations  logLik: -14534.65  diff: 0.00035      129 iterations  logLik: -14534.65  diff: 0.00032      130 iterations  logLik: -14534.65  diff: 0.00029      131 iterations  logLik: -14534.65  diff: 0.00026      132 iterations  logLik: -14534.65  diff: 0.00023      133 iterations  logLik: -14534.65  diff: 0.00021      134 iterations  logLik: -14534.65  diff: 0.00019      135 iterations  logLik: -14534.65  diff: 0.00017      136 iterations  logLik: -14534.65  diff: 0.00015      137 iterations  logLik: -14534.65  diff: 0.00014      138 iterations  logLik: -14534.65  diff: 0.00012      139 iterations  logLik: -14534.65  diff: 0.00011      140 iterations  logLik: -14534.65  diff: 0.00010      141 iterations  logLik: -14534.65  diff: 0.00009      142 iterations  logLik: -14534.65  diff: 0.00008      143 iterations  logLik: -14534.65  diff: 0.00007      144 iterations  logLik: -14534.65  diff: 0.00007      145 iterations  logLik: -14534.65  diff: 0.00006      146 iterations  logLik: -14534.65  diff: 0.00005      147 iterations  logLik: -14534.65  diff: 0.00005      148 iterations  logLik: -14534.65  diff: 0.00004      149 iterations  logLik: -14534.65  diff: 0.00004      150 iterations  logLik: -14534.65  diff: 0.00004      151 iterations  logLik: -14534.65  diff: 0.00003      152 iterations  logLik: -14534.65  diff: 0.00003      153 iterations  logLik: -14534.65  diff: 0.00003      154 iterations  logLik: -14534.65  diff: 0.00002      155 iterations  logLik: -14534.65  diff: 0.00002      156 iterations  logLik: -14534.65  diff: 0.00002      157 iterations  logLik: -14534.65  diff: 0.00002      158 iterations  logLik: -14534.65  diff: 0.00002      159 iterations  logLik: -14534.65  diff: 0.00001      160 iterations  logLik: -14534.65  diff: 0.00001      161 iterations  logLik: -14534.65  diff: 0.00001      162 iterations  logLik: -14534.65  diff: 0.00001      163 iterations  logLik: -14534.65  diff: 9.4e-06      164 iterations  logLik: -14534.65  diff: 8.4e-06      165 iterations  logLik: -14534.65  diff: 7.6e-06      166 iterations  logLik: -14534.65  diff: 6.9e-06      167 iterations  logLik: -14534.65  diff: 6.2e-06      168 iterations  logLik: -14534.65  diff: 5.6e-06      169 iterations  logLik: -14534.65  diff: 5.0e-06      170 iterations  logLik: -14534.65  diff: 4.5e-06      171 iterations  logLik: -14534.65  diff: 4.1e-06      172 iterations  logLik: -14534.65  diff: 3.7e-06      173 iterations  logLik: -14534.65  diff: 3.3e-06      174 iterations  logLik: -14534.65  diff: 3.0e-06      175 iterations  logLik: -14534.65  diff: 2.7e-06      176 iterations  logLik: -14534.65  diff: 2.4e-06      177 iterations  logLik: -14534.65  diff: 2.2e-06      178 iterations  logLik: -14534.65  diff: 2.0e-06      179 iterations  logLik: -14534.65  diff: 1.8e-06      180 iterations  logLik: -14534.65  diff: 1.6e-06      181 iterations  logLik: -14534.65  diff: 1.4e-06      182 iterations  logLik: -14534.65  diff: 1.3e-06      183 iterations  logLik: -14534.65  diff: 1.2e-06      184 iterations  logLik: -14534.65  diff: 1.1e-06      185 iterations  logLik: -14534.65  diff: 9.6e-07      
#> .. done.

compare(lta5inv, lta5, test = "chisq")
#> Analysis of Relative Model Fit
#> 
#> Model H0: lta5inv
#> Model H1: lta5
#>          Df logLik   AIC   BIC    Gsq Res. Df Pr(>Chi)
#> lta5inv  64 -14535 29197 29558                        
#> lta5    104 -14514 29235 29821 42.332      40   0.3707
lta5inv %>% param()
#> PI :
#> (DEP1)
#>   class
#>          1       2       3       4       5
#>     0.1826  0.1128  0.2307  0.3632  0.1106
#> 
#> TAU :
#> (A)
#>      parent
#> child       1       2       3       4       5
#>     1  0.4739  0.1649  0.0412  0.0857  0.0297
#>     2  0.1716  0.3595  0.0679  0.0206  0.1408
#>     3  0.0478  0.3350  0.5441  0.1242  0.1576
#>     4  0.2499  0.0326  0.3024  0.7543  0.0807
#>     5  0.0569  0.1081  0.0444  0.0152  0.5911
#>            
#> parent DEP1
#> child  DEP2
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3       4       5
#>    1(V1)  0.1184  0.6930  0.5149  0.0342  0.8118
#>    2      0.8816  0.3070  0.4851  0.9658  0.1882
#>    1(V2)  0.1827  0.8401  0.6798  0.0568  0.9393
#>    2      0.8173  0.1599  0.3202  0.9432  0.0607
#>    1(V3)  0.2704  0.8385  0.5458  0.0640  0.8472
#>    2      0.7296  0.1615  0.4542  0.9360  0.1528
#>    1(V4)  0.3407  0.9198  0.7948  0.1067  0.9423
#>    2      0.6593  0.0802  0.2052  0.8933  0.0577
#>    1(V5)  0.6809  0.7640  0.1602  0.0912  0.6491
#>    2      0.3191  0.2360  0.8398  0.9088  0.3509
#>    1(V6)  0.6704  0.8825  0.1253  0.0271  0.7926
#>    2      0.3296  0.1175  0.8747  0.9729  0.2074
#>    1(V7)  0.0767  0.1260  0.1297  0.0136  0.8885
#>    2      0.9233  0.8740  0.8703  0.9864  0.1115
#>    1(V8)  0.0475  0.0576  0.0554  0.0034  0.6961
#>    2      0.9525  0.9424  0.9446  0.9966  0.3039
#> 
#>      V1   V2   V3   V4   V5   V6   V7   V8  
#> DEP1 S1w1 S2w1 S3w1 S4w1 D1w1 D2w1 F1w1 F2w1
#> DEP2 S1w2 S2w2 S3w2 S4w2 D1w2 D2w2 F1w2 F2w2