Adolescent Depression Data from the Add Health Study

This dataset includes responses from the National Longitudinal Study of Adolescent Health (Add Health), focusing on adolescents' experiences with depression. The subjects, who were in Grades 10 and 11 during the 1994-1995 academic year, provided data on at least one measure of adolescent delinquency in Wave 1. These data can be used to replicate the latent class analysis conducted by Collins and Lanza (2008).
The dataset includes five covariates, notably grade level and sex of respondents, along with variables capturing depressive emotions: sadness (S1-S4), feeling disliked (D1-D2), and feelings of failure (F1-F2).

Format

A data frame with 2061 rows and 18 variables:

GRADE

Respondent's grade level at Wave I

SEX

Respondent's sex
levels: (1)Male, (2)Female

S1w1, S1w2

I felt that I could not shake off the blues even with help from my family and friends.

S2w1, S2w2

I felt depressed.

S3w1, S3w2

I felt lonely.

S4w1, S4w2

I felt sad.

D1w1, D1w2

People were unfriendly to me.

D2w1, D2w2

I felt that people disliked me

F1w1, F1w2

I thought my life had been a failure.

F2w1, F2w2

I felt life was not worth living

Source

https://addhealth.cpc.unc.edu/data/#public-use

Details

Responses for these variables were initially categorized as "Never," "Sometimes," "Often," or "Most or All of the Time." In this dataset, responses are recoded as "No" for "Never" and "Yes" for all other responses, providing a longitudinal perspective on adolescent depression from Waves I and II. Variables with suffix "w1" are from Wave I, and those with suffix "w2" are from Wave II.

References

Collins, L.M., & Lanza, S.T. (2009). Latent Class and Latent Transition Analysis: With Applications in the Social, Behavioral, and Health Sciences.

J.R. Udry. The National Longitudinal Study of Adolescent Health (Add Health), Waves I & II, 1994-1996. Carolina Population Center, University of North Carolina at Chapel Hill, Chapel Hill, NC, 2003.

Examples

library(magrittr)
lta5 <- slca(
   DEP1(5) ~ S1w1 + S2w1 + S3w1 + S4w1 + D1w1 + D2w1 + F1w1 + F2w1,
   DEP2(5) ~ S1w2 + S2w2 + S3w2 + S4w2 + D1w2 + D2w2 + F1w2 + F2w2,
   DEP1 ~ DEP2
) %>% estimate(addhealth)

lta5inv <- slca(
   DEP1(5) ~ S1w1 + S2w1 + S3w1 + S4w1 + D1w1 + D2w1 + F1w1 + F2w1,
   DEP2(5) ~ S1w2 + S2w2 + S3w2 + S4w2 + D1w2 + D2w2 + F1w2 + F2w2,
   DEP1 ~ DEP2,
   constraints = c("DEP1", "DEP2")
) %>% estimate(addhealth)

compare(lta5inv, lta5, test = "chisq")
#> Analysis of Relative Model Fit
#> 
#> Model H0: lta5inv
#> Model H1: lta5
#>          Df logLik   AIC   BIC    Gsq Res. Df Pr(>Chi)
#> lta5inv  64 -14535 29197 29558                        
#> lta5    104 -14514 29235 29821 42.332      40   0.3707
lta5inv %>% param()
#> PI :
#> (DEP1)
#>   class
#>          1       2       3       4       5
#>     0.1826  0.1128  0.2307  0.3632  0.1106
#> 
#> TAU :
#> (A)
#>      parent
#> child       1       2       3       4       5
#>     1  0.4739  0.1650  0.0412  0.0857  0.0297
#>     2  0.1716  0.3595  0.0679  0.0206  0.1408
#>     3  0.0478  0.3350  0.5442  0.1242  0.1576
#>     4  0.2499  0.0325  0.3024  0.7543  0.0808
#>     5  0.0569  0.1081  0.0444  0.0152  0.5911
#>            
#> parent DEP1
#> child  DEP2
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3       4       5
#>    1(V1)  0.1184  0.6930  0.5150  0.0342  0.8118
#>    2      0.8816  0.3070  0.4850  0.9658  0.1882
#>    1(V2)  0.1826  0.8401  0.6799  0.0568  0.9393
#>    2      0.8174  0.1599  0.3201  0.9432  0.0607
#>    1(V3)  0.2704  0.8385  0.5458  0.0640  0.8472
#>    2      0.7296  0.1615  0.4542  0.9360  0.1528
#>    1(V4)  0.3407  0.9198  0.7948  0.1067  0.9423
#>    2      0.6593  0.0802  0.2052  0.8933  0.0577
#>    1(V5)  0.6809  0.7641  0.1603  0.0912  0.6491
#>    2      0.3191  0.2359  0.8397  0.9088  0.3509
#>    1(V6)  0.6705  0.8826  0.1253  0.0271  0.7926
#>    2      0.3295  0.1174  0.8747  0.9729  0.2074
#>    1(V7)  0.0767  0.1260  0.1297  0.0136  0.8885
#>    2      0.9233  0.8740  0.8703  0.9864  0.1115
#>    1(V8)  0.0475  0.0576  0.0554  0.0034  0.6961
#>    2      0.9525  0.9424  0.9446  0.9966  0.3039
#> 
#>      V1   V2   V3   V4   V5   V6   V7   V8  
#> DEP1 S1w1 S2w1 S3w1 S4w1 D1w1 D2w1 F1w1 F2w1
#> DEP2 S1w2 S2w2 S3w2 S4w2 D1w2 D2w2 F1w2 F2w2