Adolescent Depression Data from the Add Health Study

This dataset contains responses from the National Longitudinal Study of Adolescent Health (Add Health), focusing on adolescents' experiences with depression. The subjects, who were in Grades 10 and 11 during the 1994–1995 academic year, provided data on at least one measure of adolescent delinquency in Wave I.
These data can be used to replicate the latent class analysis conducted by Collins and Lanza (2009).
The dataset includes five covariates, notably grade level and sex of respondents, along with variables capturing depressive emotions: sadness (S1-S4), feeling disliked (D1-D2), and feelings of failure (F1-F2).
Responses for these variables were initially categorized as "Never," "Sometimes," "Often," or "Most or All of the Time." In this dataset, responses have been recoded as "No" for "Never" and "Yes" for all other responses, providing a longitudinal perspective on adolescent depression across Waves I and II. Variables with the suffix "w1" are from Wave I, while those with the suffix "w2" are from Wave II.

Usage

addhealth

Format

A data frame with 2061 rows and 18 variables:

GRADE

Respondent's grade level at Wave I.

SEX

Respondent's sex
levels: (1)Male, (2)Female.

S1w1, S1w2

I felt that I could not shake off the blues even with help from my family and friends.

S2w1, S2w2

I felt depressed.

S3w1, S3w2

I felt lonely.

S4w1, S4w2

I felt sad.

D1w1, D1w2

People were unfriendly to me.

D2w1, D2w2

I felt that people disliked me

F1w1, F1w2

I thought my life had been a failure.

F2w1, F2w2

I felt life was not worth living

Source

https://addhealth.cpc.unc.edu/data/#public-use

References

Collins, L.M., & Lanza, S.T. (2009). Latent Class and Latent Transition Analysis: With Applications in the Social, Behavioral, and Health Sciences.

J.R. Udry. The National Longitudinal Study of Adolescent Health (Add Health), Waves I & II, 1994-1996. Carolina Population Center, University of North Carolina at Chapel Hill, Chapel Hill, NC, 2003.

Examples

library(magrittr)
data <- addhealth[1:200,]
lta5 <- slca(
   DEP1(5) ~ S1w1 + S2w1 + S3w1 + S4w1 + D1w1 + D2w1 + F1w1 + F2w1,
   DEP2(5) ~ S1w2 + S2w2 + S3w2 + S4w2 + D1w2 + D2w2 + F1w2 + F2w2,
   DEP1 ~ DEP2
) %>% estimate(data, control = list(em.tol = 1e-6))
lta5inv <- slca(
   DEP1(5) ~ S1w1 + S2w1 + S3w1 + S4w1 + D1w1 + D2w1 + F1w1 + F2w1,
   DEP2(5) ~ S1w2 + S2w2 + S3w2 + S4w2 + D1w2 + D2w2 + F1w2 + F2w2,
   DEP1 ~ DEP2,
   constraints = c("DEP1", "DEP2")
) %>% estimate(data, control = list(em.tol = 1e-6))

compare(lta5inv, lta5, test = "chisq")
#> Analysis of Relative Model Fit
#> 
#> Model H0: lta5inv
#> Model H1: lta5
#>          Df  logLik    AIC    BIC    Gsq Res. Df Pr(>Chi)
#> lta5inv  64 -1402.8 2933.7 3144.8                        
#> lta5    104 -1381.0 2970.0 3313.0 43.696      40   0.3173
lta5inv %>% param()
#> PI :
#> (DEP1)
#>   class
#>          1       2       3       4       5
#>     0.1446  0.1511  0.4153  0.1990  0.0900
#> 
#> TAU :
#> (A)
#>      parent
#> child       1       2       3       4       5
#>     1  0.2984  0.0000  0.0265  0.0355  0.0497
#>     2  0.1406  0.5003  0.0000  0.0000  0.0000
#>     3  0.2853  0.0904  0.7802  0.5689  0.2147
#>     4  0.0000  0.0979  0.1553  0.2519  0.2114
#>     5  0.2758  0.3114  0.0380  0.1437  0.5242
#>            
#> parent DEP1
#> child  DEP2
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3       4       5
#>    1(V1)  0.0862  0.7413  0.0619  0.5575  0.6346
#>    2      0.9138  0.2587  0.9381  0.4425  0.3654
#>    1(V2)  0.1766  0.8673  0.0770  0.6829  0.7441
#>    2      0.8234  0.1327  0.9230  0.3171  0.2559
#>    1(V3)  0.1851  0.8160  0.1070  0.5424  0.7278
#>    2      0.8149  0.1840  0.8930  0.4576  0.2722
#>    1(V4)  0.1935  0.8681  0.1615  0.8473  0.9848
#>    2      0.8065  0.1319  0.8385  0.1527  0.0152
#>    1(V5)  0.7109  0.6507  0.1003  0.1317  0.6783
#>    2      0.2891  0.3493  0.8997  0.8683  0.3217
#>    1(V6)  0.7317  0.9556  0.0157  0.0000  0.9058
#>    2      0.2683  0.0444  0.9843  1.0000  0.0942
#>    1(V7)  0.1500  0.7948  0.0000  0.2177  0.2090
#>    2      0.8500  0.2052  1.0000  0.7823  0.7910
#>    1(V8)  0.1290  0.6281  0.0150  0.1076  0.0000
#>    2      0.8710  0.3719  0.9850  0.8924  1.0000
#> 
#>      V1   V2   V3   V4   V5   V6   V7   V8  
#> DEP1 S1w1 S2w1 S3w1 S4w1 D1w1 D2w1 F1w1 F2w1
#> DEP2 S1w2 S2w2 S3w2 S4w2 D1w2 D2w2 F1w2 F2w2