Estimate Parameters of an slca Object

Estimates the parameters of a model created using the slca function.

Usage

estimate(x, ...)

# S3 method for class 'slca'
estimate(x,
   data,
   method = c("em", "hybrid", "nlm"),
   fix2zero = NULL,
   control = slcaControl(), ...)

Arguments

x

an slca object defining the slca model to be estimated.

...

additional arguments passed to the estimation process.

data

a data.frame containing the observed categorical variables included in the model.

method

a character string specifying the estimation method for SLCM parameters. The default is "em", which uses the expectation-maximization (EM) algorithm. The alternative "nlm" employs the Newton-Raphson algorithm via the nlm function, while "hybrid" combines both approaches, starting with EM and finishing with nlm for refined estimates.

fix2zero

a vector specifying parameters to be constrained to zero. See the 'Details' section for further information.

control

a list of control parameters for the estimation procedure. Modify default values using slcaControl().

Value

An object of class slcafit containing the following components:

model

a list describing of the model structure.

method

the estimation method used.

arg

a brief description of the model used during estimation.

mf

the data.frame used for estimation.

par

the log of the estimated paramters.

logit

the log-odds of the estimated parameters.

score

the score function for the estimated parameters.

posterior

a list of posterior probablities for each latent class variable.

convergence

a logical indicator of whether convergence was achieved.

loglikelihood

the loglikelihood value of the estimated model.

control

the control settings used during the estimation process.

The returned object can be further processed using the param function to extract the estimated parameters or their standard errors. The regress function allows for logistic regression analysis using a three-step approach to evaluate the effects of external variables on latent class variables. Additionally, several other methods are available, including predict.slcafit, reorder.slcafit, gof, and others.

Details

The fix2zero argument allows you to constrain specific parameters to zero. Each parameter is associated with a unique index, which can be identified using the param function with the argument index = TRUE. To apply constraints, provide the relevant parameter indices in the fix2zero arguments with vector.

See also

Examples

m <- slca(lc[3] ~ y1 + y2 + y3 + y4)
pi <- rep(1 / 3, 3)
rho <- c(.9, .1, .9, .1, .9, .1, .9, .1, # class 1
         .9, .1, .9, .1, .1, .9, .1, .9, # class 2
         .1, .9, .1, .9, .1, .9, .1, .9) # class 3
dt <- simulate(m, 200, parm = c(pi, rho))
estimate(m, dt$response)
#> Structural Latent Variable Model
#> 
#> Latent variables (Root*):           
#>  Label: lc*
#> nclass: 3  
#> 
#> Measurement model:                            
#>  lc -> { y1, y2, y3, y4 }  a
#> 
#> LogLik: -418.6869

# Several estimation methods
estimate(m, dt$response, method = "em",
         control = slcaControl(verbose = TRUE)) # default
#> EM iteration begin.
#> 1 iterations  logLik: -554.04  diff: inf      2 iterations  logLik: -525.87  diff: 28.16667      3 iterations  logLik: -525.85  diff: 0.02429      4 iterations  logLik: -525.75  diff: 0.10077      5 iterations  logLik: -525.25  diff: 0.49661      6 iterations  logLik: -522.86  diff: 2.39066      7 iterations  logLik: -512.76  diff: 10.10172      8 iterations  logLik: -486.04  diff: 26.71932      9 iterations  logLik: -460.94  diff: 25.09520      10 iterations  logLik: -452.69  diff: 8.25678      11 iterations  logLik: -447.72  diff: 4.96113      12 iterations  logLik: -440.28  diff: 7.44853      13 iterations  logLik: -430.01  diff: 10.26439      14 iterations  logLik: -422.50  diff: 7.50993      15 iterations  logLik: -420.23  diff: 2.27636      16 iterations  logLik: -419.87  diff: 0.35361      17 iterations  logLik: -419.79  diff: 0.07844      18 iterations  logLik: -419.75  diff: 0.04678      19 iterations  logLik: -419.71  diff: 0.04160      20 iterations  logLik: -419.67  diff: 0.03996      21 iterations  logLik: -419.63  diff: 0.03889      22 iterations  logLik: -419.59  diff: 0.03795      23 iterations  logLik: -419.55  diff: 0.03704      24 iterations  logLik: -419.52  diff: 0.03617      25 iterations  logLik: -419.48  diff: 0.03531      26 iterations  logLik: -419.45  diff: 0.03447      27 iterations  logLik: -419.41  diff: 0.03363      28 iterations  logLik: -419.38  diff: 0.03281      29 iterations  logLik: -419.35  diff: 0.03197      30 iterations  logLik: -419.32  diff: 0.03114      31 iterations  logLik: -419.29  diff: 0.03030      32 iterations  logLik: -419.26  diff: 0.02944      33 iterations  logLik: -419.23  diff: 0.02857      34 iterations  logLik: -419.20  diff: 0.02769      35 iterations  logLik: -419.17  diff: 0.02679      36 iterations  logLik: -419.15  diff: 0.02587      37 iterations  logLik: -419.12  diff: 0.02495      38 iterations  logLik: -419.10  diff: 0.02401      39 iterations  logLik: -419.08  diff: 0.02306      40 iterations  logLik: -419.05  diff: 0.02210      41 iterations  logLik: -419.03  diff: 0.02115      42 iterations  logLik: -419.01  diff: 0.02019      43 iterations  logLik: -418.99  diff: 0.01924      44 iterations  logLik: -418.97  diff: 0.01829      45 iterations  logLik: -418.96  diff: 0.01736      46 iterations  logLik: -418.94  diff: 0.01645      47 iterations  logLik: -418.92  diff: 0.01555      48 iterations  logLik: -418.91  diff: 0.01468      49 iterations  logLik: -418.90  diff: 0.01384      50 iterations  logLik: -418.88  diff: 0.01302      51 iterations  logLik: -418.87  diff: 0.01223      52 iterations  logLik: -418.86  diff: 0.01148      53 iterations  logLik: -418.85  diff: 0.01076      54 iterations  logLik: -418.84  diff: 0.01007      55 iterations  logLik: -418.83  diff: 0.00942      56 iterations  logLik: -418.82  diff: 0.00880      57 iterations  logLik: -418.81  diff: 0.00822      58 iterations  logLik: -418.80  diff: 0.00767      59 iterations  logLik: -418.80  diff: 0.00715      60 iterations  logLik: -418.79  diff: 0.00667      61 iterations  logLik: -418.78  diff: 0.00622      62 iterations  logLik: -418.78  diff: 0.00579      63 iterations  logLik: -418.77  diff: 0.00540      64 iterations  logLik: -418.77  diff: 0.00503      65 iterations  logLik: -418.76  diff: 0.00469      66 iterations  logLik: -418.76  diff: 0.00437      67 iterations  logLik: -418.76  diff: 0.00408      68 iterations  logLik: -418.75  diff: 0.00380      69 iterations  logLik: -418.75  diff: 0.00355      70 iterations  logLik: -418.74  diff: 0.00331      71 iterations  logLik: -418.74  diff: 0.00309      72 iterations  logLik: -418.74  diff: 0.00289      73 iterations  logLik: -418.74  diff: 0.00270      74 iterations  logLik: -418.73  diff: 0.00253      75 iterations  logLik: -418.73  diff: 0.00237      76 iterations  logLik: -418.73  diff: 0.00222      77 iterations  logLik: -418.73  diff: 0.00208      78 iterations  logLik: -418.72  diff: 0.00196      79 iterations  logLik: -418.72  diff: 0.00184      80 iterations  logLik: -418.72  diff: 0.00173      81 iterations  logLik: -418.72  diff: 0.00162      82 iterations  logLik: -418.72  diff: 0.00153      83 iterations  logLik: -418.72  diff: 0.00144      84 iterations  logLik: -418.72  diff: 0.00136      85 iterations  logLik: -418.71  diff: 0.00128      86 iterations  logLik: -418.71  diff: 0.00121      87 iterations  logLik: -418.71  diff: 0.00115      88 iterations  logLik: -418.71  diff: 0.00108      89 iterations  logLik: -418.71  diff: 0.00103      90 iterations  logLik: -418.71  diff: 0.00097      91 iterations  logLik: -418.71  diff: 0.00092      92 iterations  logLik: -418.71  diff: 0.00088      93 iterations  logLik: -418.71  diff: 0.00083      94 iterations  logLik: -418.70  diff: 0.00079      95 iterations  logLik: -418.70  diff: 0.00075      96 iterations  logLik: -418.70  diff: 0.00072      97 iterations  logLik: -418.70  diff: 0.00068      98 iterations  logLik: -418.70  diff: 0.00065      99 iterations  logLik: -418.70  diff: 0.00062      100 iterations  logLik: -418.70  diff: 0.00059      101 iterations  logLik: -418.70  diff: 0.00056      102 iterations  logLik: -418.70  diff: 0.00054      103 iterations  logLik: -418.70  diff: 0.00051      104 iterations  logLik: -418.70  diff: 0.00049      105 iterations  logLik: -418.70  diff: 0.00047      106 iterations  logLik: -418.70  diff: 0.00045      107 iterations  logLik: -418.70  diff: 0.00043      108 iterations  logLik: -418.70  diff: 0.00041      109 iterations  logLik: -418.70  diff: 0.00039      110 iterations  logLik: -418.70  diff: 0.00038      111 iterations  logLik: -418.70  diff: 0.00036      112 iterations  logLik: -418.70  diff: 0.00035      113 iterations  logLik: -418.70  diff: 0.00033      114 iterations  logLik: -418.69  diff: 0.00032      115 iterations  logLik: -418.69  diff: 0.00031      116 iterations  logLik: -418.69  diff: 0.00029      117 iterations  logLik: -418.69  diff: 0.00028      118 iterations  logLik: -418.69  diff: 0.00027      119 iterations  logLik: -418.69  diff: 0.00026      120 iterations  logLik: -418.69  diff: 0.00025      121 iterations  logLik: -418.69  diff: 0.00024      122 iterations  logLik: -418.69  diff: 0.00023      123 iterations  logLik: -418.69  diff: 0.00022      124 iterations  logLik: -418.69  diff: 0.00021      125 iterations  logLik: -418.69  diff: 0.00020      126 iterations  logLik: -418.69  diff: 0.00020      127 iterations  logLik: -418.69  diff: 0.00019      128 iterations  logLik: -418.69  diff: 0.00018      129 iterations  logLik: -418.69  diff: 0.00017      130 iterations  logLik: -418.69  diff: 0.00017      131 iterations  logLik: -418.69  diff: 0.00016      132 iterations  logLik: -418.69  diff: 0.00016      133 iterations  logLik: -418.69  diff: 0.00015      134 iterations  logLik: -418.69  diff: 0.00014      135 iterations  logLik: -418.69  diff: 0.00014      136 iterations  logLik: -418.69  diff: 0.00013      137 iterations  logLik: -418.69  diff: 0.00013      138 iterations  logLik: -418.69  diff: 0.00012      139 iterations  logLik: -418.69  diff: 0.00012      140 iterations  logLik: -418.69  diff: 0.00011      141 iterations  logLik: -418.69  diff: 0.00011      142 iterations  logLik: -418.69  diff: 0.00011      143 iterations  logLik: -418.69  diff: 0.00010      144 iterations  logLik: -418.69  diff: 0.00010      145 iterations  logLik: -418.69  diff: 0.00009      146 iterations  logLik: -418.69  diff: 0.00009      147 iterations  logLik: -418.69  diff: 0.00009      148 iterations  logLik: -418.69  diff: 0.00008      149 iterations  logLik: -418.69  diff: 0.00008      150 iterations  logLik: -418.69  diff: 0.00008      151 iterations  logLik: -418.69  diff: 0.00008      152 iterations  logLik: -418.69  diff: 0.00007      153 iterations  logLik: -418.69  diff: 0.00007      154 iterations  logLik: -418.69  diff: 0.00007      155 iterations  logLik: -418.69  diff: 0.00007      156 iterations  logLik: -418.69  diff: 0.00006      157 iterations  logLik: -418.69  diff: 0.00006      158 iterations  logLik: -418.69  diff: 0.00006      159 iterations  logLik: -418.69  diff: 0.00006      160 iterations  logLik: -418.69  diff: 0.00005      161 iterations  logLik: -418.69  diff: 0.00005      162 iterations  logLik: -418.69  diff: 0.00005      163 iterations  logLik: -418.69  diff: 0.00005      164 iterations  logLik: -418.69  diff: 0.00005      165 iterations  logLik: -418.69  diff: 0.00005      166 iterations  logLik: -418.69  diff: 0.00004      167 iterations  logLik: -418.69  diff: 0.00004      168 iterations  logLik: -418.69  diff: 0.00004      169 iterations  logLik: -418.69  diff: 0.00004      170 iterations  logLik: -418.69  diff: 0.00004      171 iterations  logLik: -418.69  diff: 0.00004      172 iterations  logLik: -418.69  diff: 0.00004      173 iterations  logLik: -418.69  diff: 0.00003      174 iterations  logLik: -418.69  diff: 0.00003      175 iterations  logLik: -418.69  diff: 0.00003      176 iterations  logLik: -418.69  diff: 0.00003      177 iterations  logLik: -418.69  diff: 0.00003      178 iterations  logLik: -418.69  diff: 0.00003      179 iterations  logLik: -418.69  diff: 0.00003      180 iterations  logLik: -418.69  diff: 0.00003      181 iterations  logLik: -418.69  diff: 0.00003      182 iterations  logLik: -418.69  diff: 0.00002      183 iterations  logLik: -418.69  diff: 0.00002      184 iterations  logLik: -418.69  diff: 0.00002      185 iterations  logLik: -418.69  diff: 0.00002      186 iterations  logLik: -418.69  diff: 0.00002      187 iterations  logLik: -418.69  diff: 0.00002      188 iterations  logLik: -418.69  diff: 0.00002      189 iterations  logLik: -418.69  diff: 0.00002      190 iterations  logLik: -418.69  diff: 0.00002      191 iterations  logLik: -418.69  diff: 0.00002      192 iterations  logLik: -418.69  diff: 0.00002      193 iterations  logLik: -418.69  diff: 0.00002      194 iterations  logLik: -418.69  diff: 0.00002      195 iterations  logLik: -418.69  diff: 0.00002      196 iterations  logLik: -418.69  diff: 0.00001      197 iterations  logLik: -418.69  diff: 0.00001      198 iterations  logLik: -418.69  diff: 0.00001      199 iterations  logLik: -418.69  diff: 0.00001      200 iterations  logLik: -418.69  diff: 0.00001      201 iterations  logLik: -418.69  diff: 0.00001      202 iterations  logLik: -418.69  diff: 0.00001      203 iterations  logLik: -418.69  diff: 0.00001      204 iterations  logLik: -418.69  diff: 0.00001      205 iterations  logLik: -418.69  diff: 0.00001      206 iterations  logLik: -418.69  diff: 0.00001      207 iterations  logLik: -418.69  diff: 9.8e-06      208 iterations  logLik: -418.69  diff: 9.5e-06      209 iterations  logLik: -418.69  diff: 9.1e-06      210 iterations  logLik: -418.69  diff: 8.8e-06      211 iterations  logLik: -418.69  diff: 8.5e-06      212 iterations  logLik: -418.69  diff: 8.2e-06      213 iterations  logLik: -418.69  diff: 7.9e-06      214 iterations  logLik: -418.69  diff: 7.6e-06      215 iterations  logLik: -418.69  diff: 7.3e-06      216 iterations  logLik: -418.69  diff: 7.1e-06      217 iterations  logLik: -418.69  diff: 6.8e-06      218 iterations  logLik: -418.69  diff: 6.6e-06      219 iterations  logLik: -418.69  diff: 6.4e-06      220 iterations  logLik: -418.69  diff: 6.1e-06      221 iterations  logLik: -418.69  diff: 5.9e-06      222 iterations  logLik: -418.69  diff: 5.7e-06      223 iterations  logLik: -418.69  diff: 5.5e-06      224 iterations  logLik: -418.69  diff: 5.3e-06      225 iterations  logLik: -418.69  diff: 5.1e-06      226 iterations  logLik: -418.69  diff: 4.9e-06      227 iterations  logLik: -418.69  diff: 4.8e-06      228 iterations  logLik: -418.69  diff: 4.6e-06      229 iterations  logLik: -418.69  diff: 4.4e-06      230 iterations  logLik: -418.69  diff: 4.3e-06      231 iterations  logLik: -418.69  diff: 4.1e-06      232 iterations  logLik: -418.69  diff: 4.0e-06      233 iterations  logLik: -418.69  diff: 3.8e-06      234 iterations  logLik: -418.69  diff: 3.7e-06      235 iterations  logLik: -418.69  diff: 3.6e-06      236 iterations  logLik: -418.69  diff: 3.4e-06      237 iterations  logLik: -418.69  diff: 3.3e-06      238 iterations  logLik: -418.69  diff: 3.2e-06      239 iterations  logLik: -418.69  diff: 3.1e-06      240 iterations  logLik: -418.69  diff: 3.0e-06      241 iterations  logLik: -418.69  diff: 2.9e-06      242 iterations  logLik: -418.69  diff: 2.8e-06      243 iterations  logLik: -418.69  diff: 2.7e-06      244 iterations  logLik: -418.69  diff: 2.6e-06      245 iterations  logLik: -418.69  diff: 2.5e-06      246 iterations  logLik: -418.69  diff: 2.4e-06      247 iterations  logLik: -418.69  diff: 2.3e-06      248 iterations  logLik: -418.69  diff: 2.2e-06      249 iterations  logLik: -418.69  diff: 2.1e-06      250 iterations  logLik: -418.69  diff: 2.1e-06      251 iterations  logLik: -418.69  diff: 2.0e-06      252 iterations  logLik: -418.69  diff: 1.9e-06      253 iterations  logLik: -418.69  diff: 1.9e-06      254 iterations  logLik: -418.69  diff: 1.8e-06      255 iterations  logLik: -418.69  diff: 1.7e-06      256 iterations  logLik: -418.69  diff: 1.7e-06      257 iterations  logLik: -418.69  diff: 1.6e-06      258 iterations  logLik: -418.69  diff: 1.6e-06      259 iterations  logLik: -418.69  diff: 1.5e-06      260 iterations  logLik: -418.69  diff: 1.4e-06      261 iterations  logLik: -418.69  diff: 1.4e-06      262 iterations  logLik: -418.69  diff: 1.3e-06      263 iterations  logLik: -418.69  diff: 1.3e-06      264 iterations  logLik: -418.69  diff: 1.3e-06      265 iterations  logLik: -418.69  diff: 1.2e-06      266 iterations  logLik: -418.69  diff: 1.2e-06      267 iterations  logLik: -418.69  diff: 1.1e-06      268 iterations  logLik: -418.69  diff: 1.1e-06      269 iterations  logLik: -418.69  diff: 1.0e-06      270 iterations  logLik: -418.69  diff: 1.0e-06      271 iterations  logLik: -418.69  diff: 9.7e-07      272 iterations  logLik: -418.69  diff: 9.4e-07      273 iterations  logLik: -418.69  diff: 9.0e-07      274 iterations  logLik: -418.69  diff: 8.7e-07      275 iterations  logLik: -418.69  diff: 8.4e-07      276 iterations  logLik: -418.69  diff: 8.1e-07      277 iterations  logLik: -418.69  diff: 7.8e-07      278 iterations  logLik: -418.69  diff: 7.5e-07      279 iterations  logLik: -418.69  diff: 7.3e-07      280 iterations  logLik: -418.69  diff: 7.0e-07      281 iterations  logLik: -418.69  diff: 6.8e-07      282 iterations  logLik: -418.69  diff: 6.5e-07      283 iterations  logLik: -418.69  diff: 6.3e-07      284 iterations  logLik: -418.69  diff: 6.1e-07      285 iterations  logLik: -418.69  diff: 5.9e-07      286 iterations  logLik: -418.69  diff: 5.6e-07      287 iterations  logLik: -418.69  diff: 5.4e-07      288 iterations  logLik: -418.69  diff: 5.3e-07      289 iterations  logLik: -418.69  diff: 5.1e-07      290 iterations  logLik: -418.69  diff: 4.9e-07      291 iterations  logLik: -418.69  diff: 4.7e-07      292 iterations  logLik: -418.69  diff: 4.5e-07      293 iterations  logLik: -418.69  diff: 4.4e-07      294 iterations  logLik: -418.69  diff: 4.2e-07      295 iterations  logLik: -418.69  diff: 4.1e-07      296 iterations  logLik: -418.69  diff: 3.9e-07      297 iterations  logLik: -418.69  diff: 3.8e-07      298 iterations  logLik: -418.69  diff: 3.7e-07      299 iterations  logLik: -418.69  diff: 3.5e-07      300 iterations  logLik: -418.69  diff: 3.4e-07      301 iterations  logLik: -418.69  diff: 3.3e-07      302 iterations  logLik: -418.69  diff: 3.2e-07      303 iterations  logLik: -418.69  diff: 3.1e-07      304 iterations  logLik: -418.69  diff: 2.9e-07      305 iterations  logLik: -418.69  diff: 2.8e-07      306 iterations  logLik: -418.69  diff: 2.7e-07      307 iterations  logLik: -418.69  diff: 2.6e-07      308 iterations  logLik: -418.69  diff: 2.6e-07      309 iterations  logLik: -418.69  diff: 2.5e-07      310 iterations  logLik: -418.69  diff: 2.4e-07      311 iterations  logLik: -418.69  diff: 2.3e-07      312 iterations  logLik: -418.69  diff: 2.2e-07      313 iterations  logLik: -418.69  diff: 2.1e-07      314 iterations  logLik: -418.69  diff: 2.1e-07      315 iterations  logLik: -418.69  diff: 2.0e-07      316 iterations  logLik: -418.69  diff: 1.9e-07      317 iterations  logLik: -418.69  diff: 1.8e-07      318 iterations  logLik: -418.69  diff: 1.8e-07      319 iterations  logLik: -418.69  diff: 1.7e-07      320 iterations  logLik: -418.69  diff: 1.7e-07      321 iterations  logLik: -418.69  diff: 1.6e-07      322 iterations  logLik: -418.69  diff: 1.5e-07      323 iterations  logLik: -418.69  diff: 1.5e-07      324 iterations  logLik: -418.69  diff: 1.4e-07      325 iterations  logLik: -418.69  diff: 1.4e-07      326 iterations  logLik: -418.69  diff: 1.3e-07      327 iterations  logLik: -418.69  diff: 1.3e-07      328 iterations  logLik: -418.69  diff: 1.2e-07      329 iterations  logLik: -418.69  diff: 1.2e-07      330 iterations  logLik: -418.69  diff: 1.2e-07      331 iterations  logLik: -418.69  diff: 1.1e-07      332 iterations  logLik: -418.69  diff: 1.1e-07      333 iterations  logLik: -418.69  diff: 1.0e-07      334 iterations  logLik: -418.69  diff: 1.0e-07      335 iterations  logLik: -418.69  diff: 9.6e-08      336 iterations  logLik: -418.69  diff: 9.3e-08      337 iterations  logLik: -418.69  diff: 9.0e-08      338 iterations  logLik: -418.69  diff: 8.6e-08      339 iterations  logLik: -418.69  diff: 8.3e-08      340 iterations  logLik: -418.69  diff: 8.0e-08      341 iterations  logLik: -418.69  diff: 7.8e-08      342 iterations  logLik: -418.69  diff: 7.5e-08      343 iterations  logLik: -418.69  diff: 7.2e-08      344 iterations  logLik: -418.69  diff: 7.0e-08      345 iterations  logLik: -418.69  diff: 6.7e-08      346 iterations  logLik: -418.69  diff: 6.5e-08      347 iterations  logLik: -418.69  diff: 6.2e-08      348 iterations  logLik: -418.69  diff: 6.0e-08      349 iterations  logLik: -418.69  diff: 5.8e-08      350 iterations  logLik: -418.69  diff: 5.6e-08      351 iterations  logLik: -418.69  diff: 5.4e-08      352 iterations  logLik: -418.69  diff: 5.2e-08      353 iterations  logLik: -418.69  diff: 5.0e-08      354 iterations  logLik: -418.69  diff: 4.8e-08      355 iterations  logLik: -418.69  diff: 4.7e-08      356 iterations  logLik: -418.69  diff: 4.5e-08      357 iterations  logLik: -418.69  diff: 4.4e-08      358 iterations  logLik: -418.69  diff: 4.2e-08      359 iterations  logLik: -418.69  diff: 4.0e-08      360 iterations  logLik: -418.69  diff: 3.9e-08      361 iterations  logLik: -418.69  diff: 3.8e-08      362 iterations  logLik: -418.69  diff: 3.6e-08      363 iterations  logLik: -418.69  diff: 3.5e-08      364 iterations  logLik: -418.69  diff: 3.4e-08      365 iterations  logLik: -418.69  diff: 3.3e-08      366 iterations  logLik: -418.69  diff: 3.1e-08      367 iterations  logLik: -418.69  diff: 3.0e-08      368 iterations  logLik: -418.69  diff: 2.9e-08      369 iterations  logLik: -418.69  diff: 2.8e-08      370 iterations  logLik: -418.69  diff: 2.7e-08      371 iterations  logLik: -418.69  diff: 2.6e-08      372 iterations  logLik: -418.69  diff: 2.5e-08      373 iterations  logLik: -418.69  diff: 2.4e-08      374 iterations  logLik: -418.69  diff: 2.4e-08      375 iterations  logLik: -418.69  diff: 2.3e-08      376 iterations  logLik: -418.69  diff: 2.2e-08      377 iterations  logLik: -418.69  diff: 2.1e-08      378 iterations  logLik: -418.69  diff: 2.0e-08      379 iterations  logLik: -418.69  diff: 2.0e-08      380 iterations  logLik: -418.69  diff: 1.9e-08      381 iterations  logLik: -418.69  diff: 1.8e-08      382 iterations  logLik: -418.69  diff: 1.8e-08      383 iterations  logLik: -418.69  diff: 1.7e-08      384 iterations  logLik: -418.69  diff: 1.6e-08      385 iterations  logLik: -418.69  diff: 1.6e-08      386 iterations  logLik: -418.69  diff: 1.5e-08      387 iterations  logLik: -418.69  diff: 1.5e-08      388 iterations  logLik: -418.69  diff: 1.4e-08      389 iterations  logLik: -418.69  diff: 1.4e-08      390 iterations  logLik: -418.69  diff: 1.3e-08      391 iterations  logLik: -418.69  diff: 1.3e-08      392 iterations  logLik: -418.69  diff: 1.2e-08      393 iterations  logLik: -418.69  diff: 1.2e-08      394 iterations  logLik: -418.69  diff: 1.1e-08      395 iterations  logLik: -418.69  diff: 1.1e-08      396 iterations  logLik: -418.69  diff: 1.1e-08      397 iterations  logLik: -418.69  diff: 1.0e-08      398 iterations  logLik: -418.69  diff: 9.9e-09      
#> .. done.
#> Structural Latent Variable Model
#> 
#> Latent variables (Root*):           
#>  Label: lc*
#> nclass: 3  
#> 
#> Measurement model:                            
#>  lc -> { y1, y2, y3, y4 }  a
#> 
#> LogLik: -418.6869
estimate(m, dt$response, method = "nlm",
         control = slcaControl(verbose = TRUE))
#> nlm iteration begin.
#> .. done.
#> Structural Latent Variable Model
#> 
#> Latent variables (Root*):           
#>  Label: lc*
#> nclass: 3  
#> 
#> Measurement model:                            
#>  lc -> { y1, y2, y3, y4 }  a
#> 
#> LogLik: -419.449
estimate(m, dt$response, method = "hybrid",
         control = slcaControl(verbose = TRUE))
#> EM iteration begin.
#> 1 iterations  logLik: -553.95  diff: inf      2 iterations  logLik: -525.81  diff: 28.13425      3 iterations  logLik: -525.57  diff: 0.24010      4 iterations  logLik: -524.40  diff: 1.17046      5 iterations  logLik: -519.01  diff: 5.39735      6 iterations  logLik: -500.04  diff: 18.96421      7 iterations  logLik: -469.39  diff: 30.64874      8 iterations  logLik: -453.52  diff: 15.87814      9 iterations  logLik: -442.87  diff: 10.64149      10 iterations  logLik: -429.74  diff: 13.13561      11 iterations  logLik: -421.75  diff: 7.98906      12 iterations  logLik: -419.99  diff: 1.76209      13 iterations  logLik: -419.80  diff: 0.18196      14 iterations  logLik: -419.78  diff: 0.02572      15 iterations  logLik: -419.76  diff: 0.01456      16 iterations  logLik: -419.75  diff: 0.01309      17 iterations  logLik: -419.74  diff: 0.01227      18 iterations  logLik: -419.73  diff: 0.01154      19 iterations  logLik: -419.72  diff: 0.01088      20 iterations  logLik: -419.71  diff: 0.01028      21 iterations  logLik: -419.70  diff: 0.00973      22 iterations  logLik: -419.69  diff: 0.00923      23 iterations  logLik: -419.68  diff: 0.00877      24 iterations  logLik: -419.67  diff: 0.00836      25 iterations  logLik: -419.66  diff: 0.00800      26 iterations  logLik: -419.65  diff: 0.00767      27 iterations  logLik: -419.65  diff: 0.00737      28 iterations  logLik: -419.64  diff: 0.00711      29 iterations  logLik: -419.63  diff: 0.00689      30 iterations  logLik: -419.63  diff: 0.00669      31 iterations  logLik: -419.62  diff: 0.00652      32 iterations  logLik: -419.61  diff: 0.00638      33 iterations  logLik: -419.61  diff: 0.00626      34 iterations  logLik: -419.60  diff: 0.00616      35 iterations  logLik: -419.60  diff: 0.00609      36 iterations  logLik: -419.59  diff: 0.00605      37 iterations  logLik: -419.58  diff: 0.00602      38 iterations  logLik: -419.58  diff: 0.00601      39 iterations  logLik: -419.57  diff: 0.00602      40 iterations  logLik: -419.57  diff: 0.00605      41 iterations  logLik: -419.56  diff: 0.00609      42 iterations  logLik: -419.55  diff: 0.00615      43 iterations  logLik: -419.55  diff: 0.00623      44 iterations  logLik: -419.54  diff: 0.00631      45 iterations  logLik: -419.53  diff: 0.00642      46 iterations  logLik: -419.53  diff: 0.00653      47 iterations  logLik: -419.52  diff: 0.00666      48 iterations  logLik: -419.51  diff: 0.00680      49 iterations  logLik: -419.51  diff: 0.00694      50 iterations  logLik: -419.50  diff: 0.00710      51 iterations  logLik: -419.49  diff: 0.00726      52 iterations  logLik: -419.49  diff: 0.00743      53 iterations  logLik: -419.48  diff: 0.00761      54 iterations  logLik: -419.47  diff: 0.00779      55 iterations  logLik: -419.46  diff: 0.00797      56 iterations  logLik: -419.45  diff: 0.00816      57 iterations  logLik: -419.45  diff: 0.00834      58 iterations  logLik: -419.44  diff: 0.00853      59 iterations  logLik: -419.43  diff: 0.00871      60 iterations  logLik: -419.42  diff: 0.00889      61 iterations  logLik: -419.41  diff: 0.00906      62 iterations  logLik: -419.40  diff: 0.00923      63 iterations  logLik: -419.39  diff: 0.00939      64 iterations  logLik: -419.38  diff: 0.00954      65 iterations  logLik: -419.37  diff: 0.00968      66 iterations  logLik: -419.36  diff: 0.00981      67 iterations  logLik: -419.35  diff: 0.00993      68 iterations  logLik: -419.34  diff: 0.01003      69 iterations  logLik: -419.33  diff: 0.01012      70 iterations  logLik: -419.32  diff: 0.01019      71 iterations  logLik: -419.31  diff: 0.01025      72 iterations  logLik: -419.30  diff: 0.01029      73 iterations  logLik: -419.29  diff: 0.01031      74 iterations  logLik: -419.28  diff: 0.01032      75 iterations  logLik: -419.27  diff: 0.01030      76 iterations  logLik: -419.26  diff: 0.01027      77 iterations  logLik: -419.25  diff: 0.01022      78 iterations  logLik: -419.24  diff: 0.01016      79 iterations  logLik: -419.23  diff: 0.01008      80 iterations  logLik: -419.22  diff: 0.00998      81 iterations  logLik: -419.21  diff: 0.00987      82 iterations  logLik: -419.20  diff: 0.00975      83 iterations  logLik: -419.19  diff: 0.00961      84 iterations  logLik: -419.18  diff: 0.00946      85 iterations  logLik: -419.17  diff: 0.00930      86 iterations  logLik: -419.16  diff: 0.00913      87 iterations  logLik: -419.15  diff: 0.00895      88 iterations  logLik: -419.15  diff: 0.00877      89 iterations  logLik: -419.14  diff: 0.00857      90 iterations  logLik: -419.13  diff: 0.00838      91 iterations  logLik: -419.12  diff: 0.00817      92 iterations  logLik: -419.11  diff: 0.00797      93 iterations  logLik: -419.10  diff: 0.00776      94 iterations  logLik: -419.10  diff: 0.00755      95 iterations  logLik: -419.09  diff: 0.00734      96 iterations  logLik: -419.08  diff: 0.00713      97 iterations  logLik: -419.08  diff: 0.00692      98 iterations  logLik: -419.07  diff: 0.00672      99 iterations  logLik: -419.06  diff: 0.00651      100 iterations  logLik: -419.06  diff: 0.00631      101 iterations  logLik: -419.05  diff: 0.00611      102 iterations  logLik: -419.04  diff: 0.00591      103 iterations  logLik: -419.04  diff: 0.00572      104 iterations  logLik: -419.03  diff: 0.00553      105 iterations  logLik: -419.03  diff: 0.00534      106 iterations  logLik: -419.02  diff: 0.00516      107 iterations  logLik: -419.02  diff: 0.00498      108 iterations  logLik: -419.01  diff: 0.00481      109 iterations  logLik: -419.01  diff: 0.00465      110 iterations  logLik: -419.00  diff: 0.00448      111 iterations  logLik: -419.00  diff: 0.00433      112 iterations  logLik: -418.99  diff: 0.00417      113 iterations  logLik: -418.99  diff: 0.00403      114 iterations  logLik: -418.99  diff: 0.00388      115 iterations  logLik: -418.98  diff: 0.00375      116 iterations  logLik: -418.98  diff: 0.00361      117 iterations  logLik: -418.98  diff: 0.00348      118 iterations  logLik: -418.97  diff: 0.00336      119 iterations  logLik: -418.97  diff: 0.00324      120 iterations  logLik: -418.97  diff: 0.00312      121 iterations  logLik: -418.96  diff: 0.00301      122 iterations  logLik: -418.96  diff: 0.00290      123 iterations  logLik: -418.96  diff: 0.00280      124 iterations  logLik: -418.95  diff: 0.00270      125 iterations  logLik: -418.95  diff: 0.00260      126 iterations  logLik: -418.95  diff: 0.00251      127 iterations  logLik: -418.95  diff: 0.00242      128 iterations  logLik: -418.95  diff: 0.00233      129 iterations  logLik: -418.94  diff: 0.00225      130 iterations  logLik: -418.94  diff: 0.00217      131 iterations  logLik: -418.94  diff: 0.00210      132 iterations  logLik: -418.94  diff: 0.00202      133 iterations  logLik: -418.93  diff: 0.00195      134 iterations  logLik: -418.93  diff: 0.00188      135 iterations  logLik: -418.93  diff: 0.00182      136 iterations  logLik: -418.93  diff: 0.00175      137 iterations  logLik: -418.93  diff: 0.00169      138 iterations  logLik: -418.93  diff: 0.00163      139 iterations  logLik: -418.92  diff: 0.00158      140 iterations  logLik: -418.92  diff: 0.00152      141 iterations  logLik: -418.92  diff: 0.00147      142 iterations  logLik: -418.92  diff: 0.00142      143 iterations  logLik: -418.92  diff: 0.00137      144 iterations  logLik: -418.92  diff: 0.00133      145 iterations  logLik: -418.92  diff: 0.00128      146 iterations  logLik: -418.91  diff: 0.00124      147 iterations  logLik: -418.91  diff: 0.00120      148 iterations  logLik: -418.91  diff: 0.00116      149 iterations  logLik: -418.91  diff: 0.00112      150 iterations  logLik: -418.91  diff: 0.00108      151 iterations  logLik: -418.91  diff: 0.00105      152 iterations  logLik: -418.91  diff: 0.00101      153 iterations  logLik: -418.91  diff: 0.00098      154 iterations  logLik: -418.91  diff: 0.00095      155 iterations  logLik: -418.91  diff: 0.00091      156 iterations  logLik: -418.90  diff: 0.00088      157 iterations  logLik: -418.90  diff: 0.00086      158 iterations  logLik: -418.90  diff: 0.00083      159 iterations  logLik: -418.90  diff: 0.00080      160 iterations  logLik: -418.90  diff: 0.00078      161 iterations  logLik: -418.90  diff: 0.00075      162 iterations  logLik: -418.90  diff: 0.00073      163 iterations  logLik: -418.90  diff: 0.00070      164 iterations  logLik: -418.90  diff: 0.00068      165 iterations  logLik: -418.90  diff: 0.00066      166 iterations  logLik: -418.90  diff: 0.00064      167 iterations  logLik: -418.90  diff: 0.00062      168 iterations  logLik: -418.90  diff: 0.00060      169 iterations  logLik: -418.89  diff: 0.00058      170 iterations  logLik: -418.89  diff: 0.00056      171 iterations  logLik: -418.89  diff: 0.00054      172 iterations  logLik: -418.89  diff: 0.00053      173 iterations  logLik: -418.89  diff: 0.00051      174 iterations  logLik: -418.89  diff: 0.00049      175 iterations  logLik: -418.89  diff: 0.00048      176 iterations  logLik: -418.89  diff: 0.00046      177 iterations  logLik: -418.89  diff: 0.00045      178 iterations  logLik: -418.89  diff: 0.00044      179 iterations  logLik: -418.89  diff: 0.00042      180 iterations  logLik: -418.89  diff: 0.00041      181 iterations  logLik: -418.89  diff: 0.00040      182 iterations  logLik: -418.89  diff: 0.00038      183 iterations  logLik: -418.89  diff: 0.00037      184 iterations  logLik: -418.89  diff: 0.00036      185 iterations  logLik: -418.89  diff: 0.00035      186 iterations  logLik: -418.89  diff: 0.00034      187 iterations  logLik: -418.89  diff: 0.00033      188 iterations  logLik: -418.89  diff: 0.00032      189 iterations  logLik: -418.89  diff: 0.00031      190 iterations  logLik: -418.89  diff: 0.00030      191 iterations  logLik: -418.89  diff: 0.00029      192 iterations  logLik: -418.89  diff: 0.00028      193 iterations  logLik: -418.89  diff: 0.00027      194 iterations  logLik: -418.89  diff: 0.00027      195 iterations  logLik: -418.88  diff: 0.00026      196 iterations  logLik: -418.88  diff: 0.00025      197 iterations  logLik: -418.88  diff: 0.00024      198 iterations  logLik: -418.88  diff: 0.00024      199 iterations  logLik: -418.88  diff: 0.00023      200 iterations  logLik: -418.88  diff: 0.00022      201 iterations  logLik: -418.88  diff: 0.00022      202 iterations  logLik: -418.88  diff: 0.00021      203 iterations  logLik: -418.88  diff: 0.00020      204 iterations  logLik: -418.88  diff: 0.00020      205 iterations  logLik: -418.88  diff: 0.00019      206 iterations  logLik: -418.88  diff: 0.00019      207 iterations  logLik: -418.88  diff: 0.00018      208 iterations  logLik: -418.88  diff: 0.00018      209 iterations  logLik: -418.88  diff: 0.00017      210 iterations  logLik: -418.88  diff: 0.00017      211 iterations  logLik: -418.88  diff: 0.00016      212 iterations  logLik: -418.88  diff: 0.00016      213 iterations  logLik: -418.88  diff: 0.00015      214 iterations  logLik: -418.88  diff: 0.00015      215 iterations  logLik: -418.88  diff: 0.00014      216 iterations  logLik: -418.88  diff: 0.00014      217 iterations  logLik: -418.88  diff: 0.00013      218 iterations  logLik: -418.88  diff: 0.00013      219 iterations  logLik: -418.88  diff: 0.00013      220 iterations  logLik: -418.88  diff: 0.00012      221 iterations  logLik: -418.88  diff: 0.00012      222 iterations  logLik: -418.88  diff: 0.00012      223 iterations  logLik: -418.88  diff: 0.00011      224 iterations  logLik: -418.88  diff: 0.00011      225 iterations  logLik: -418.88  diff: 0.00011      226 iterations  logLik: -418.88  diff: 0.00010      227 iterations  logLik: -418.88  diff: 0.00010      228 iterations  logLik: -418.88  diff: 0.00010      229 iterations  logLik: -418.88  diff: 0.00009      230 iterations  logLik: -418.88  diff: 0.00009      231 iterations  logLik: -418.88  diff: 0.00009      232 iterations  logLik: -418.88  diff: 0.00009      233 iterations  logLik: -418.88  diff: 0.00008      234 iterations  logLik: -418.88  diff: 0.00008      235 iterations  logLik: -418.88  diff: 0.00008      236 iterations  logLik: -418.88  diff: 0.00008      237 iterations  logLik: -418.88  diff: 0.00008      238 iterations  logLik: -418.88  diff: 0.00007      239 iterations  logLik: -418.88  diff: 0.00007      240 iterations  logLik: -418.88  diff: 0.00007      241 iterations  logLik: -418.88  diff: 0.00007      242 iterations  logLik: -418.88  diff: 0.00007      243 iterations  logLik: -418.88  diff: 0.00006      244 iterations  logLik: -418.88  diff: 0.00006      245 iterations  logLik: -418.88  diff: 0.00006      246 iterations  logLik: -418.88  diff: 0.00006      247 iterations  logLik: -418.88  diff: 0.00006      248 iterations  logLik: -418.88  diff: 0.00005      249 iterations  logLik: -418.88  diff: 0.00005      250 iterations  logLik: -418.88  diff: 0.00005      251 iterations  logLik: -418.88  diff: 0.00005      252 iterations  logLik: -418.88  diff: 0.00005      253 iterations  logLik: -418.88  diff: 0.00005      254 iterations  logLik: -418.88  diff: 0.00005      255 iterations  logLik: -418.88  diff: 0.00004      256 iterations  logLik: -418.88  diff: 0.00004      257 iterations  logLik: -418.88  diff: 0.00004      258 iterations  logLik: -418.88  diff: 0.00004      259 iterations  logLik: -418.88  diff: 0.00004      260 iterations  logLik: -418.88  diff: 0.00004      261 iterations  logLik: -418.88  diff: 0.00004      262 iterations  logLik: -418.88  diff: 0.00004      263 iterations  logLik: -418.88  diff: 0.00004      264 iterations  logLik: -418.88  diff: 0.00003      265 iterations  logLik: -418.88  diff: 0.00003      266 iterations  logLik: -418.88  diff: 0.00003      267 iterations  logLik: -418.88  diff: 0.00003      268 iterations  logLik: -418.88  diff: 0.00003      269 iterations  logLik: -418.88  diff: 0.00003      270 iterations  logLik: -418.88  diff: 0.00003      271 iterations  logLik: -418.88  diff: 0.00003      272 iterations  logLik: -418.88  diff: 0.00003      273 iterations  logLik: -418.88  diff: 0.00003      274 iterations  logLik: -418.88  diff: 0.00003      275 iterations  logLik: -418.88  diff: 0.00003      276 iterations  logLik: -418.88  diff: 0.00002      277 iterations  logLik: -418.88  diff: 0.00002      278 iterations  logLik: -418.88  diff: 0.00002      279 iterations  logLik: -418.88  diff: 0.00002      280 iterations  logLik: -418.88  diff: 0.00002      281 iterations  logLik: -418.88  diff: 0.00002      282 iterations  logLik: -418.88  diff: 0.00002      283 iterations  logLik: -418.88  diff: 0.00002      284 iterations  logLik: -418.88  diff: 0.00002      285 iterations  logLik: -418.88  diff: 0.00002      286 iterations  logLik: -418.88  diff: 0.00002      287 iterations  logLik: -418.88  diff: 0.00002      288 iterations  logLik: -418.88  diff: 0.00002      289 iterations  logLik: -418.88  diff: 0.00002      290 iterations  logLik: -418.88  diff: 0.00002      291 iterations  logLik: -418.88  diff: 0.00002      292 iterations  logLik: -418.88  diff: 0.00002      293 iterations  logLik: -418.88  diff: 0.00002      294 iterations  logLik: -418.88  diff: 0.00001      295 iterations  logLik: -418.88  diff: 0.00001      296 iterations  logLik: -418.88  diff: 0.00001      297 iterations  logLik: -418.88  diff: 0.00001      298 iterations  logLik: -418.88  diff: 0.00001      299 iterations  logLik: -418.88  diff: 0.00001      300 iterations  logLik: -418.88  diff: 0.00001      301 iterations  logLik: -418.88  diff: 0.00001      302 iterations  logLik: -418.88  diff: 0.00001      303 iterations  logLik: -418.88  diff: 0.00001      304 iterations  logLik: -418.88  diff: 0.00001      305 iterations  logLik: -418.88  diff: 0.00001      306 iterations  logLik: -418.88  diff: 0.00001      307 iterations  logLik: -418.88  diff: 0.00001      308 iterations  logLik: -418.88  diff: 0.00001      309 iterations  logLik: -418.88  diff: 9.9e-06      310 iterations  logLik: -418.88  diff: 9.6e-06      311 iterations  logLik: -418.88  diff: 9.3e-06      312 iterations  logLik: -418.88  diff: 9.1e-06      313 iterations  logLik: -418.88  diff: 8.8e-06      314 iterations  logLik: -418.88  diff: 8.6e-06      315 iterations  logLik: -418.88  diff: 8.3e-06      316 iterations  logLik: -418.88  diff: 8.1e-06      317 iterations  logLik: -418.88  diff: 7.9e-06      318 iterations  logLik: -418.88  diff: 7.7e-06      319 iterations  logLik: -418.88  diff: 7.5e-06      320 iterations  logLik: -418.88  diff: 7.3e-06      321 iterations  logLik: -418.88  diff: 7.1e-06      322 iterations  logLik: -418.88  diff: 6.9e-06      323 iterations  logLik: -418.88  diff: 6.7e-06      324 iterations  logLik: -418.88  diff: 6.5e-06      325 iterations  logLik: -418.88  diff: 6.3e-06      326 iterations  logLik: -418.88  diff: 6.2e-06      327 iterations  logLik: -418.88  diff: 6.0e-06      328 iterations  logLik: -418.88  diff: 5.8e-06      329 iterations  logLik: -418.88  diff: 5.7e-06      330 iterations  logLik: -418.88  diff: 5.5e-06      331 iterations  logLik: -418.88  diff: 5.4e-06      332 iterations  logLik: -418.88  diff: 5.2e-06      333 iterations  logLik: -418.88  diff: 5.1e-06      334 iterations  logLik: -418.88  diff: 4.9e-06      335 iterations  logLik: -418.88  diff: 4.8e-06      336 iterations  logLik: -418.88  diff: 4.7e-06      337 iterations  logLik: -418.88  diff: 4.5e-06      338 iterations  logLik: -418.88  diff: 4.4e-06      339 iterations  logLik: -418.88  diff: 4.3e-06      340 iterations  logLik: -418.88  diff: 4.2e-06      341 iterations  logLik: -418.88  diff: 4.1e-06      342 iterations  logLik: -418.88  diff: 4.0e-06      343 iterations  logLik: -418.88  diff: 3.8e-06      344 iterations  logLik: -418.88  diff: 3.7e-06      345 iterations  logLik: -418.88  diff: 3.6e-06      346 iterations  logLik: -418.88  diff: 3.5e-06      347 iterations  logLik: -418.88  diff: 3.4e-06      348 iterations  logLik: -418.88  diff: 3.4e-06      349 iterations  logLik: -418.88  diff: 3.3e-06      350 iterations  logLik: -418.88  diff: 3.2e-06      351 iterations  logLik: -418.88  diff: 3.1e-06      352 iterations  logLik: -418.88  diff: 3.0e-06      353 iterations  logLik: -418.88  diff: 2.9e-06      354 iterations  logLik: -418.88  diff: 2.8e-06      355 iterations  logLik: -418.88  diff: 2.8e-06      356 iterations  logLik: -418.88  diff: 2.7e-06      357 iterations  logLik: -418.88  diff: 2.6e-06      358 iterations  logLik: -418.88  diff: 2.5e-06      359 iterations  logLik: -418.88  diff: 2.5e-06      360 iterations  logLik: -418.88  diff: 2.4e-06      361 iterations  logLik: -418.88  diff: 2.3e-06      362 iterations  logLik: -418.88  diff: 2.3e-06      363 iterations  logLik: -418.88  diff: 2.2e-06      364 iterations  logLik: -418.88  diff: 2.2e-06      365 iterations  logLik: -418.88  diff: 2.1e-06      366 iterations  logLik: -418.88  diff: 2.0e-06      367 iterations  logLik: -418.88  diff: 2.0e-06      368 iterations  logLik: -418.88  diff: 1.9e-06      369 iterations  logLik: -418.88  diff: 1.9e-06      370 iterations  logLik: -418.88  diff: 1.8e-06      371 iterations  logLik: -418.88  diff: 1.8e-06      372 iterations  logLik: -418.88  diff: 1.7e-06      373 iterations  logLik: -418.88  diff: 1.7e-06      374 iterations  logLik: -418.88  diff: 1.6e-06      375 iterations  logLik: -418.88  diff: 1.6e-06      376 iterations  logLik: -418.88  diff: 1.6e-06      377 iterations  logLik: -418.88  diff: 1.5e-06      378 iterations  logLik: -418.88  diff: 1.5e-06      379 iterations  logLik: -418.88  diff: 1.4e-06      380 iterations  logLik: -418.88  diff: 1.4e-06      381 iterations  logLik: -418.88  diff: 1.4e-06      382 iterations  logLik: -418.88  diff: 1.3e-06      383 iterations  logLik: -418.88  diff: 1.3e-06      384 iterations  logLik: -418.88  diff: 1.3e-06      385 iterations  logLik: -418.88  diff: 1.2e-06      386 iterations  logLik: -418.88  diff: 1.2e-06      387 iterations  logLik: -418.88  diff: 1.2e-06      388 iterations  logLik: -418.88  diff: 1.1e-06      389 iterations  logLik: -418.88  diff: 1.1e-06      390 iterations  logLik: -418.88  diff: 1.1e-06      391 iterations  logLik: -418.88  diff: 1.0e-06      392 iterations  logLik: -418.88  diff: 1.0e-06      393 iterations  logLik: -418.88  diff: 9.8e-07      394 iterations  logLik: -418.88  diff: 9.5e-07      395 iterations  logLik: -418.88  diff: 9.3e-07      396 iterations  logLik: -418.88  diff: 9.0e-07      397 iterations  logLik: -418.88  diff: 8.8e-07      398 iterations  logLik: -418.88  diff: 8.6e-07      399 iterations  logLik: -418.88  diff: 8.3e-07      400 iterations  logLik: -418.88  diff: 8.1e-07      401 iterations  logLik: -418.88  diff: 7.9e-07      402 iterations  logLik: -418.88  diff: 7.7e-07      403 iterations  logLik: -418.88  diff: 7.5e-07      404 iterations  logLik: -418.88  diff: 7.3e-07      405 iterations  logLik: -418.88  diff: 7.1e-07      406 iterations  logLik: -418.88  diff: 6.9e-07      407 iterations  logLik: -418.88  diff: 6.7e-07      408 iterations  logLik: -418.88  diff: 6.5e-07      409 iterations  logLik: -418.88  diff: 6.3e-07      410 iterations  logLik: -418.88  diff: 6.2e-07      411 iterations  logLik: -418.88  diff: 6.0e-07      412 iterations  logLik: -418.88  diff: 5.9e-07      413 iterations  logLik: -418.88  diff: 5.7e-07      414 iterations  logLik: -418.88  diff: 5.5e-07      415 iterations  logLik: -418.88  diff: 5.4e-07      416 iterations  logLik: -418.88  diff: 5.3e-07      417 iterations  logLik: -418.88  diff: 5.1e-07      418 iterations  logLik: -418.88  diff: 5.0e-07      419 iterations  logLik: -418.88  diff: 4.8e-07      420 iterations  logLik: -418.88  diff: 4.7e-07      421 iterations  logLik: -418.88  diff: 4.6e-07      422 iterations  logLik: -418.88  diff: 4.5e-07      423 iterations  logLik: -418.88  diff: 4.3e-07      424 iterations  logLik: -418.88  diff: 4.2e-07      425 iterations  logLik: -418.88  diff: 4.1e-07      426 iterations  logLik: -418.88  diff: 4.0e-07      427 iterations  logLik: -418.88  diff: 3.9e-07      428 iterations  logLik: -418.88  diff: 3.8e-07      429 iterations  logLik: -418.88  diff: 3.7e-07      430 iterations  logLik: -418.88  diff: 3.6e-07      431 iterations  logLik: -418.88  diff: 3.5e-07      432 iterations  logLik: -418.88  diff: 3.4e-07      433 iterations  logLik: -418.88  diff: 3.3e-07      434 iterations  logLik: -418.88  diff: 3.2e-07      435 iterations  logLik: -418.88  diff: 3.1e-07      436 iterations  logLik: -418.88  diff: 3.1e-07      437 iterations  logLik: -418.88  diff: 3.0e-07      438 iterations  logLik: -418.88  diff: 2.9e-07      439 iterations  logLik: -418.88  diff: 2.8e-07      440 iterations  logLik: -418.88  diff: 2.7e-07      441 iterations  logLik: -418.88  diff: 2.7e-07      442 iterations  logLik: -418.88  diff: 2.6e-07      443 iterations  logLik: -418.88  diff: 2.5e-07      444 iterations  logLik: -418.88  diff: 2.5e-07      445 iterations  logLik: -418.88  diff: 2.4e-07      446 iterations  logLik: -418.88  diff: 2.3e-07      447 iterations  logLik: -418.88  diff: 2.3e-07      448 iterations  logLik: -418.88  diff: 2.2e-07      449 iterations  logLik: -418.88  diff: 2.2e-07      450 iterations  logLik: -418.88  diff: 2.1e-07      451 iterations  logLik: -418.88  diff: 2.0e-07      452 iterations  logLik: -418.88  diff: 2.0e-07      453 iterations  logLik: -418.88  diff: 1.9e-07      454 iterations  logLik: -418.88  diff: 1.9e-07      455 iterations  logLik: -418.88  diff: 1.8e-07      456 iterations  logLik: -418.88  diff: 1.8e-07      457 iterations  logLik: -418.88  diff: 1.7e-07      458 iterations  logLik: -418.88  diff: 1.7e-07      459 iterations  logLik: -418.88  diff: 1.7e-07      460 iterations  logLik: -418.88  diff: 1.6e-07      461 iterations  logLik: -418.88  diff: 1.6e-07      462 iterations  logLik: -418.88  diff: 1.5e-07      463 iterations  logLik: -418.88  diff: 1.5e-07      464 iterations  logLik: -418.88  diff: 1.4e-07      465 iterations  logLik: -418.88  diff: 1.4e-07      466 iterations  logLik: -418.88  diff: 1.4e-07      467 iterations  logLik: -418.88  diff: 1.3e-07      468 iterations  logLik: -418.88  diff: 1.3e-07      469 iterations  logLik: -418.88  diff: 1.3e-07      470 iterations  logLik: -418.88  diff: 1.2e-07      471 iterations  logLik: -418.88  diff: 1.2e-07      472 iterations  logLik: -418.88  diff: 1.2e-07      473 iterations  logLik: -418.88  diff: 1.1e-07      474 iterations  logLik: -418.88  diff: 1.1e-07      475 iterations  logLik: -418.88  diff: 1.1e-07      476 iterations  logLik: -418.88  diff: 1.0e-07      477 iterations  logLik: -418.88  diff: 1.0e-07      478 iterations  logLik: -418.88  diff: 1.0e-07      479 iterations  logLik: -418.88  diff: 9.7e-08      480 iterations  logLik: -418.88  diff: 9.4e-08      481 iterations  logLik: -418.88  diff: 9.2e-08      482 iterations  logLik: -418.88  diff: 8.9e-08      483 iterations  logLik: -418.88  diff: 8.7e-08      484 iterations  logLik: -418.88  diff: 8.5e-08      485 iterations  logLik: -418.88  diff: 8.3e-08      486 iterations  logLik: -418.88  diff: 8.0e-08      487 iterations  logLik: -418.88  diff: 7.8e-08      488 iterations  logLik: -418.88  diff: 7.6e-08      489 iterations  logLik: -418.88  diff: 7.4e-08      490 iterations  logLik: -418.88  diff: 7.2e-08      491 iterations  logLik: -418.88  diff: 7.0e-08      492 iterations  logLik: -418.88  diff: 6.9e-08      493 iterations  logLik: -418.88  diff: 6.7e-08      494 iterations  logLik: -418.88  diff: 6.5e-08      495 iterations  logLik: -418.88  diff: 6.3e-08      496 iterations  logLik: -418.88  diff: 6.2e-08      497 iterations  logLik: -418.88  diff: 6.0e-08      498 iterations  logLik: -418.88  diff: 5.9e-08      499 iterations  logLik: -418.88  diff: 5.7e-08      500 iterations  logLik: -418.88  diff: 5.6e-08      501 iterations  logLik: -418.88  diff: 5.4e-08      502 iterations  logLik: -418.88  diff: 5.3e-08      503 iterations  logLik: -418.88  diff: 5.1e-08      504 iterations  logLik: -418.88  diff: 5.0e-08      505 iterations  logLik: -418.88  diff: 4.9e-08      506 iterations  logLik: -418.88  diff: 4.7e-08      507 iterations  logLik: -418.88  diff: 4.6e-08      508 iterations  logLik: -418.88  diff: 4.5e-08      509 iterations  logLik: -418.88  diff: 4.4e-08      510 iterations  logLik: -418.88  diff: 4.3e-08      511 iterations  logLik: -418.88  diff: 4.2e-08      512 iterations  logLik: -418.88  diff: 4.0e-08      513 iterations  logLik: -418.88  diff: 3.9e-08      514 iterations  logLik: -418.88  diff: 3.8e-08      515 iterations  logLik: -418.88  diff: 3.7e-08      516 iterations  logLik: -418.88  diff: 3.6e-08      517 iterations  logLik: -418.88  diff: 3.5e-08      518 iterations  logLik: -418.88  diff: 3.5e-08      519 iterations  logLik: -418.88  diff: 3.4e-08      520 iterations  logLik: -418.88  diff: 3.3e-08      521 iterations  logLik: -418.88  diff: 3.2e-08      522 iterations  logLik: -418.88  diff: 3.1e-08      523 iterations  logLik: -418.88  diff: 3.0e-08      524 iterations  logLik: -418.88  diff: 2.9e-08      525 iterations  logLik: -418.88  diff: 2.9e-08      526 iterations  logLik: -418.88  diff: 2.8e-08      527 iterations  logLik: -418.88  diff: 2.7e-08      528 iterations  logLik: -418.88  diff: 2.7e-08      529 iterations  logLik: -418.88  diff: 2.6e-08      530 iterations  logLik: -418.88  diff: 2.5e-08      531 iterations  logLik: -418.88  diff: 2.5e-08      532 iterations  logLik: -418.88  diff: 2.4e-08      533 iterations  logLik: -418.88  diff: 2.3e-08      534 iterations  logLik: -418.88  diff: 2.3e-08      535 iterations  logLik: -418.88  diff: 2.2e-08      536 iterations  logLik: -418.88  diff: 2.2e-08      537 iterations  logLik: -418.88  diff: 2.1e-08      538 iterations  logLik: -418.88  diff: 2.0e-08      539 iterations  logLik: -418.88  diff: 2.0e-08      540 iterations  logLik: -418.88  diff: 1.9e-08      541 iterations  logLik: -418.88  diff: 1.9e-08      542 iterations  logLik: -418.88  diff: 1.8e-08      543 iterations  logLik: -418.88  diff: 1.8e-08      544 iterations  logLik: -418.88  diff: 1.7e-08      545 iterations  logLik: -418.88  diff: 1.7e-08      546 iterations  logLik: -418.88  diff: 1.7e-08      547 iterations  logLik: -418.88  diff: 1.6e-08      548 iterations  logLik: -418.88  diff: 1.6e-08      549 iterations  logLik: -418.88  diff: 1.5e-08      550 iterations  logLik: -418.88  diff: 1.5e-08      551 iterations  logLik: -418.88  diff: 1.5e-08      552 iterations  logLik: -418.88  diff: 1.4e-08      553 iterations  logLik: -418.88  diff: 1.4e-08      554 iterations  logLik: -418.88  diff: 1.3e-08      555 iterations  logLik: -418.88  diff: 1.3e-08      556 iterations  logLik: -418.88  diff: 1.3e-08      557 iterations  logLik: -418.88  diff: 1.2e-08      558 iterations  logLik: -418.88  diff: 1.2e-08      559 iterations  logLik: -418.88  diff: 1.2e-08      560 iterations  logLik: -418.88  diff: 1.1e-08      561 iterations  logLik: -418.88  diff: 1.1e-08      562 iterations  logLik: -418.88  diff: 1.1e-08      563 iterations  logLik: -418.88  diff: 1.1e-08      564 iterations  logLik: -418.88  diff: 1.0e-08      565 iterations  logLik: -418.88  diff: 1.0e-08      566 iterations  logLik: -418.88  diff: 9.8e-09      
#> .. done. 
#> nlm iteration begin.
#> .. done.
#> Structural Latent Variable Model
#> 
#> Latent variables (Root*):           
#>  Label: lc*
#> nclass: 3  
#> 
#> Measurement model:                            
#>  lc -> { y1, y2, y3, y4 }  a
#> 
#> LogLik: -418.8761

# Parameter restriction
mf <- estimate(m, dt$response)
param(mf, index = TRUE)
#> PI :
#> (lc)
#>   class
#>              1           2           3
#>     0.3165 (1)  0.2838 (2)  0.3998 (3)
#> 
#> RHO :
#> (a)
#>         class
#> response            1            2            3
#>    1(V1)   0.0000 (4)  1.0000 (12)  0.9286 (20)
#>    2       1.0000 (5)  0.0000 (13)  0.0714 (21)
#>    1(V2)   0.1501 (6)  0.8090 (14)  0.8703 (22)
#>    2       0.8499 (7)  0.1910 (15)  0.1297 (23)
#>    1(V3)   0.1025 (8)  0.0000 (16)  0.8069 (24)
#>    2       0.8975 (9)  1.0000 (17)  0.1931 (25)
#>    1(V4)  0.1107 (10)  0.0000 (18)  0.9129 (26)
#>    2      0.8893 (11)  1.0000 (19)  0.0871 (27)
#> 
#>    V1 V2 V3 V4
#> lc y1 y2 y3 y4
mf0 <- estimate(mf, fix2zero = c(4, 6, 8, 10))
param(mf0)
#> PI :
#> (lc)
#>   class
#>          1       2       3
#>     0.2067  0.3119  0.4814
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  0.0000  0.8754  0.7935
#>    2      1.0000  0.1246  0.2065
#>    1(V2)  0.0000  0.8127  0.7717
#>    2      1.0000  0.1873  0.2283
#>    1(V3)  0.0000  0.0000  0.7375
#>    2      1.0000  1.0000  0.2625
#>    1(V4)  0.0000  0.0000  0.8310
#>    2      1.0000  1.0000  0.1690
#> 
#>    V1 V2 V3 V4
#> lc y1 y2 y3 y4

# Estimation control
estimate(m, dt$response, control = slcaControl(nrep = 3, verbose = TRUE))
#> Inital parameter test: 
#> 1 / 3  logLik: -418.6869 
#> 2 / 3  logLik: -418.6869 
#> 3 / 3  logLik: -418.8761 
#> 
#> 1th parameter set has been selected.
#> 
#> EM iteration begin.
#> 1 iterations  logLik: -418.69  diff: inf      2 iterations  logLik: -418.69  diff: 2.4e-10      
#> .. done.
#> Structural Latent Variable Model
#> 
#> Latent variables (Root*):           
#>  Label: lc*
#> nclass: 3  
#> 
#> Measurement model:                            
#>  lc -> { y1, y2, y3, y4 }  a
#> 
#> LogLik: -418.6869
estimate(m, dt$response, control = slcaControl(init.param = c(pi, rho)))
#> Structural Latent Variable Model
#> 
#> Latent variables (Root*):           
#>  Label: lc*
#> nclass: 3  
#> 
#> Measurement model:                            
#>  lc -> { y1, y2, y3, y4 }  a
#> 
#> LogLik: -418.6869