Estimate Parameters of slca Object

Estimate the parameters of model constructed using the slca function.

Usage

estimate(x, ...)

# S3 method for slca
estimate(x, data,
    method = c("em", "hybrid", "nlm"),
    fix2zero = NULL,
    control = slcaControl(), ...)

Arguments

x

an slca object defining SLCM model to be estimated.

...

additional arguments.

data

a data.frame object containing observed categorical variables incorporated in the model.

method

estimation method for SLCM parameters. The default is "em", which employs expecation-maximization (EM) algorithm for estimation; the alternative "nlm", utilizes nlm function for Newton-Raphson algorithm. The "hybrid" method begins with the EM algorithm and concludes with the nlm function for refined estimation.

fix2zero

a vector of parameters to be restricted to zero. The details of restriction is given under 'Details'

control

a list of control for the estimation procedure. Used to modify default values in slcaControl.

Value

An object of class slca and estimated with an following elements:

model

a list describing of the model.

method

the method used for estimation

arg

the brief model description used during the estimation.

mf

the data.frame used for estimation.

par

the log of the estimated paramters.

logit

the log-odds of the estimated parameters.

score

the score function for the estimated parameters.

posterior

the list of posterior probablities for each latent class variable.

convergence

a logical indicator of whether convergence was achieved.

loglikelihood

the loglikelihood of the estimated model.

control

the control values used during the estimation process.

This returned object can be further processed using the param functions to extract the estimated parameters or their respective standard errors. Additionally, the regress function enables logistic regression analysis using three-step approach to evaluate the effect of external variables on latent class variables.

Details

To constrain certain parameters to zero, use the fix2zero argument. Each parameter is associated with a unique index. You can identify the index of a specific parameter by invoking the param function with the index = TRUE argument. To apply these constraints, include the relevant parameter indices in the fix2zero argument.

See also