GSS 1976-1977 Data on Social Status and Tolerance towards Minorities

This dataset contains responses from the General Social Survey (GSS) for the years 1976 and 1977, focusing on social status and tolerance towards minorities. The dataset can be used to replicate the analyses conducted in McCutcheon (1985) and Bakk et al. (2014).
It includes covariates such as interview year, age, sex, race, education level, and income. Social status-related variables include father's occupation and education level, as well as mother's education level. Tolerance towards minorities is measured by agreement with three questions: (1) allowing public speaking, (2) allowing teaching, and (3) allowing literature publication.

Usage

gss7677

Format

A data frame with 2942 rows and 14 variables:

YEAR

Interview year (1976, 1977).

COHORT

Respondent's age cohort.
Levels: (1) YOUNG, (2) YOUNG-MIDDLE, (4) MIDDLE, (5) OLD.

SEX

Respondent's sex.
Levels: (1) MALE, (2) FEMALE.

RACE

Respondent's race.
Levels: (1) WHITE, (2) BLACK, (3) OTHER.

DEGREE

Respondent's education level.
Levels: (1) LT HS, (2) HIGH-SCH, (3) HIGHER.

REALRINC

Respondent's income.

PAPRES

Father's occupational prestige.
Levels: (1) LOW, (2) MEDIUM, (3) HIGH.

PADEG

Father's education level.
Levels: (1) LT HS, (2) HIGH-SCH, (3) COLLEGE, (4) BACHELOR, (5) GRADUATE.

MADEG

Mother's education level.
Levels: (1) LT HS, (2) HIGH-SCH, (3) COLLEGE, (4) BACHELOR, (5) GRADUATE.

TOLRAC

Tolerance towards racists.

TOLCOM

Tolerance towards communists.

TOLHOMO

Tolerance towards homosexuals.

TOLATH

Tolerance towards atheists.

TOLMIL

Tolerance towards militarists.

Source

General Social Survey (GSS) 1976, 1977

References

Bakk Z, Kuha J. (2021) Relating latent class membership to external variables: An overview. Br J Math Stat Psychol. 74(2):340-362.

McCutcheon, A. L. (1985). A latent class analysis of tolerance for nonconformity in the American public. Public Opinion Quarterly, 49, 474–488.

Examples

library(magrittr)
data <- gss7677[gss7677$RACE == "BLACK",]
model_stat <- slca(status(3) ~ PAPRES + PADEG + MADEG) %>%
   estimate(data = data, control = list(em.tol = 1e-6))
summary(model_stat)
#> Structural Latent Class Model
#> 
#> Summary of model structure
#>                                    
#>  Number of manifest variables     3
#>  Number of latent class variables 1
#> 
#>  Latent variables (Root*):                
#>   Label: status*
#>  nclass: 3      
#> 
#>  Measurement model:                                       
#>   status -> { PAPRES, PADEG, MADEG }  a
#> 
#> 
#> Summary of manifest variables
#> 
#>  Categories for each variable:
#>          response
#>            1      2         3        4         5       
#>    PAPRES  LOW    MIDIUM    HIGH                       
#>    PADEG   LT-HS  HIGH-SCH  COLLEGE  BACHELOR  GRADUATE
#>    MADEG   LT-HS  HIGH-SCH  COLLEGE  BACHELOR  GRADUATE
#> 
#>  Frequencies for each categories:
#>          response
#>              1    2  3  4  5  <NA>
#>    PAPRES   80  114  8          85
#>    PADEG   155   27  1     2   102
#>    MADEG   183   44  4  4  1    51
#> 
#> 
#> Summary of model fit
#>                                            
#>  Number of observations                 287
#>  Number of free parameters               32
#>  Log-likelihood                    -383.308
#>  Information criteria                      
#>    Akaike (AIC)                     830.616
#>    Bayesian (BIC)                   947.720
#>  Chi-squared Tests                         
#>    Residual degree of freedom (df)       42
#>    Pearson Chi-squared (X-squared)   39.335
#>      P(>Chi)                          0.589
#>    Likelihood Ratio (G-squared)      22.691
#>      P(>Chi)                          0.993
param(model_stat)
#> PI :
#> (status)
#>   class
#>          1       2       3
#>     0.8401  0.0248  0.1351
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  0.4135  0.0000  0.3655
#>    2      0.5721  0.0000  0.6345
#>    3      0.0144  1.0000  0.0000
#>    1(V2)  1.0000  0.0000  0.0000
#>    2      0.0000  0.4214  1.0000
#>    3      0.0000  0.1929  0.0000
#>    4      0.0000  0.0000  0.0000
#>    5      0.0000  0.3857  0.0000
#>    1(V3)  0.8690  0.4966  0.2446
#>    2      0.1243  0.3627  0.5411
#>    3      0.0000  0.0000  0.1285
#>    4      0.0067  0.0000  0.0859
#>    5      0.0000  0.1406  0.0000
#> 
#>        V1     V2    V3   
#> status PAPRES PADEG MADEG

model_tol <- slca(tol(4) ~ TOLRAC + TOLCOM + TOLHOMO + TOLATH + TOLMIL) %>%
   estimate(data = data, control = list(em.tol = 1e-6))
summary(model_tol)
#> Structural Latent Class Model
#> 
#> Summary of model structure
#>                                    
#>  Number of manifest variables     5
#>  Number of latent class variables 1
#> 
#>  Latent variables (Root*):             
#>   Label: tol*
#>  nclass: 4   
#> 
#>  Measurement model:                                                       
#>   tol -> { TOLRAC, TOLCOM, TOLHOMO, TOLATH, TOLMIL }  a
#> 
#> 
#> Summary of manifest variables
#> 
#>  Categories for each variable:
#>           response
#>             1         2         
#>    TOLRAC   TOLERANT  INTOLERANT
#>    TOLCOM   TOLERANT  INTOLERANT
#>    TOLHOMO  TOLERANT  INTOLERANT
#>    TOLATH   TOLERANT  INTOLERANT
#>    TOLMIL   TOLERANT  INTOLERANT
#> 
#>  Frequencies for each categories:
#>           response
#>               1    2  <NA>
#>    TOLRAC    64  216     7
#>    TOLCOM    83  184    20
#>    TOLHOMO  103  164    20
#>    TOLATH    67  212     8
#>    TOLMIL    74  203    10
#> 
#> 
#> Summary of model fit
#>                                            
#>  Number of observations                 287
#>  Number of free parameters               23
#>  Log-likelihood                    -644.804
#>  Information criteria                      
#>    Akaike (AIC)                    1335.609
#>    Bayesian (BIC)                  1419.777
#>  Chi-squared Tests                         
#>    Residual degree of freedom (df)        8
#>    Pearson Chi-squared (X-squared)    6.773
#>      P(>Chi)                          0.561
#>    Likelihood Ratio (G-squared)       7.099
#>      P(>Chi)                          0.526
param(model_tol)
#> PI :
#> (tol)
#>   class
#>          1       2       3       4
#>     0.5747  0.1248  0.0733  0.2272
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3       4
#>    1(V1)  0.0000  1.0000  0.1747  0.4018
#>    2      1.0000  0.0000  0.8253  0.5982
#>    1(V2)  0.0654  0.9714  1.0000  0.3509
#>    2      0.9346  0.0286  0.0000  0.6491
#>    1(V3)  0.1342  0.9003  0.9531  0.5254
#>    2      0.8658  0.0997  0.0469  0.4746
#>    1(V4)  0.0084  1.0000  0.1852  0.4540
#>    2      0.9916  0.0000  0.8148  0.5460
#>    1(V5)  0.1057  0.8657  0.6944  0.2190
#>    2      0.8943  0.1343  0.3056  0.7810
#> 
#>     V1     V2     V3      V4     V5    
#> tol TOLRAC TOLCOM TOLHOMO TOLATH TOLMIL

model_lta <- slca(
   status(3) ~ PAPRES + PADEG + MADEG,
   tol(4) ~ TOLRAC + TOLCOM + TOLHOMO + TOLATH + TOLMIL,
   status ~ tol
) %>% estimate(data = data, control = list(em.tol = 1e-6))
summary(model_lta)
#> Structural Latent Class Model
#> 
#> Summary of model structure
#>                                    
#>  Number of manifest variables     8
#>  Number of latent class variables 2
#> 
#>  Latent variables (Root*):                    
#>   Label: status* tol
#>  nclass: 3       4  
#> 
#>  Measurement model:                                                          
#>   status -> { PAPRES, PADEG, MADEG }                     a
#>   tol    -> { TOLRAC, TOLCOM, TOLHOMO, TOLATH, TOLMIL }  b
#> 
#>  Structural model:                   
#>   status -> { tol }
#> 
#>  Dependency constraints:
#>   A            
#>   status -> tol
#> 
#>  Tree of structural model:                
#>   status  -> tol
#> 
#> 
#> Summary of manifest variables
#> 
#>  Categories for each variable:
#>           response
#>             1         2           3        4         5       
#>    PAPRES   LOW       MIDIUM      HIGH                       
#>    PADEG    LT-HS     HIGH-SCH    COLLEGE  BACHELOR  GRADUATE
#>    MADEG    LT-HS     HIGH-SCH    COLLEGE  BACHELOR  GRADUATE
#>    TOLRAC   TOLERANT  INTOLERANT                             
#>    TOLCOM   TOLERANT  INTOLERANT                             
#>    TOLHOMO  TOLERANT  INTOLERANT                             
#>    TOLATH   TOLERANT  INTOLERANT                             
#>    TOLMIL   TOLERANT  INTOLERANT                             
#> 
#>  Frequencies for each categories:
#>           response
#>               1    2  3  4  5  <NA>
#>    PAPRES    80  114  8          85
#>    PADEG    155   27  1     2   102
#>    MADEG    183   44  4  4  1    51
#>    TOLRAC    64  216              7
#>    TOLCOM    83  184             20
#>    TOLHOMO  103  164             20
#>    TOLATH    67  212              8
#>    TOLMIL    74  203             10
#> 
#> 
#> Summary of model fit
#>                                                                                                                                                                                                                                                                                 
#>  Number of observations                                                                                                                                                                                                                                                      287
#>  Number of free parameters                                                                                                                                                                                                                                                    61
#>  Log-likelihood                                                                                                                                                                                                                                                        -1025.425
#>  Information criteria                                                                                                                                                                                                                                                           
#>    Akaike (AIC)                                                                                                                                                                                                                                                         2172.849
#>    Bayesian (BIC)                                                                                                                                                                                                                                                       2396.078
#>  Chi-squared Tests                                                                                                                                                                                                                                                              
#>    Residual degree of freedom (df)                                                                                                                                                                                                                                          2338
#>    Pearson Chi-squared (X-squared) 19661246208730520495373505396657235393177407952771820797382919779510847070522225252898069812498708512351352948583100018876338925262531004990118577748400009089285864903200196542365988500283327334842136812055310319701462839095103324160.000
#>      P(>Chi)                                                                                                                                                                                                                                                               0.000
#>    Likelihood Ratio (G-squared)                                                                                                                                                                                                                                          221.469
#>      P(>Chi)                                                                                                                                                                                                                                                               1.000
param(model_lta)
#> PI :
#> (status)
#>   class
#>          1       2       3
#>     0.6945  0.1591  0.1464
#> 
#> TAU :
#> (A)
#>      parent
#> child       1       2       3
#>     1  0.1600  0.4145  0.6216
#>     2  0.3541  0.4033  0.0000
#>     3  0.0420  0.1823  0.3784
#>     4  0.4439  0.0000  0.0001
#>              
#> parent status
#> child  tol   
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  0.3401  0.3164  0.7560
#>    2      0.6468  0.5392  0.1830
#>    3      0.0131  0.1444  0.0610
#>    1(V2)  1.0000  0.0000  0.9598
#>    2      0.0000  0.9310  0.0000
#>    3      0.0000  0.0000  0.0402
#>    4      0.0000  0.0000  0.0000
#>    5      0.0000  0.0690  0.0000
#>    1(V3)  0.9284  0.3046  0.5666
#>    2      0.0647  0.5141  0.4044
#>    3      0.0000  0.1057  0.0000
#>    4      0.0069  0.0756  0.0000
#>    5      0.0000  0.0000  0.0290
#> 
#>        V1     V2    V3   
#> status PAPRES PADEG MADEG
#> (b)
#>         class
#> response       1       2       3       4
#>    1(V1)  0.3812  0.0445  1.0000  0.0000
#>    2      0.6188  0.9555  0.0000  1.0000
#>    1(V2)  0.5989  0.1007  0.9875  0.0222
#>    2      0.4011  0.8993  0.0125  0.9778
#>    1(V3)  0.6907  0.2901  0.9038  0.0000
#>    2      0.3093  0.7099  0.0962  1.0000
#>    1(V4)  0.4168  0.0606  1.0000  0.0096
#>    2      0.5832  0.9394  0.0000  0.9904
#>    1(V5)  0.4097  0.0000  0.8879  0.1909
#>    2      0.5903  1.0000  0.1121  0.8091
#> 
#>     V1     V2     V3      V4     V5    
#> tol TOLRAC TOLCOM TOLHOMO TOLATH TOLMIL

# \donttest{
regress(model_lta, status ~ SEX, data)
#> Coefficients:     
#> class  (Intercept)  SEXFEMALE
#>   1/3   1.5640      -0.0263  
#>   2/3   0.1054      -0.5909  
regress(model_lta, status ~ SEX, data, method = "BCH")
#> Coefficients:     
#> class  (Intercept)  SEXFEMALE
#>   1/3   1.5692      -0.0609  
#>   2/3   0.1167      -0.9402  
regress(model_lta, status ~ SEX, data, method = "ML")
#> Coefficients:     
#> class  (Intercept)  SEXFEMALE
#>   1/3   1.4363      -0.0531  
#>   2/3  -0.4609      -1.2074  
# }