GSS 1976-1977 Data on Social Status and Tolerance towards Minorities

This dataset contains responses from the General Social Survey (GSS) for the years 1976 and 1977, focusing on social status and tolerance towards minorities. The dataset can be used to replicate the analyses conducted in McCutcheon (1985) and Bakk et al. (2014).
It includes covariates such as interview year, age, sex, race, education level, and income. Social status-related variables include father's occupation and education level, as well as mother's education level. Tolerance towards minorities is measured by agreement with three questions: (1) allowing public speaking, (2) allowing teaching, and (3) allowing literature publication.

Usage

gss7677

Format

A data frame with 2942 rows and 14 variables:

YEAR

Interview year (1976, 1977).

COHORT

Respondent's age cohort.
Levels: (1) YOUNG, (2) YOUNG-MIDDLE, (4) MIDDLE, (5) OLD.

SEX

Respondent's sex.
Levels: (1) MALE, (2) FEMALE.

RACE

Respondent's race.
Levels: (1) WHITE, (2) BLACK, (3) OTHER.

DEGREE

Respondent's education level.
Levels: (1) LT HS, (2) HIGH-SCH, (3) HIGHER.

REALRINC

Respondent's income.

PAPRES

Father's occupational prestige.
Levels: (1) LOW, (2) MEDIUM, (3) HIGH.

PADEG

Father's education level.
Levels: (1) LT HS, (2) HIGH-SCH, (3) COLLEGE, (4) BACHELOR, (5) GRADUATE.

MADEG

Mother's education level.
Levels: (1) LT HS, (2) HIGH-SCH, (3) COLLEGE, (4) BACHELOR, (5) GRADUATE.

TOLRAC

Tolerance towards racists.

TOLCOM

Tolerance towards communists.

TOLHOMO

Tolerance towards homosexuals.

TOLATH

Tolerance towards atheists.

TOLMIL

Tolerance towards militarists.

Source

General Social Survey (GSS) 1976, 1977

References

Bakk Z, Kuha J. (2021) Relating latent class membership to external variables: An overview. Br J Math Stat Psychol. 74(2):340-362.

McCutcheon, A. L. (1985). A latent class analysis of tolerance for nonconformity in the American public. Public Opinion Quarterly, 49, 474–488.

Examples

library(magrittr)
gss500 <- gss7677[1:500,] %>% na.omit
model_stat <- slca(status(3) ~ PAPRES + PADEG + MADEG) %>%
   estimate(data = gss500, control = list(em.tol = 1e-6))
summary(model_stat)
#> Structural Latent Class Model
#> 
#> Summary of model structure
#>                                    
#>  Number of manifest variables     3
#>  Number of latent class variables 1
#> 
#>  Latent variables (Root*):                
#>   Label: status*
#>  nclass: 3      
#> 
#>  Measurement model:                                       
#>   status -> { PAPRES, PADEG, MADEG }  a
#> 
#> 
#> Summary of manifest variables
#> 
#>  Categories for each variable:
#>          response
#>            1      2         3        4         5       
#>    PAPRES  LOW    MIDIUM    HIGH                       
#>    PADEG   LT-HS  HIGH-SCH  COLLEGE  BACHELOR  GRADUATE
#>    MADEG   LT-HS  HIGH-SCH  COLLEGE  BACHELOR  GRADUATE
#> 
#>  Frequencies for each categories:
#>          response
#>              1    2   3   4  5  <NA>
#>    PAPRES   62  102  19            0
#>    PADEG   104   63      11  5     0
#>    MADEG    86   84   4   8  1     0
#> 
#> 
#> Summary of model fit
#>                                            
#>  Number of observations                 183
#>  Number of free parameters               32
#>  Log-likelihood                    -477.892
#>  Information criteria                      
#>    Akaike (AIC)                    1019.785
#>    Bayesian (BIC)                  1122.488
#>  Chi-squared Tests                         
#>    Residual degree of freedom (df)       42
#>    Pearson Chi-squared (X-squared)   21.976
#>      P(>Chi)                          0.995
#>    Likelihood Ratio (G-squared)      23.705
#>      P(>Chi)                          0.990
param(model_stat)
#> PI :
#> (status)
#>   class
#>          1       2       3
#>     0.4289  0.1893  0.3817
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  0.4602  0.0000  0.3704
#>    2      0.4777  0.6718  0.5901
#>    3      0.0621  0.3282  0.0395
#>    1(V2)  0.4401  0.0000  0.9942
#>    2      0.5344  0.6076  0.0000
#>    3      0.0000  0.0000  0.0000
#>    4      0.0255  0.2598  0.0000
#>    5      0.0000  0.1326  0.0058
#>    1(V3)  0.1460  0.1496  0.9929
#>    2      0.8104  0.5884  0.0000
#>    3      0.0000  0.1155  0.0000
#>    4      0.0436  0.1177  0.0071
#>    5      0.0000  0.0289  0.0000
#> 
#>        V1     V2    V3   
#> status PAPRES PADEG MADEG

model_tol <- slca(tol(4) ~ TOLRAC + TOLCOM + TOLHOMO + TOLATH + TOLMIL) %>%
   estimate(data = gss500, control = list(em.tol = 1e-6))
summary(model_tol)
#> Structural Latent Class Model
#> 
#> Summary of model structure
#>                                    
#>  Number of manifest variables     5
#>  Number of latent class variables 1
#> 
#>  Latent variables (Root*):             
#>   Label: tol*
#>  nclass: 4   
#> 
#>  Measurement model:                                                       
#>   tol -> { TOLRAC, TOLCOM, TOLHOMO, TOLATH, TOLMIL }  a
#> 
#> 
#> Summary of manifest variables
#> 
#>  Categories for each variable:
#>           response
#>             1         2         
#>    TOLRAC   TOLERANT  INTOLERANT
#>    TOLCOM   TOLERANT  INTOLERANT
#>    TOLHOMO  TOLERANT  INTOLERANT
#>    TOLATH   TOLERANT  INTOLERANT
#>    TOLMIL   TOLERANT  INTOLERANT
#> 
#>  Frequencies for each categories:
#>           response
#>              1    2  <NA>
#>    TOLRAC   80  103     0
#>    TOLCOM   78  105     0
#>    TOLHOMO  92   91     0
#>    TOLATH   88   95     0
#>    TOLMIL   83  100     0
#> 
#> 
#> Summary of model fit
#>                                            
#>  Number of observations                 183
#>  Number of free parameters               23
#>  Log-likelihood                    -439.750
#>  Information criteria                      
#>    Akaike (AIC)                     925.500
#>    Bayesian (BIC)                   999.318
#>  Chi-squared Tests                         
#>    Residual degree of freedom (df)        8
#>    Pearson Chi-squared (X-squared)   10.766
#>      P(>Chi)                          0.215
#>    Likelihood Ratio (G-squared)      11.299
#>      P(>Chi)                          0.185
param(model_tol)
#> PI :
#> (tol)
#>   class
#>          1       2       3       4
#>     0.3590  0.0782  0.3569  0.2059
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3       4
#>    1(V1)  1.0000  0.9999  0.0000  0.0000
#>    2      0.0000  0.0001  1.0000  1.0000
#>    1(V2)  0.8620  0.2359  0.0541  0.3839
#>    2      0.1380  0.7641  0.9459  0.6161
#>    1(V3)  0.8924  0.3059  0.0191  0.7365
#>    2      0.1076  0.6941  0.9809  0.2635
#>    1(V4)  0.9536  0.4442  0.0113  0.4845
#>    2      0.0464  0.5558  0.9887  0.5155
#>    1(V5)  0.9894  0.0000  0.0889  0.3237
#>    2      0.0106  1.0000  0.9111  0.6763
#> 
#>     V1     V2     V3      V4     V5    
#> tol TOLRAC TOLCOM TOLHOMO TOLATH TOLMIL

model_lta <- slca(
   status(3) ~ PAPRES + PADEG + MADEG,
   tol(4) ~ TOLRAC + TOLCOM + TOLHOMO + TOLATH + TOLMIL,
   status ~ tol
) %>% estimate(data = gss500, control = list(em.tol = 1e-6))
summary(model_lta)
#> Structural Latent Class Model
#> 
#> Summary of model structure
#>                                    
#>  Number of manifest variables     8
#>  Number of latent class variables 2
#> 
#>  Latent variables (Root*):                    
#>   Label: status* tol
#>  nclass: 3       4  
#> 
#>  Measurement model:                                                          
#>   status -> { PAPRES, PADEG, MADEG }                     a
#>   tol    -> { TOLRAC, TOLCOM, TOLHOMO, TOLATH, TOLMIL }  b
#> 
#>  Structural model:                   
#>   status -> { tol }
#> 
#>  Dependency constraints:
#>   A            
#>   status -> tol
#> 
#>  Tree of structural model:                
#>   status  -> tol
#> 
#> 
#> Summary of manifest variables
#> 
#>  Categories for each variable:
#>           response
#>             1         2           3        4         5       
#>    PAPRES   LOW       MIDIUM      HIGH                       
#>    PADEG    LT-HS     HIGH-SCH    COLLEGE  BACHELOR  GRADUATE
#>    MADEG    LT-HS     HIGH-SCH    COLLEGE  BACHELOR  GRADUATE
#>    TOLRAC   TOLERANT  INTOLERANT                             
#>    TOLCOM   TOLERANT  INTOLERANT                             
#>    TOLHOMO  TOLERANT  INTOLERANT                             
#>    TOLATH   TOLERANT  INTOLERANT                             
#>    TOLMIL   TOLERANT  INTOLERANT                             
#> 
#>  Frequencies for each categories:
#>           response
#>               1    2   3   4  5  <NA>
#>    PAPRES    62  102  19            0
#>    PADEG    104   63      11  5     0
#>    MADEG     86   84   4   8  1     0
#>    TOLRAC    80  103                0
#>    TOLCOM    78  105                0
#>    TOLHOMO   92   91                0
#>    TOLATH    88   95                0
#>    TOLMIL    83  100                0
#> 
#> 
#> Summary of model fit
#>                                            
#>  Number of observations                 183
#>  Number of free parameters               61
#>  Log-likelihood                    -898.402
#>  Information criteria                      
#>    Akaike (AIC)                    1918.803
#>    Bayesian (BIC)                  2114.582
#>  Chi-squared Tests                         
#>    Residual degree of freedom (df)     2338
#>    Pearson Chi-squared (X-squared)  676.456
#>      P(>Chi)                          1.000
#>    Likelihood Ratio (G-squared)     277.653
#>      P(>Chi)                          1.000
param(model_lta)
#> PI :
#> (status)
#>   class
#>          1       2       3
#>     0.2623  0.4016  0.3360
#> 
#> TAU :
#> (A)
#>      parent
#> child       1       2       3
#>     1  0.2875  0.3595  0.0000
#>     2  0.1408  0.0000  0.0974
#>     3  0.0431  0.3596  0.7592
#>     4  0.5286  0.2809  0.1434
#>              
#> parent status
#> child  tol   
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  0.1179  0.5140  0.3019
#>    2      0.7166  0.3356  0.6981
#>    3      0.1655  0.1504  0.0000
#>    1(V2)  0.0000  0.5924  0.9832
#>    2      0.7687  0.3551  0.0000
#>    3      0.0000  0.0000  0.0000
#>    4      0.2076  0.0000  0.0168
#>    5      0.0237  0.0525  0.0000
#>    1(V3)  0.0000  0.4202  0.8963
#>    2      0.7712  0.5798  0.0709
#>    3      0.0833  0.0000  0.0000
#>    4      0.1246  0.0000  0.0328
#>    5      0.0208  0.0000  0.0000
#> 
#>        V1     V2    V3   
#> status PAPRES PADEG MADEG
#> (b)
#>         class
#> response       1       2       3       4
#>    1(V1)  0.3321  0.6362  0.0491  1.0000
#>    2      0.6679  0.3638  0.9509  0.0000
#>    1(V2)  0.3338  0.9526  0.0645  0.8676
#>    2      0.6662  0.0474  0.9355  0.1324
#>    1(V3)  0.6531  0.2546  0.1017  1.0000
#>    2      0.3469  0.7454  0.8983  0.0000
#>    1(V4)  0.6248  0.7385  0.0000  0.9747
#>    2      0.3752  0.2615  1.0000  0.0253
#>    1(V5)  0.1873  1.0000  0.1048  1.0000
#>    2      0.8127  0.0000  0.8952  0.0000
#> 
#>     V1     V2     V3      V4     V5    
#> tol TOLRAC TOLCOM TOLHOMO TOLATH TOLMIL

# \donttest{
regress(model_lta, status ~ SEX, gss500)
#> Coefficients:     
#> class  (Intercept)  SEXFEMALE
#>   1/3  -0.272       -0.091   
#>   2/3  -0.127        0.488   
regress(model_lta, status ~ SEX, gss500, method = "BCH")
#> Coefficients:     
#> class  (Intercept)  SEXFEMALE
#>   1/3  -0.271       -0.115   
#>   2/3  -0.280        0.733   
regress(model_lta, status ~ SEX, gss500, method = "ML")
#> Coefficients:     
#> class  (Intercept)  SEXFEMALE
#>   1/3  -0.271       -0.112   
#>   2/3   0.184        0.745   
# }