Regress Exogenous Variables on Latent Variables

Performs regression analysis to examine the influence of exogenous (external) variables on latent class variables in an estimated slca model. The function uses logistic regression with a three-step approach to account for measurement error.

Usage

regress(object, ...)

# S3 method for class 'slcafit'
regress(
   object, formula, data = parent.frame(),
   imputation = c("modal", "prob"),
   method = c("naive", "BCH", "ML"), ...
)

# S3 method for class 'slcafit'
regress(
  object,
  formula,
  data = parent.frame(),
  imputation = c("modal", "prob"),
  method = c("naive", "BCH", "ML"),
  ...
)

Arguments

object

an object of class slcafit.

...

additional arguments.

formula

a formula specifying the regression model, including both latent class variables (from the estimated model) and exogenous variables.

data

an optional data.frame containing the exogenous variables of interest. If omitted, the variables are taken from the parent environment.

imputation

a character string specifying the imputation method for latent class assignment. Options include:

"modal": Assigns each individual to the latent class with the highest posterior probability.

"prob": Assigns classes probabilistically based on the posterior probability distribution.

method

a character string specifying the method to adjust for bias in the three-step approach. Options include:

"naive": A simple approach without correction for classification error.

"BCH": The bias-adjusted Bolck, Croon, and Hagenaars method.

"ML": A maximum likelihood approach that accounts for classification error.

Value

A list of class reg.slca with the following components:

coefficients

A matrix of regression coefficients representing the odds ratios for each latent class against the baseline class (the last class).

std.err

A matrix of standard errors corresponding to the regression coefficients.

vcov

The variance-covariance matrix of the regression coefficients.

dim

The dimensions of the coefficients matrix.

ll

The log-likelihood of the regression model.

The summary function can be used to display the regression coefficients, standard errors, Wald statistics, and p-values. The standard errors are derived by numerically calculated Hessian matrix from nlm function.

References

Vermunt, J. K. (2010). Latent Class Modeling with Covariates: Two Improved Three-Step Approaches. Political Analysis, 18(4), 450–469. http://www.jstor.org/stable/25792024

Examples

library(magrittr)
names(nlsy97)
#>  [1] "SEX"     "RACE"    "ESMK_98" "FSMK_98" "DSMK_98" "HSMK_98" "EDRK_98"
#>  [8] "CDRK_98" "WDRK_98" "BDRK_98" "EMRJ_98" "CMRJ_98" "OMRJ_98" "SMRJ_98"
#> [15] "ESMK_03" "FSMK_03" "DSMK_03" "HSMK_03" "EDRK_03" "CDRK_03" "WDRK_03"
#> [22] "BDRK_03" "EMRJ_03" "CMRJ_03" "OMRJ_03" "SMRJ_03" "ESMK_08" "FSMK_08"
#> [29] "DSMK_08" "HSMK_08" "EDRK_08" "CDRK_08" "WDRK_08" "BDRK_08" "EMRJ_08"
#> [36] "CMRJ_08" "OMRJ_08" "SMRJ_08"
debug(regress)
nlsy_jlcpa %>% regress(SMK_98 ~ SEX, nlsy97)
#> debugging in: regress(., SMK_98 ~ SEX, nlsy97)
#> debug: UseMethod("regress")
#> debugging in: regress.slcafit(., SMK_98 ~ SEX, nlsy97)
#> debug: {
#>     model <- object$model
#>     labels <- all.vars(formula)
#>     latent <- labels[labels %in% row.names(model$latent)]
#>     imputation <- match.arg(imputation)
#>     method <- match.arg(method)
#>     model$reg$x <- setdiff(labels, latent)
#>     model$reg$y <- latent
#>     impute <- function(x, imputation) {
#>         if (imputation == "modal") 
#>             imputed <- as.factor(apply(x, 1, which.max))
#>         else imputed <- as.factor(apply(x, 1, function(y) sample(seq_len(ncol(x)), 
#>             1, prob = y)))
#>         return(imputed)
#>     }
#>     if (missing(data)) 
#>         data <- object$mf
#>     else data <- data.frame(object$mf, data)
#>     imputed <- lapply(object$posterior$marginal[latent], impute, 
#>         imputation)
#>     data <- data.frame(data, imputed)
#>     mf <- stats::model.frame(formula, data)
#>     cprobs <- function(X, b, ref) {
#>         beta <- matrix(nrow = nrow(b), ncol = ncol(b) + 1)
#>         beta[, ref] = 0
#>         beta[, -ref] = b
#>         xb <- X %*% beta
#>         exb <- exp(xb)
#>         denom <- rowSums(exb)
#>         xb - log(denom)
#>     }
#>     p <- object$posterior$marginal[[latent]][rownames(mf), ]
#>     w <- switch(imputation, modal = apply(p, 1, function(x) as.numeric(x == 
#>         max(x))), prob = t(p))
#>     y <- stats::model.response(mf)
#>     X <- stats::model.matrix(formula, mf, ...)
#>     nr <- nlevels(y) - 1
#>     nc <- ncol(X)
#>     init <- numeric(nc * nr)
#>     if (method == "naive") {
#>         naive_ll <- function(par, X, w, ref) {
#>             b <- matrix(par, nrow = ncol(X))
#>             prob <- cprobs(X, b, ref)
#>             -sum(prob * t(w))
#>         }
#>         fit1 <- try(suppressWarnings(stats::nlm(naive_ll, init, 
#>             X = X, y = y, ref = nlevels(y), hessian = TRUE)), 
#>             TRUE)
#>         if (!inherits(fit1, "try-error")) {
#>             ll <- fit1$minimum
#>             par <- list(fit1$estimate)
#>             hess <- list(fit1$hessian)
#>         }
#>         else {
#>             ll <- c()
#>             par <- list()
#>             hess <- list()
#>         }
#>         fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>             "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>             naive_ll, X = X, y = y, method = x, ref = nlevels(y), 
#>             hessian = TRUE)), TRUE))
#>         fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>         ll <- c(ll, sapply(fit2, "[[", "value"))
#>         par <- c(par, lapply(fit2, "[[", "par"))
#>         hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>     }
#>     else {
#>         d <- (w %*% p)/colSums(p)
#>         if (method == "BCH") {
#>             w_ <- t(w) %*% ginv(d)
#>             bch_ll <- function(par, X, w_, ref) {
#>                 b <- matrix(par, ncol(X))
#>                 prob <- cprobs(X, b, ref)
#>                 -sum(w_ * prob)
#>             }
#>             fit1 <- try(suppressWarnings(stats::nlm(bch_ll, init, 
#>                 X = X, w_ = w_, ref = nlevels(y), hessian = TRUE, 
#>                 iterlim = 2)), silent = TRUE)
#>             if (!inherits(fit1, "try-error")) {
#>                 ll <- fit1$minimum
#>                 par <- list(fit1$estimate)
#>                 hess <- list(fit1$hessian)
#>             }
#>             else {
#>                 ll <- c()
#>                 par <- list()
#>                 hess <- list()
#>             }
#>             fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>                 "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>                 bch_ll, X = X, w_ = w_, method = x, ref = nlevels(y), 
#>                 hessian = TRUE)), silent = TRUE))
#>             fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>             ll <- c(ll, sapply(fit2, "[[", "value"))
#>             par <- c(par, lapply(fit2, "[[", "par"))
#>             hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>         }
#>         else if (method == "ML") {
#>             w_ <- log(sapply(y, function(x) d[, x]))
#>             ml_ll <- function(par, X, w_, ref) {
#>                 b <- matrix(par, ncol(X))
#>                 prob <- t(cprobs(X, b, ref))
#>                 ll <- colSums(exp(prob + w_))
#>                 -sum(log(ll))
#>             }
#>             fit1 <- try(suppressWarnings(stats::nlm(ml_ll, init, 
#>                 X = X, w_ = w_, ref = nlevels(y), hessian = TRUE)), 
#>                 silent = TRUE)
#>             if (!inherits(fit1, "try-error")) {
#>                 ll <- fit1$minimum
#>                 par <- list(fit1$estimate)
#>                 hess <- list(fit1$hessian)
#>             }
#>             else {
#>                 ll <- c()
#>                 par <- list()
#>                 hess <- list()
#>             }
#>             fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>                 "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>                 ml_ll, X = X, w_ = w_, method = x, ref = nlevels(y), 
#>                 hessian = TRUE)), TRUE))
#>             fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>             ll <- c(ll, sapply(fit2, "[[", "value"))
#>             par <- c(par, lapply(fit2, "[[", "par"))
#>             hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>         }
#>     }
#>     par <- par[[which.min(ll)]]
#>     hess <- hess[[which.min(ll)]]
#>     rn <- paste0(seq_len(nr), "/", nr + 1)
#>     cn <- colnames(X)
#>     coef <- matrix(par, nr, nc, byrow = TRUE, dimnames = list(class = rn, 
#>         cn))
#>     dn <- paste0(rep(cn, nr), "|", rep(rn, each = nc))
#>     vcov <- matrix(ginv(hess), nc * nr, nc * nr, dimnames = list(dn, 
#>         dn))
#>     se <- matrix(diag(vcov), nr, nc, byrow = TRUE, dimnames = list(class = rn, 
#>         cn))
#>     res <- list()
#>     res$model <- model
#>     res$coefficients <- coef
#>     res$std.err <- se
#>     res$vcov <- vcov
#>     res$dim <- c(nr, nc)
#>     res$ll <- -min(ll)
#>     class(res) <- "reg.slca"
#>     return(res)
#> }
#> debug: model <- object$model
#> debug: labels <- all.vars(formula)
#> debug: latent <- labels[labels %in% row.names(model$latent)]
#> debug: imputation <- match.arg(imputation)
#> debug: method <- match.arg(method)
#> debug: model$reg$x <- setdiff(labels, latent)
#> debug: model$reg$y <- latent
#> debug: impute <- function(x, imputation) {
#>     if (imputation == "modal") 
#>         imputed <- as.factor(apply(x, 1, which.max))
#>     else imputed <- as.factor(apply(x, 1, function(y) sample(seq_len(ncol(x)), 
#>         1, prob = y)))
#>     return(imputed)
#> }
#> debug: if (missing(data)) data <- object$mf else data <- data.frame(object$mf, 
#>     data)
#> debug: data <- data.frame(object$mf, data)
#> debug: imputed <- lapply(object$posterior$marginal[latent], impute, 
#>     imputation)
#> debug: data <- data.frame(data, imputed)
#> debug: mf <- stats::model.frame(formula, data)
#> debug: cprobs <- function(X, b, ref) {
#>     beta <- matrix(nrow = nrow(b), ncol = ncol(b) + 1)
#>     beta[, ref] = 0
#>     beta[, -ref] = b
#>     xb <- X %*% beta
#>     exb <- exp(xb)
#>     denom <- rowSums(exb)
#>     xb - log(denom)
#> }
#> debug: p <- object$posterior$marginal[[latent]][rownames(mf), ]
#> debug: w <- switch(imputation, modal = apply(p, 1, function(x) as.numeric(x == 
#>     max(x))), prob = t(p))
#> debug: y <- stats::model.response(mf)
#> debug: X <- stats::model.matrix(formula, mf, ...)
#> debug: nr <- nlevels(y) - 1
#> debug: nc <- ncol(X)
#> debug: init <- numeric(nc * nr)
#> debug: if (method == "naive") {
#>     naive_ll <- function(par, X, w, ref) {
#>         b <- matrix(par, nrow = ncol(X))
#>         prob <- cprobs(X, b, ref)
#>         -sum(prob * t(w))
#>     }
#>     fit1 <- try(suppressWarnings(stats::nlm(naive_ll, init, X = X, 
#>         y = y, ref = nlevels(y), hessian = TRUE)), TRUE)
#>     if (!inherits(fit1, "try-error")) {
#>         ll <- fit1$minimum
#>         par <- list(fit1$estimate)
#>         hess <- list(fit1$hessian)
#>     }
#>     else {
#>         ll <- c()
#>         par <- list()
#>         hess <- list()
#>     }
#>     fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>         "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>         naive_ll, X = X, y = y, method = x, ref = nlevels(y), 
#>         hessian = TRUE)), TRUE))
#>     fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>     ll <- c(ll, sapply(fit2, "[[", "value"))
#>     par <- c(par, lapply(fit2, "[[", "par"))
#>     hess <- c(hess, lapply(fit2, "[[", "hessian"))
#> } else {
#>     d <- (w %*% p)/colSums(p)
#>     if (method == "BCH") {
#>         w_ <- t(w) %*% ginv(d)
#>         bch_ll <- function(par, X, w_, ref) {
#>             b <- matrix(par, ncol(X))
#>             prob <- cprobs(X, b, ref)
#>             -sum(w_ * prob)
#>         }
#>         fit1 <- try(suppressWarnings(stats::nlm(bch_ll, init, 
#>             X = X, w_ = w_, ref = nlevels(y), hessian = TRUE, 
#>             iterlim = 2)), silent = TRUE)
#>         if (!inherits(fit1, "try-error")) {
#>             ll <- fit1$minimum
#>             par <- list(fit1$estimate)
#>             hess <- list(fit1$hessian)
#>         }
#>         else {
#>             ll <- c()
#>             par <- list()
#>             hess <- list()
#>         }
#>         fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>             "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>             bch_ll, X = X, w_ = w_, method = x, ref = nlevels(y), 
#>             hessian = TRUE)), silent = TRUE))
#>         fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>         ll <- c(ll, sapply(fit2, "[[", "value"))
#>         par <- c(par, lapply(fit2, "[[", "par"))
#>         hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>     }
#>     else if (method == "ML") {
#>         w_ <- log(sapply(y, function(x) d[, x]))
#>         ml_ll <- function(par, X, w_, ref) {
#>             b <- matrix(par, ncol(X))
#>             prob <- t(cprobs(X, b, ref))
#>             ll <- colSums(exp(prob + w_))
#>             -sum(log(ll))
#>         }
#>         fit1 <- try(suppressWarnings(stats::nlm(ml_ll, init, 
#>             X = X, w_ = w_, ref = nlevels(y), hessian = TRUE)), 
#>             silent = TRUE)
#>         if (!inherits(fit1, "try-error")) {
#>             ll <- fit1$minimum
#>             par <- list(fit1$estimate)
#>             hess <- list(fit1$hessian)
#>         }
#>         else {
#>             ll <- c()
#>             par <- list()
#>             hess <- list()
#>         }
#>         fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>             "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>             ml_ll, X = X, w_ = w_, method = x, ref = nlevels(y), 
#>             hessian = TRUE)), TRUE))
#>         fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>         ll <- c(ll, sapply(fit2, "[[", "value"))
#>         par <- c(par, lapply(fit2, "[[", "par"))
#>         hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>     }
#> }
#> debug: naive_ll <- function(par, X, w, ref) {
#>     b <- matrix(par, nrow = ncol(X))
#>     prob <- cprobs(X, b, ref)
#>     -sum(prob * t(w))
#> }
#> debug: fit1 <- try(suppressWarnings(stats::nlm(naive_ll, init, X = X, 
#>     y = y, ref = nlevels(y), hessian = TRUE)), TRUE)
#> debug: if (!inherits(fit1, "try-error")) {
#>     ll <- fit1$minimum
#>     par <- list(fit1$estimate)
#>     hess <- list(fit1$hessian)
#> } else {
#>     ll <- c()
#>     par <- list()
#>     hess <- list()
#> }
#> debug: ll <- c()
#> debug: par <- list()
#> debug: hess <- list()
#> debug: fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN"), 
#>     function(x) try(suppressWarnings(stats::optim(init, naive_ll, 
#>         X = X, y = y, method = x, ref = nlevels(y), hessian = TRUE)), 
#>         TRUE))
#> debug: fit2 <- fit2[sapply(fit2, class) != "try-error"]
#> debug: ll <- c(ll, sapply(fit2, "[[", "value"))
#> debug: par <- c(par, lapply(fit2, "[[", "par"))
#> debug: hess <- c(hess, lapply(fit2, "[[", "hessian"))
#> debug: par <- par[[which.min(ll)]]
#> Error in par[[which.min(ll)]]: attempt to select less than one element in get1index
# \donttest{
nlsy_jlcpa %>% regress(PROF ~ SEX, nlsy97)
#> debugging in: regress(., PROF ~ SEX, nlsy97)
#> debug: UseMethod("regress")
#> debugging in: regress.slcafit(., PROF ~ SEX, nlsy97)
#> debug: {
#>     model <- object$model
#>     labels <- all.vars(formula)
#>     latent <- labels[labels %in% row.names(model$latent)]
#>     imputation <- match.arg(imputation)
#>     method <- match.arg(method)
#>     model$reg$x <- setdiff(labels, latent)
#>     model$reg$y <- latent
#>     impute <- function(x, imputation) {
#>         if (imputation == "modal") 
#>             imputed <- as.factor(apply(x, 1, which.max))
#>         else imputed <- as.factor(apply(x, 1, function(y) sample(seq_len(ncol(x)), 
#>             1, prob = y)))
#>         return(imputed)
#>     }
#>     if (missing(data)) 
#>         data <- object$mf
#>     else data <- data.frame(object$mf, data)
#>     imputed <- lapply(object$posterior$marginal[latent], impute, 
#>         imputation)
#>     data <- data.frame(data, imputed)
#>     mf <- stats::model.frame(formula, data)
#>     cprobs <- function(X, b, ref) {
#>         beta <- matrix(nrow = nrow(b), ncol = ncol(b) + 1)
#>         beta[, ref] = 0
#>         beta[, -ref] = b
#>         xb <- X %*% beta
#>         exb <- exp(xb)
#>         denom <- rowSums(exb)
#>         xb - log(denom)
#>     }
#>     p <- object$posterior$marginal[[latent]][rownames(mf), ]
#>     w <- switch(imputation, modal = apply(p, 1, function(x) as.numeric(x == 
#>         max(x))), prob = t(p))
#>     y <- stats::model.response(mf)
#>     X <- stats::model.matrix(formula, mf, ...)
#>     nr <- nlevels(y) - 1
#>     nc <- ncol(X)
#>     init <- numeric(nc * nr)
#>     if (method == "naive") {
#>         naive_ll <- function(par, X, w, ref) {
#>             b <- matrix(par, nrow = ncol(X))
#>             prob <- cprobs(X, b, ref)
#>             -sum(prob * t(w))
#>         }
#>         fit1 <- try(suppressWarnings(stats::nlm(naive_ll, init, 
#>             X = X, y = y, ref = nlevels(y), hessian = TRUE)), 
#>             TRUE)
#>         if (!inherits(fit1, "try-error")) {
#>             ll <- fit1$minimum
#>             par <- list(fit1$estimate)
#>             hess <- list(fit1$hessian)
#>         }
#>         else {
#>             ll <- c()
#>             par <- list()
#>             hess <- list()
#>         }
#>         fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>             "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>             naive_ll, X = X, y = y, method = x, ref = nlevels(y), 
#>             hessian = TRUE)), TRUE))
#>         fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>         ll <- c(ll, sapply(fit2, "[[", "value"))
#>         par <- c(par, lapply(fit2, "[[", "par"))
#>         hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>     }
#>     else {
#>         d <- (w %*% p)/colSums(p)
#>         if (method == "BCH") {
#>             w_ <- t(w) %*% ginv(d)
#>             bch_ll <- function(par, X, w_, ref) {
#>                 b <- matrix(par, ncol(X))
#>                 prob <- cprobs(X, b, ref)
#>                 -sum(w_ * prob)
#>             }
#>             fit1 <- try(suppressWarnings(stats::nlm(bch_ll, init, 
#>                 X = X, w_ = w_, ref = nlevels(y), hessian = TRUE, 
#>                 iterlim = 2)), silent = TRUE)
#>             if (!inherits(fit1, "try-error")) {
#>                 ll <- fit1$minimum
#>                 par <- list(fit1$estimate)
#>                 hess <- list(fit1$hessian)
#>             }
#>             else {
#>                 ll <- c()
#>                 par <- list()
#>                 hess <- list()
#>             }
#>             fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>                 "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>                 bch_ll, X = X, w_ = w_, method = x, ref = nlevels(y), 
#>                 hessian = TRUE)), silent = TRUE))
#>             fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>             ll <- c(ll, sapply(fit2, "[[", "value"))
#>             par <- c(par, lapply(fit2, "[[", "par"))
#>             hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>         }
#>         else if (method == "ML") {
#>             w_ <- log(sapply(y, function(x) d[, x]))
#>             ml_ll <- function(par, X, w_, ref) {
#>                 b <- matrix(par, ncol(X))
#>                 prob <- t(cprobs(X, b, ref))
#>                 ll <- colSums(exp(prob + w_))
#>                 -sum(log(ll))
#>             }
#>             fit1 <- try(suppressWarnings(stats::nlm(ml_ll, init, 
#>                 X = X, w_ = w_, ref = nlevels(y), hessian = TRUE)), 
#>                 silent = TRUE)
#>             if (!inherits(fit1, "try-error")) {
#>                 ll <- fit1$minimum
#>                 par <- list(fit1$estimate)
#>                 hess <- list(fit1$hessian)
#>             }
#>             else {
#>                 ll <- c()
#>                 par <- list()
#>                 hess <- list()
#>             }
#>             fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>                 "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>                 ml_ll, X = X, w_ = w_, method = x, ref = nlevels(y), 
#>                 hessian = TRUE)), TRUE))
#>             fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>             ll <- c(ll, sapply(fit2, "[[", "value"))
#>             par <- c(par, lapply(fit2, "[[", "par"))
#>             hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>         }
#>     }
#>     par <- par[[which.min(ll)]]
#>     hess <- hess[[which.min(ll)]]
#>     rn <- paste0(seq_len(nr), "/", nr + 1)
#>     cn <- colnames(X)
#>     coef <- matrix(par, nr, nc, byrow = TRUE, dimnames = list(class = rn, 
#>         cn))
#>     dn <- paste0(rep(cn, nr), "|", rep(rn, each = nc))
#>     vcov <- matrix(ginv(hess), nc * nr, nc * nr, dimnames = list(dn, 
#>         dn))
#>     se <- matrix(diag(vcov), nr, nc, byrow = TRUE, dimnames = list(class = rn, 
#>         cn))
#>     res <- list()
#>     res$model <- model
#>     res$coefficients <- coef
#>     res$std.err <- se
#>     res$vcov <- vcov
#>     res$dim <- c(nr, nc)
#>     res$ll <- -min(ll)
#>     class(res) <- "reg.slca"
#>     return(res)
#> }
#> debug: model <- object$model
#> debug: labels <- all.vars(formula)
#> debug: latent <- labels[labels %in% row.names(model$latent)]
#> debug: imputation <- match.arg(imputation)
#> debug: method <- match.arg(method)
#> debug: model$reg$x <- setdiff(labels, latent)
#> debug: model$reg$y <- latent
#> debug: impute <- function(x, imputation) {
#>     if (imputation == "modal") 
#>         imputed <- as.factor(apply(x, 1, which.max))
#>     else imputed <- as.factor(apply(x, 1, function(y) sample(seq_len(ncol(x)), 
#>         1, prob = y)))
#>     return(imputed)
#> }
#> debug: if (missing(data)) data <- object$mf else data <- data.frame(object$mf, 
#>     data)
#> debug: data <- data.frame(object$mf, data)
#> debug: imputed <- lapply(object$posterior$marginal[latent], impute, 
#>     imputation)
#> debug: data <- data.frame(data, imputed)
#> debug: mf <- stats::model.frame(formula, data)
#> debug: cprobs <- function(X, b, ref) {
#>     beta <- matrix(nrow = nrow(b), ncol = ncol(b) + 1)
#>     beta[, ref] = 0
#>     beta[, -ref] = b
#>     xb <- X %*% beta
#>     exb <- exp(xb)
#>     denom <- rowSums(exb)
#>     xb - log(denom)
#> }
#> debug: p <- object$posterior$marginal[[latent]][rownames(mf), ]
#> debug: w <- switch(imputation, modal = apply(p, 1, function(x) as.numeric(x == 
#>     max(x))), prob = t(p))
#> debug: y <- stats::model.response(mf)
#> debug: X <- stats::model.matrix(formula, mf, ...)
#> debug: nr <- nlevels(y) - 1
#> debug: nc <- ncol(X)
#> debug: init <- numeric(nc * nr)
#> debug: if (method == "naive") {
#>     naive_ll <- function(par, X, w, ref) {
#>         b <- matrix(par, nrow = ncol(X))
#>         prob <- cprobs(X, b, ref)
#>         -sum(prob * t(w))
#>     }
#>     fit1 <- try(suppressWarnings(stats::nlm(naive_ll, init, X = X, 
#>         y = y, ref = nlevels(y), hessian = TRUE)), TRUE)
#>     if (!inherits(fit1, "try-error")) {
#>         ll <- fit1$minimum
#>         par <- list(fit1$estimate)
#>         hess <- list(fit1$hessian)
#>     }
#>     else {
#>         ll <- c()
#>         par <- list()
#>         hess <- list()
#>     }
#>     fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>         "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>         naive_ll, X = X, y = y, method = x, ref = nlevels(y), 
#>         hessian = TRUE)), TRUE))
#>     fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>     ll <- c(ll, sapply(fit2, "[[", "value"))
#>     par <- c(par, lapply(fit2, "[[", "par"))
#>     hess <- c(hess, lapply(fit2, "[[", "hessian"))
#> } else {
#>     d <- (w %*% p)/colSums(p)
#>     if (method == "BCH") {
#>         w_ <- t(w) %*% ginv(d)
#>         bch_ll <- function(par, X, w_, ref) {
#>             b <- matrix(par, ncol(X))
#>             prob <- cprobs(X, b, ref)
#>             -sum(w_ * prob)
#>         }
#>         fit1 <- try(suppressWarnings(stats::nlm(bch_ll, init, 
#>             X = X, w_ = w_, ref = nlevels(y), hessian = TRUE, 
#>             iterlim = 2)), silent = TRUE)
#>         if (!inherits(fit1, "try-error")) {
#>             ll <- fit1$minimum
#>             par <- list(fit1$estimate)
#>             hess <- list(fit1$hessian)
#>         }
#>         else {
#>             ll <- c()
#>             par <- list()
#>             hess <- list()
#>         }
#>         fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>             "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>             bch_ll, X = X, w_ = w_, method = x, ref = nlevels(y), 
#>             hessian = TRUE)), silent = TRUE))
#>         fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>         ll <- c(ll, sapply(fit2, "[[", "value"))
#>         par <- c(par, lapply(fit2, "[[", "par"))
#>         hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>     }
#>     else if (method == "ML") {
#>         w_ <- log(sapply(y, function(x) d[, x]))
#>         ml_ll <- function(par, X, w_, ref) {
#>             b <- matrix(par, ncol(X))
#>             prob <- t(cprobs(X, b, ref))
#>             ll <- colSums(exp(prob + w_))
#>             -sum(log(ll))
#>         }
#>         fit1 <- try(suppressWarnings(stats::nlm(ml_ll, init, 
#>             X = X, w_ = w_, ref = nlevels(y), hessian = TRUE)), 
#>             silent = TRUE)
#>         if (!inherits(fit1, "try-error")) {
#>             ll <- fit1$minimum
#>             par <- list(fit1$estimate)
#>             hess <- list(fit1$hessian)
#>         }
#>         else {
#>             ll <- c()
#>             par <- list()
#>             hess <- list()
#>         }
#>         fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", 
#>             "SANN"), function(x) try(suppressWarnings(stats::optim(init, 
#>             ml_ll, X = X, w_ = w_, method = x, ref = nlevels(y), 
#>             hessian = TRUE)), TRUE))
#>         fit2 <- fit2[sapply(fit2, class) != "try-error"]
#>         ll <- c(ll, sapply(fit2, "[[", "value"))
#>         par <- c(par, lapply(fit2, "[[", "par"))
#>         hess <- c(hess, lapply(fit2, "[[", "hessian"))
#>     }
#> }
#> debug: naive_ll <- function(par, X, w, ref) {
#>     b <- matrix(par, nrow = ncol(X))
#>     prob <- cprobs(X, b, ref)
#>     -sum(prob * t(w))
#> }
#> debug: fit1 <- try(suppressWarnings(stats::nlm(naive_ll, init, X = X, 
#>     y = y, ref = nlevels(y), hessian = TRUE)), TRUE)
#> debug: if (!inherits(fit1, "try-error")) {
#>     ll <- fit1$minimum
#>     par <- list(fit1$estimate)
#>     hess <- list(fit1$hessian)
#> } else {
#>     ll <- c()
#>     par <- list()
#>     hess <- list()
#> }
#> debug: ll <- c()
#> debug: par <- list()
#> debug: hess <- list()
#> debug: fit2 <- lapply(c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN"), 
#>     function(x) try(suppressWarnings(stats::optim(init, naive_ll, 
#>         X = X, y = y, method = x, ref = nlevels(y), hessian = TRUE)), 
#>         TRUE))
#> debug: fit2 <- fit2[sapply(fit2, class) != "try-error"]
#> debug: ll <- c(ll, sapply(fit2, "[[", "value"))
#> debug: par <- c(par, lapply(fit2, "[[", "par"))
#> debug: hess <- c(hess, lapply(fit2, "[[", "hessian"))
#> debug: par <- par[[which.min(ll)]]
#> Error in par[[which.min(ll)]]: attempt to select less than one element in get1index
# }