Simulate Data from slca Model

This function simulates data from a constructed slca model. If the model is not already estimated, parameters can be specified by the user or generated randomly.

Usage

# S3 method for class 'slca'
simulate(object, nsim = 500, seed = NULL, parm, nlevel, ...)

Arguments

object

a slca object representing the model from which data will be simulated.

nsim

the number of response observations to simulate. Defaults to 500.

seed

a random seed for reproducibility of the

parm

a set of parameters provided by the user to guide the simulation, if the model has not been estimated.

nlevel

the number of levels for each manifest item declared in the model. If not provided, the default is 2.

...

additional arguments.

Value

A list of two components:

class

A data.frame providing the assigned latent class for each individual across different latent class variables.

response

A data.frame containing the manifest items that were simulated.

Examples

m1 <- slca(lc1[3] ~ x1 + x2 + x3 + x4 + x5,
              lc2[4] ~ y1 + y2 + y3 + y4 + y5)
sim <- simulate(m1, 1000)
sapply(sim$class, table)
#> $lc1
#> 
#>   1   2   3 
#> 326 355 319 
#> 
#> $lc2
#> 
#>   1   2   3   4 
#> 269 374  62 295 
#> 

# simulate data with defined number of levels of manifest items
m2 <- slca(lc1[3] ~ x1 + x2 + x3 + x4)
sim <- simulate(m2, nlevel = c(3, 3, 3, 3))
d <- sim$response
sapply(d, table)
#>    x1  x2  x3  x4
#> 1 101 238 195 182
#> 2 230 215  97 180
#> 3 169  47 208 138

sim <- simulate(m2, nlevel = c(x1 = 2, x3 = 3, x4 = 4, x5 = 5))
d <- sim$response
sapply(d, table)
#> $x1
#> 
#>   1   2 
#> 334 166 
#> 
#> $x2
#> 
#>   1   2   3 
#> 162 153 185 
#> 
#> $x3
#> 
#>   1   2   3   4 
#> 139  91 139 131 
#> 
#> $x4
#> 
#>   1   2   3   4   5 
#>  61 153  99 152  35 
#> 

# simulate data with user-defined parameters
pi <- rep(1 / 3, 3)
rho <- c(.9, .1, .9, .1, .9, .1, .9, .1,
         .9, .1, .9, .1, .1, .9, .1, .9,
         .1, .9, .1, .9, .1, .9, .1, .9)
par <- c(pi, rho)
m3 <- slca(lc[3] ~ y1 + y2 + y3 + y4)
sim <- simulate(m3, parm = par)
mf <- estimate(m3, sim$response)
param(mf)
#> PI :
#> (lc)
#>   class
#>          1       2       3
#>     0.4206  0.3384  0.2411
#> 
#> RHO :
#> (a)
#>         class
#> response       1       2       3
#>    1(V1)  1.0000  0.0047  0.8870
#>    2      0.0000  0.9953  0.1130
#>    1(V2)  0.8439  0.1499  0.8393
#>    2      0.1561  0.8501  0.1607
#>    1(V3)  0.1285  0.0854  1.0000
#>    2      0.8715  0.9146  0.0000
#>    1(V4)  0.1143  0.1090  1.0000
#>    2      0.8857  0.8910  0.0000
#> 
#>    V1 V2 V3 V4
#> lc y1 y2 y3 y4